cor_lower {correlation}	R Documentation

Return the upper or lower triangular part

Description

Return the upper or lower triangular part of the correlation matrix.

Usage

cor_lower(x, diag = FALSE, ...)

Arguments

x
A correlation object.

diag
Should the diagonal be included?

...
Other arguments to be passed to or from other functions.

Examples

x <- correlation(mtcars, redundant = TRUE) # Generate full matrix
x <- cor_lower(x)

if (require("ggplot2")) {
  ggplot(x, aes(x = Parameter2, y = Parameter1, fill = r)) +
    geom_tile()
}

# Sorted
x <- correlation(mtcars, redundant = TRUE) # Generate full matrix
x <- cor_sort(x)
x <- cor_lower(x)

if (require("ggplot2")) {
  ggplot(x, aes(x = Parameter2, y = Parameter1, fill = r)) +
    geom_tile()
}

[Package correlation version 0.8.8 Index]