cor_text {correlation}	R Documentation

Correlation text

Description

This function returns a formatted character of correlation statistics.

Usage

cor_text(x, show_ci = TRUE, show_statistic = TRUE, show_sig = TRUE, ...)

Arguments

x
A dataframe with correlation statistics.

show_ci, show_statistic, show_sig
Toggle on/off different parts of the text.

...
Other arguments to be passed to or from other functions.

Examples

rez <- cor_test(mtcars, "mpg", "wt")

cor_text(rez)
cor_text(rez, show_statistic = FALSE, show_ci = FALSE, stars = TRUE)

rez <- correlation(mtcars)

cor_text(rez)

[Package correlation version 0.8.8 Index]