Correlation Analysis

Correlation Analysis

correlation(data, data2 = NULL, ci = "default", method = "pearson",
  p_adjust = "holm", partial = FALSE, bayesian = FALSE,
  iterations = 10^4, rope_full = TRUE, rope_bounds = c(-0.05, 0.05),
  prior = "medium", ...)

Arguments

data
A dataframe.

data2
An optional dataframe.

ci
Confidence/Credible Interval level. If "default", then 0.95 for Frequentist and 0.9 for Bayesian.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson" (default), "kendall", or "spearman", can be abbreviated.

p_adjust
Correction method for frequentist correlations. One of "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".

partial
Can be TRUE or "semi" for partial and semi-partial correlations, respectively. This only works for Frequentist correlations.

bayesian
If TRUE, the arguments below apply.

iterations
The number of iterations to sample.

rope_full
If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the ci argument.

rope_bounds
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.05, 0.05)).

prior
For the prior argument, several named values are recognized: "medium.narrow", "medium", "wide", and "ultrawide". These correspond to r scale values of 1/sqrt(27), 1/3, 1/sqrt(3) and 1, respectively. See the BayesFactor::correlationBF function.

...
Arguments passed to or from other methods.

Examples

library(dplyr)

#> Warning: package 'dplyr' was built under R version 3.5.2

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

correlation(iris)

#>      Parameter1   Parameter2          r         t DoF            p     CI_low
#> 1  Sepal.Length Sepal.Length  1.0000000       Inf 148 0.000000e+00  1.0000000
#> 2   Sepal.Width Sepal.Length -0.1175698 -1.440287 148 3.037965e-01 -0.2726932
#> 3  Petal.Length Sepal.Length  0.8717538 21.646019 148 1.038667e-46  0.8270363
#> 4   Petal.Width Sepal.Length  0.8179411 17.296454 148 1.860398e-36  0.7568971
#> 5  Sepal.Length  Sepal.Width -0.1175698 -1.440287 148 3.037965e-01 -0.2726932
#> 6   Sepal.Width  Sepal.Width  1.0000000       Inf 148 0.000000e+00  1.0000000
#> 7  Petal.Length  Sepal.Width -0.4284401 -5.768449 148 2.707989e-07 -0.5508771
#> 8   Petal.Width  Sepal.Width -0.3661259 -4.786461 148 1.629291e-05 -0.4972130
#> 9  Sepal.Length Petal.Length  0.8717538 21.646019 148 1.038667e-46  0.8270363
#> 10  Sepal.Width Petal.Length -0.4284401 -5.768449 148 2.707989e-07 -0.5508771
#> 11 Petal.Length Petal.Length  1.0000000       Inf 148 0.000000e+00  1.0000000
#> 12  Petal.Width Petal.Length  0.9628654 43.387237 148 5.610005e-85  0.9490525
#> 13 Sepal.Length  Petal.Width  0.8179411 17.296454 148 1.860398e-36  0.7568971
#> 14  Sepal.Width  Petal.Width -0.3661259 -4.786461 148 1.629291e-05 -0.4972130
#> 15 Petal.Length  Petal.Width  0.9628654 43.387237 148 5.610005e-85  0.9490525
#> 16  Petal.Width  Petal.Width  1.0000000       Inf 148 0.000000e+00  1.0000000
#>        CI_high CI_level  Method
#> 1   1.00000000     0.95 Pearson
#> 2   0.04351158     0.95 Pearson
#> 3   0.90550805     0.95 Pearson
#> 4   0.86483606     0.95 Pearson
#> 5   0.04351158     0.95 Pearson
#> 6   1.00000000     0.95 Pearson
#> 7  -0.28794993     0.95 Pearson
#> 8  -0.21869663     0.95 Pearson
#> 9   0.90550805     0.95 Pearson
#> 10 -0.28794993     0.95 Pearson
#> 11  1.00000000     0.95 Pearson
#> 12  0.97298532     0.95 Pearson
#> 13  0.86483606     0.95 Pearson
#> 14 -0.21869663     0.95 Pearson
#> 15  0.97298532     0.95 Pearson
#> 16  1.00000000     0.95 Pearson

iris %>%
  select(starts_with("Sepal")) %>%
  correlation()

#>     Parameter1   Parameter2          r         t DoF         p     CI_low
#> 1 Sepal.Length Sepal.Length  1.0000000       Inf 148 0.0000000  1.0000000
#> 2  Sepal.Width Sepal.Length -0.1175698 -1.440287 148 0.3037965 -0.2726932
#> 3 Sepal.Length  Sepal.Width -0.1175698 -1.440287 148 0.3037965 -0.2726932
#> 4  Sepal.Width  Sepal.Width  1.0000000       Inf 148 0.0000000  1.0000000
#>      CI_high CI_level  Method
#> 1 1.00000000     0.95 Pearson
#> 2 0.04351158     0.95 Pearson
#> 3 0.04351158     0.95 Pearson
#> 4 1.00000000     0.95 Pearson

iris %>%
  select(starts_with("Sepal")) %>%
  correlation(select(iris, starts_with("Petal")))

#>     Parameter1   Parameter2          r         t DoF            p     CI_low
#> 1 Sepal.Length Petal.Length  0.8717538 21.646019 148 4.154670e-47  0.8270363
#> 2  Sepal.Width Petal.Length -0.4284401 -5.768449 148 9.026629e-08 -0.5508771
#> 3 Sepal.Length  Petal.Width  0.8179411 17.296454 148 6.976494e-37  0.7568971
#> 4  Sepal.Width  Petal.Width -0.3661259 -4.786461 148 4.073229e-06 -0.4972130
#>      CI_high CI_level  Method
#> 1  0.9055080     0.95 Pearson
#> 2 -0.2879499     0.95 Pearson
#> 3  0.8648361     0.95 Pearson
#> 4 -0.2186966     0.95 Pearson

iris %>%
  group_by(Species) %>%
  correlation()

#>         Group   Parameter1   Parameter2         r         t DoF            p
#> 1      setosa Sepal.Length Sepal.Length 1.0000000       Inf  48 0.000000e+00
#> 2      setosa  Sepal.Width Sepal.Length 0.7425467  7.680738  48 8.051812e-09
#> 3      setosa Petal.Length Sepal.Length 0.2671758  1.920876  48 4.042115e-01
#> 4      setosa  Petal.Width Sepal.Length 0.2780984  2.005847  48 4.042115e-01
#> 5      setosa Sepal.Length  Sepal.Width 0.7425467  7.680738  48 8.051812e-09
#> 6      setosa  Sepal.Width  Sepal.Width 1.0000000       Inf  48 0.000000e+00
#> 7      setosa Petal.Length  Sepal.Width 0.1777000  1.251052  48 4.339578e-01
#> 8      setosa  Petal.Width  Sepal.Width 0.2327520  1.658091  48 4.152846e-01
#> 9      setosa Sepal.Length Petal.Length 0.2671758  1.920876  48 4.042115e-01
#> 10     setosa  Sepal.Width Petal.Length 0.1777000  1.251052  48 4.339578e-01
#> 11     setosa Petal.Length Petal.Length 1.0000000       Inf  48 0.000000e+00
#> 12     setosa  Petal.Width Petal.Length 0.3316300  2.435422  48 1.863892e-01
#> 13     setosa Sepal.Length  Petal.Width 0.2780984  2.005847  48 4.042115e-01
#> 14     setosa  Sepal.Width  Petal.Width 0.2327520  1.658091  48 4.152846e-01
#> 15     setosa Petal.Length  Petal.Width 0.3316300  2.435422  48 1.863892e-01
#> 16     setosa  Petal.Width  Petal.Width 1.0000000       Inf  48 0.000000e+00
#> 17 versicolor Sepal.Length Sepal.Length 1.0000000       Inf  48 0.000000e+00
#> 18 versicolor  Sepal.Width Sepal.Length 0.5259107  4.283887  48 1.754372e-04
#> 19 versicolor Petal.Length Sepal.Length 0.7540490  7.953806  48 2.586190e-09
#> 20 versicolor  Petal.Width Sepal.Length 0.5464611  4.520673  48 1.614169e-04
#> 21 versicolor Sepal.Length  Sepal.Width 0.5259107  4.283887  48 1.754372e-04
#> 22 versicolor  Sepal.Width  Sepal.Width 1.0000000       Inf  48 0.000000e+00
#> 23 versicolor Petal.Length  Sepal.Width 0.5605221  4.689320  48 1.381301e-04
#> 24 versicolor  Petal.Width  Sepal.Width 0.6639987  6.152348  48 1.173329e-06
#> 25 versicolor Sepal.Length Petal.Length 0.7540490  7.953806  48 2.586190e-09
#> 26 versicolor  Sepal.Width Petal.Length 0.5605221  4.689320  48 1.381301e-04
#> 27 versicolor Petal.Length Petal.Length 1.0000000       Inf  48 0.000000e+00
#> 28 versicolor  Petal.Width Petal.Length 0.7866681  8.827999  48 1.526299e-10
#> 29 versicolor Sepal.Length  Petal.Width 0.5464611  4.520673  48 1.614169e-04
#> 30 versicolor  Sepal.Width  Petal.Width 0.6639987  6.152348  48 1.173329e-06
#> 31 versicolor Petal.Length  Petal.Width 0.7866681  8.827999  48 1.526299e-10
#> 32 versicolor  Petal.Width  Petal.Width 1.0000000       Inf  48 0.000000e+00
#> 33  virginica Sepal.Length Sepal.Length 1.0000000       Inf  48 0.000000e+00
#> 34  virginica  Sepal.Width Sepal.Length 0.4572278  3.561892  48 6.747700e-03
#> 35  virginica Petal.Length Sepal.Length 0.8642247 11.901120  48 7.557343e-15
#> 36  virginica  Petal.Width Sepal.Length 0.2811077  2.029405  48 9.596299e-02
#> 37  virginica Sepal.Length  Sepal.Width 0.4572278  3.561892  48 6.747700e-03
#> 38  virginica  Sepal.Width  Sepal.Width 1.0000000       Inf  48 0.000000e+00
#> 39  virginica Petal.Length  Sepal.Width 0.4010446  3.033123  48 2.338622e-02
#> 40  virginica  Petal.Width  Sepal.Width 0.5377280  4.418702  48 5.647610e-04
#> 41  virginica Sepal.Length Petal.Length 0.8642247 11.901120  48 7.557343e-15
#> 42  virginica  Sepal.Width Petal.Length 0.4010446  3.033123  48 2.338622e-02
#> 43  virginica Petal.Length Petal.Length 1.0000000       Inf  48 0.000000e+00
#> 44  virginica  Petal.Width Petal.Length 0.3221082  2.357267  48 9.014307e-02
#> 45  virginica Sepal.Length  Petal.Width 0.2811077  2.029405  48 9.596299e-02
#> 46  virginica  Sepal.Width  Petal.Width 0.5377280  4.418702  48 5.647610e-04
#> 47  virginica Petal.Length  Petal.Width 0.3221082  2.357267  48 9.014307e-02
#> 48  virginica  Petal.Width  Petal.Width 1.0000000       Inf  48 0.000000e+00
#>           CI_low   CI_high CI_level  Method
#> 1   1.0000000000 1.0000000     0.95 Pearson
#> 2   0.5851391379 0.8460314     0.95 Pearson
#> 3  -0.0120695447 0.5077623     0.95 Pearson
#> 4  -0.0002702881 0.5164673     0.95 Pearson
#> 5   0.5851391379 0.8460314     0.95 Pearson
#> 6   1.0000000000 1.0000000     0.95 Pearson
#> 7  -0.1058850798 0.4345536     0.95 Pearson
#> 8  -0.0487543030 0.4800023     0.95 Pearson
#> 9  -0.0120695447 0.5077623     0.95 Pearson
#> 10 -0.1058850798 0.4345536     0.95 Pearson
#> 11  1.0000000000 1.0000000     0.95 Pearson
#> 12  0.0587009135 0.5584299     0.95 Pearson
#> 13 -0.0002702881 0.5164673     0.95 Pearson
#> 14 -0.0487543030 0.4800023     0.95 Pearson
#> 15  0.0587009135 0.5584299     0.95 Pearson
#> 16  1.0000000000 1.0000000     0.95 Pearson
#> 17  1.0000000000 1.0000000     0.95 Pearson
#> 18  0.2900175170 0.7015599     0.95 Pearson
#> 19  0.6020679651 0.8532995     0.95 Pearson
#> 20  0.3162110051 0.7159139     0.95 Pearson
#> 21  0.2900175170 0.7015599     0.95 Pearson
#> 22  1.0000000000 1.0000000     0.95 Pearson
#> 23  0.3343377667 0.7256533     0.95 Pearson
#> 24  0.4730883950 0.7953482     0.95 Pearson
#> 25  0.6020679651 0.8532995     0.95 Pearson
#> 26  0.3343377667 0.7256533     0.95 Pearson
#> 27  1.0000000000 1.0000000     0.95 Pearson
#> 28  0.6508310714 0.8737034     0.95 Pearson
#> 29  0.3162110051 0.7159139     0.95 Pearson
#> 30  0.4730883950 0.7953482     0.95 Pearson
#> 31  0.6508310714 0.8737034     0.95 Pearson
#> 32  1.0000000000 1.0000000     0.95 Pearson
#> 33  1.0000000000 1.0000000     0.95 Pearson
#> 34  0.2049657436 0.6525292     0.95 Pearson
#> 35  0.7714541554 0.9210172     0.95 Pearson
#> 36  0.0029942801 0.5188571     0.95 Pearson
#> 37  0.2049657436 0.6525292     0.95 Pearson
#> 38  1.0000000000 1.0000000     0.95 Pearson
#> 39  0.1381145662 0.6111677     0.95 Pearson
#> 40  0.3050368040 0.7098315     0.95 Pearson
#> 41  0.7714541554 0.9210172     0.95 Pearson
#> 42  0.1381145662 0.6111677     0.95 Pearson
#> 43  1.0000000000 1.0000000     0.95 Pearson
#> 44  0.0480703972 0.5510499     0.95 Pearson
#> 45  0.0029942801 0.5188571     0.95 Pearson
#> 46  0.3050368040 0.7098315     0.95 Pearson
#> 47  0.0480703972 0.5510499     0.95 Pearson
#> 48  1.0000000000 1.0000000     0.95 Pearson