correlation-package {correlation}	R Documentation

correlation: Methods for correlation analysis

Description

Lightweight package for computing different kinds of correlations, such as partial correlations, Bayesian correlations, multilevel correlations, polychoric correlations, biweight correlations, distance correlations and more. Part of the 'easystats' ecosystem.

References: Makowski et al. (2020) doi: 10.21105/joss.02306.

Details

correlation

Author(s)

Maintainer: Brenton M. Wiernik brenton@wiernik.org (ORCID)

Authors:

Dominique Makowski dom.makowski@gmail.com (ORCID) [inventor]

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID)

Daniel Lüdecke d.luedecke@uke.de (ORCID)

Mattan S. Ben-Shachar matanshm@post.bgu.ac.il (ORCID)

Rémi Thériault remi.theriault@mail.mcgill.ca (ORCID)

Other contributors:

Mark White markhwhiteii@gmail.com [reviewer]

Maximilian M. Rabe maximilian.rabe@uni-potsdam.de (ORCID) [reviewer]

See Also

Useful links:

https://easystats.github.io/correlation/

Report bugs at https://github.com/easystats/correlation/issues

[Package correlation version 0.8.8 Index]