display.easycormatrix {correlation}	R Documentation

Export tables into different output formats

Description

Export tables (i.e. data frame) into different output formats. print_md() is a alias for display(format = "markdown"). Note that you can use format() to get the formatted table as a dataframe.

Usage

## S3 method for class 'easycormatrix'
display(
  object,
  format = "markdown",
  digits = 2,
  p_digits = 3,
  stars = TRUE,
  include_significance = NULL,
  ...
)

## S3 method for class 'easycorrelation'
print_md(x, digits = NULL, p_digits = NULL, stars = NULL, ...)

## S3 method for class 'easycorrelation'
print_html(x, digits = NULL, p_digits = NULL, stars = NULL, ...)

## S3 method for class 'easycormatrix'
print_md(
  x,
  digits = NULL,
  p_digits = NULL,
  stars = NULL,
  include_significance = NULL,
  ...
)

## S3 method for class 'easycormatrix'
print_html(
  x,
  digits = NULL,
  p_digits = NULL,
  stars = NULL,
  include_significance = NULL,
  ...
)

Arguments

object, x
An object returned by correlation() or its summary.

format
String, indicating the output format. Currently, only "markdown" is supported.

digits, p_digits
To do...

stars
To do...

include_significance
To do...

...
Currently not used.

Details

display() is useful when the table-output from functions, which is usually printed as formatted text-table to console, should be formatted for pretty table-rendering in markdown documents, or if knitted from rmarkdown to PDF or Word files.

Value

A character vector. If format = "markdown", the return value will be a character vector in markdown-table format.

Examples

data(iris)
corr <- correlation(iris)
display(corr)

s <- summary(corr)
display(s)

[Package correlation version 0.8.8 Index]