Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A dataframe.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(iris[, 1:4])

#>       Distance     CI_low    CI_high
#> 1    2.1818990  1.4820848  2.9795163
#> 2    2.9291871  2.2529685  3.7803565
#> 3    2.1525199  1.5584474  2.7431538
#> 4    2.5323924  1.9519146  3.1703203
#> 5    2.5159740  1.7557286  3.4087846
#> 6    3.9971712  2.8058196  5.3803409
#> 7    2.9761756  2.3530638  3.7425456
#> 8    1.8841677  1.3316875  2.5579169
#> 9    3.4875385  2.7083250  4.3556599
#> 10   2.4423767  1.9730931  3.0177403
#> 11   3.3563884  2.4639707  4.6358445
#> 12   2.8876836  2.2220945  3.7293674
#> 13   2.6852262  2.0809564  3.3346428
#> 14   3.7181485  3.0081375  4.7138835
#> 15   9.0598275  6.9492167 11.8906740
#> 16   9.9907511  7.3643967 12.9919363
#> 17   5.9436400  4.3197247  7.7263019
#> 18   2.3891745  1.6811261  3.1984306
#> 19   4.6672172  3.4083446  6.1797628
#> 20   3.5431546  2.5315355  4.6956161
#> 21   2.7222501  2.1105582  3.5238479
#> 22   3.0147368  2.2546429  4.0721634
#> 23   3.7354126  2.9292754  4.7161059
#> 24   2.3158231  1.7058599  2.9997281
#> 25   5.4915003  4.2154384  6.9289075
#> 26   2.5112060  1.9088243  3.1779537
#> 27   1.8258497  1.3121902  2.4582785
#> 28   2.2455944  1.6199127  3.0682013
#> 29   2.5822548  1.8612365  3.3489861
#> 30   2.5507717  1.9770203  3.2515766
#> 31   2.0464103  1.5811744  2.6204475
#> 32   4.7619280  3.6496278  6.0813552
#> 33   8.6183879  6.4940419 10.9491952
#> 34   7.4452041  5.4353377  9.7887178
#> 35   2.0414453  1.5947997  2.6407362
#> 36   3.5240006  2.6525157  4.3941389
#> 37   5.4899401  4.3060430  6.9440725
#> 38   3.5660032  2.7372693  4.6649366
#> 39   3.2506732  2.5993754  4.1477331
#> 40   1.9455523  1.3609762  2.6237605
#> 41   2.6292897  1.9426993  3.4863694
#> 42  11.7326454  8.8812009 15.5656519
#> 43   3.4528482  2.7630982  4.3032858
#> 44   3.8489452  3.0657112  4.8351538
#> 45   4.6561007  3.5587077  5.9340872
#> 46   3.0609480  2.2487996  3.9192252
#> 47   4.5715689  3.3710826  5.8795756
#> 48   2.4442314  1.9137625  3.1586214
#> 49   3.0733892  2.1331009  4.1879555
#> 50   1.9859226  1.4383950  2.6043479
#> 51   4.6821750  3.4082478  6.3116384
#> 52   0.6918153  0.3562691  1.0404017
#> 53   3.1822347  2.2348977  4.4291955
#> 54   3.6545902  2.7344326  4.9867753
#> 55   2.1300780  1.5220715  3.0080698
#> 56   3.4144873  2.4219042  4.5098846
#> 57   1.3462568  0.8600181  1.9266314
#> 58   4.4106165  3.4039425  5.7744364
#> 59   2.8822926  2.0204034  3.9800324
#> 60   3.2937900  2.3406867  4.2341267
#> 61   7.7764395  5.9012576 10.1000618
#> 62   0.4683252  0.1880124  0.7490009
#> 63   7.7159041  6.1295069 10.2505267
#> 64   1.5965037  1.1026426  2.3656427
#> 65   1.0983088  0.7526785  1.5645550
#> 66   2.9404934  2.0529149  3.8529002
#> 67   3.6057927  2.6826589  4.7372089
#> 68   3.4389982  2.5948813  4.5749867
#> 69   7.5579466  5.8619379  9.9337573
#> 70   2.2308452  1.6858002  3.0646353
#> 71   3.3760555  2.3340478  4.5384280
#> 72   1.4060764  0.9279879  1.9585474
#> 73   2.4525101  1.7818358  3.3327735
#> 74   4.6565942  3.4794155  6.2962459
#> 75   1.6766037  1.1210487  2.3559552
#> 76   2.4518171  1.6613079  3.2658873
#> 77   4.3803567  3.1808237  5.8797273
#> 78   1.5080117  0.9329638  2.1252960
#> 79   0.3453285  0.1845008  0.5557530
#> 80   2.1591699  1.5452482  3.0079233
#> 81   2.8402976  2.1473716  3.8854236
#> 82   3.1894739  2.3425313  4.2040498
#> 83   0.8807166  0.5614325  1.3080655
#> 84   2.5919033  1.8886204  3.5187292
#> 85   6.4265124  4.8711422  8.1065058
#> 86   2.9033029  2.0361911  3.8507454
#> 87   1.9065766  1.2458914  2.6821675
#> 88   6.7046533  5.1738966  8.8001671
#> 89   1.7262785  1.2318628  2.4225686
#> 90   2.0381863  1.4544627  2.7577081
#> 91   5.0934637  3.7613372  6.6388364
#> 92   1.2239620  0.7399287  1.7843970
#> 93   1.3995689  0.9526702  1.9714643
#> 94   4.6593809  3.4742778  6.1207864
#> 95   1.5324227  0.9932127  2.0567955
#> 96   2.7742319  1.9417428  3.7151737
#> 97   1.2287001  0.7900342  1.7401758
#> 98   0.6977073  0.3523517  1.0530568
#> 99   4.9643029  3.6332928  6.5898800
#> 100  0.6601987  0.3923025  0.9837687
#> 101  9.2355374  6.5144478 12.0444888
#> 102  3.0834580  2.1276626  4.0382030
#> 103  2.5852279  1.8292602  3.4177389
#> 104  3.6595716  2.6375802  4.8057456
#> 105  2.4571924  1.6900286  3.3287739
#> 106  6.3076547  4.4721031  8.6928330
#> 107 10.5166826  8.0185695 13.3355391
#> 108  7.8707118  5.3904539 10.6095713
#> 109  3.4226585  2.4685760  4.4833457
#> 110  7.6357743  5.8866571 10.3066197
#> 111  2.3775183  1.5376339  3.4365528
#> 112  1.2894292  0.9594762  1.7108211
#> 113  2.5612023  1.7416559  3.6591287
#> 114  4.9482680  3.7267126  6.3353537
#> 115 11.8955140  8.5129731 15.4213373
#> 116  6.1365761  4.0112839  8.2858831
#> 117  2.0445989  1.4071098  2.8572626
#> 118 13.1873888  9.6376568 18.0950940
#> 119  7.4562617  5.4798396  9.7230884
#> 120  4.4814276  3.4135387  5.7644241
#> 121  4.2241839  2.6896275  5.6611875
#> 122  5.5776723  3.7933810  7.2759300
#> 123  9.1562175  6.6024686 12.2271650
#> 124  1.8711577  1.3295307  2.5850523
#> 125  2.9887187  2.0801650  4.0274177
#> 126  6.0675886  3.9711759  8.3525907
#> 127  1.5063915  1.0094150  2.1507189
#> 128  1.1907258  0.6963304  1.7455118
#> 129  1.7456458  1.2632318  2.3863014
#> 130  7.3318091  5.1916250 10.0791594
#> 131  5.9338365  4.2412360  7.8669905
#> 132 13.5478540  9.6285294 18.1786312
#> 133  2.6489136  1.8257508  3.6524362
#> 134  2.4812957  1.6414217  3.3934399
#> 135 13.2284530 10.0777404 17.2568849
#> 136 10.0708089  7.5689680 13.1310963
#> 137  8.4700710  6.1438290 11.4528911
#> 138  3.3604271  2.4509530  4.5458458
#> 139  1.5291780  0.9530703  2.2753130
#> 140  3.6359843  2.4302517  4.9983082
#> 141  6.2401432  4.0661262  8.3948176
#> 142 12.7726684  9.4263587 17.5836174
#> 143  3.0834580  2.1276626  4.0382030
#> 144  3.2312385  2.1353184  4.3649308
#> 145  7.7940334  5.2952981 10.4361409
#> 146  9.4106775  6.7315037 12.7797513
#> 147  4.1833915  3.1193523  5.3782148
#> 148  1.8685954  1.0767542  2.6322983
#> 149  7.9436292  5.6976951 10.7149226
#> 150  3.5972247  2.6489620  4.8127617

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688