Smooth a non-positive definite correlation matrix to make it positive definite

Make correlations positive definite using psych::cor.smooth. Note that this modifies the correlation values, and does not recompute the remaining indices (p-values, confidence intervals, etc.). Thus, once you used cor_smooth(), any other indices becomes irrelevant.

cor_smooth(x, method = "psych", verbose = TRUE, ...)

is.positive_definite(x, tol = .Machine$double.eps, ...)

Arguments

x
A correlation matrix.

method
Smoothing method. Can be psych (will use psych::cor.smooth()), hj (Jorjani et al., 2003) or lrs (Schaeffer, 2014). For the two last, will use mbend::bend() (check its documentation for details).

verbose
Set to FALSE to silence the function.

...
Other arguments to be passed to or from other functions.

tol
The minimum eigen values to be considered as acceptable.

Examples

set.seed(1)
data <- cbind(mtcars,
              mtcars * matrix(rnorm(32 * 11, sd = 0.05), ncol=11),
              mtcars * matrix(rnorm(32 * 11, sd = 0.05), ncol=11),
              mtcars * -matrix(rnorm(32 * 11, sd = 0.01), ncol=11))
x <- correlation(data)
is.positive_definite(x)
#> [1] FALSE

cor_smooth(as.matrix(x))
#>                  mpg          cyl        disp           hp        drat
#> mpg     1.0000000000 -0.852161959 -0.84755138 -0.776168372  0.68117191
#> cyl    -0.8521619594  1.000000000  0.90203287  0.832447453 -0.69993811
#> disp   -0.8475513793  0.902032872  1.00000000  0.790948586 -0.71021393
#> hp     -0.7761683718  0.832447453  0.79094859  1.000000000 -0.44875912
#> drat    0.6811719078 -0.699938114 -0.71021393 -0.448759117  1.00000000
#> wt     -0.8676593765  0.782495794  0.88797992  0.658747887 -0.71244065
#> qsec    0.4186840339 -0.591242074 -0.43369788 -0.708223389  0.09120476
#> vs      0.6640389191 -0.810811796 -0.71041589 -0.723096737  0.44027846
#> am      0.5998324295 -0.522607047 -0.59122704 -0.243204257  0.71271113
#> gear    0.4802847573 -0.492686599 -0.55556920 -0.125704258  0.69961013
#> carb   -0.5509250739  0.526988294  0.39497686  0.749812472 -0.09078980
#> mpg.1   0.1639611458 -0.130534259 -0.06377360 -0.162318615  0.10133199
#> cyl.1  -0.3001917787  0.263282953  0.28020264  0.450529211  0.17942350
#> disp.1 -0.0891710867  0.170409724  0.10479377  0.238041674 -0.19611128
#> hp.1   -0.0792314979  0.110831159 -0.03820947  0.032117697 -0.05592451
#> drat.1  0.0593023983  0.002694454 -0.04322827 -0.013150627  0.32741183
#> wt.1    0.1564186578 -0.097130238 -0.13120594 -0.276147988 -0.04779743
#> qsec.1 -0.1178709955  0.115573018  0.08634945  0.057299373  0.06220324
#> vs.1   -0.0034533932 -0.009644146 -0.08727569  0.063391591  0.13903605
#> am.1    0.3659435220 -0.311258411 -0.21209451 -0.324876781  0.20152985
#> gear.1  0.1552821458 -0.069817097  0.01017577 -0.180736229  0.04216641
#> carb.1 -0.1389857208  0.072195691 -0.04830612  0.331340522  0.13899528
#> mpg.2   0.0221977621 -0.025877541 -0.05650953  0.006905703 -0.01277756
#> cyl.2  -0.0234613664  0.094191428  0.15205302  0.129165087  0.02449422
#> disp.2  0.2927847657 -0.215589987 -0.25336988 -0.280986237  0.22999320
#> hp.2   -0.0731629156  0.024241534  0.06360487  0.138960634 -0.08145906
#> drat.2 -0.2396536275  0.310544878  0.36955596  0.100793138 -0.31150952
#> wt.2    0.0062940947 -0.023387189  0.11935702 -0.013136456 -0.13874957
#> qsec.2  0.0688607619 -0.122046887 -0.21528650 -0.065645357  0.19951691
#> vs.2   -0.0558734919  0.061338989  0.05164889  0.053086334 -0.13338091
#> am.2   -0.1910073181  0.115107334  0.04448810 -0.037207751  0.03669881
#> gear.2  0.1302736517 -0.105498547 -0.18187896 -0.014710492  0.01659415
#> carb.2 -0.0292541443 -0.069794457 -0.03114476  0.101138498  0.05950794
#> mpg.3  -0.0057327787 -0.080286176 -0.19769068 -0.084248592  0.11049273
#> cyl.3   0.0394416194  0.011830073  0.20938328 -0.028222970  0.04518604
#> disp.3  0.0716892572 -0.048073850 -0.08977965  0.089926849  0.01830283
#> hp.3    0.0008116447  0.109210118 -0.07051807  0.123008609  0.23152730
#> drat.3 -0.1069740085  0.109931709  0.11443882  0.141169316 -0.14625729
#> wt.3    0.1487085587 -0.168379495 -0.04969629 -0.255130218 -0.05350804
#> qsec.3 -0.0518622547  0.009597514  0.06066935  0.097826851 -0.21898312
#> vs.3   -0.4289104765  0.205518527  0.16457702  0.209958233 -0.25565395
#> am.3   -0.1136072418 -0.093409382 -0.04548350 -0.005939245 -0.15108353
#> gear.3 -0.2327788675  0.199263207  0.28309739  0.046949787 -0.16291493
#> carb.3  0.2428338772 -0.163843910 -0.10213400 -0.266491685  0.25722680
#>                   wt        qsec            vs           am        gear
#> mpg    -0.8676593765  0.41868403  0.6640389191  0.599832429  0.48028476
#> cyl     0.7824957945 -0.59124207 -0.8108117961 -0.522607047 -0.49268660
#> disp    0.8879799221 -0.43369788 -0.7104158908 -0.591227040 -0.55556920
#> hp      0.6587478873 -0.70822339 -0.7230967374 -0.243204257 -0.12570426
#> drat   -0.7124406467  0.09120476  0.4402784650  0.712711127  0.69961013
#> wt      1.0000000000 -0.17471588 -0.5549156777 -0.692495259 -0.58328700
#> qsec   -0.1747158787  1.00000000  0.7445354435 -0.229860862 -0.21268223
#> vs     -0.5549156777  0.74453544  1.0000000000  0.168345125  0.20602335
#> am     -0.6924952588 -0.22986086  0.1683451246  1.000000000  0.79405876
#> gear   -0.5832869965 -0.21268223  0.2060233487  0.794058760  1.00000000
#> carb    0.4276059377 -0.65624923 -0.5696071410  0.057534351  0.27407284
#> mpg.1  -0.0021656960  0.30043441  0.2073668119 -0.075917500 -0.05462384
#> cyl.1   0.1678632295 -0.41178290 -0.2598144531  0.022443808  0.14498421
#> disp.1 -0.1149211638 -0.17007686 -0.0444072266  0.060825730  0.08919764
#> hp.1    0.1781528260  0.04600708  0.0008689411 -0.168511414 -0.02154334
#> drat.1 -0.1564818175 -0.06317186  0.0597832017  0.187310624 -0.02478296
#> wt.1   -0.0337433744  0.27958519  0.2533199233  0.010446887 -0.06871176
#> qsec.1  0.1751487496  0.01656137 -0.1074671963 -0.225847176 -0.07649561
#> vs.1   -0.1301251826 -0.17618324 -0.0275421877  0.245041108  0.02358547
#> am.1   -0.1445944467  0.35781259  0.2848386651  0.067138300 -0.06556994
#> gear.1  0.0413539649  0.11332661  0.0260287979 -0.058313719 -0.10251884
#> carb.1 -0.0007907968 -0.42939953 -0.2814420885  0.433275886  0.50920296
#> mpg.2  -0.1148856845 -0.32222403  0.0301463755  0.262249650  0.25115992
#> cyl.2   0.1503935115 -0.03640281 -0.0826661396 -0.058855582 -0.01396184
#> disp.2 -0.3938102156  0.10211894  0.2485640588  0.048373950 -0.01217339
#> hp.2    0.1122477829  0.02480106 -0.1396911987  0.159577131  0.15663927
#> drat.2  0.2602032241 -0.06801638 -0.1054631476 -0.279801488 -0.41687113
#> wt.2    0.1868386425  0.05529408  0.0088192307 -0.017976208 -0.01501386
#> qsec.2 -0.2597513164 -0.15185238  0.2193599585  0.301949258  0.25800893
#> vs.2    0.1338915166  0.02292831 -0.0671380622 -0.162704408  0.10812293
#> am.2    0.0521382538 -0.06294346 -0.0767629782  0.017222122 -0.14696305
#> gear.2 -0.1973886309 -0.03731372  0.1293870006  0.159276971  0.12522591
#> carb.2  0.0599751305 -0.05942452 -0.0630900186  0.119186099  0.17878877
#> mpg.3  -0.0777139494 -0.03181357 -0.0564241020  0.218248358  0.15425278
#> cyl.3   0.2123085267  0.13508819  0.0336076661 -0.100224470 -0.09349183
#> disp.3 -0.1396789850 -0.16603035  0.0136927975  0.285588134  0.38098387
#> hp.3   -0.1803799604 -0.23803904 -0.1631759455  0.183544509  0.21854655
#> drat.3  0.1032566592 -0.23680717 -0.1455359142 -0.077732484 -0.14796191
#> wt.3    0.0961612790  0.37416550  0.2861039491 -0.108203688 -0.01486571
#> qsec.3  0.1226797102 -0.02869651  0.0197319192 -0.020201962  0.08459760
#> vs.3    0.2511873345 -0.08753350 -0.0734830332 -0.181082628 -0.12908784
#> am.3    0.0272340604  0.01386083 -0.0006565719  0.066527038  0.01581843
#> gear.3  0.2389875537  0.03275011 -0.1702682362 -0.355192418 -0.43278850
#> carb.3 -0.1974109158  0.21307671  0.2013800709  0.009918475 -0.12628115
#>               carb         mpg.1         cyl.1        disp.1          hp.1
#> mpg    -0.55092507  0.1639611458 -0.3001917787 -0.0891710867 -0.0792314979
#> cyl     0.52698829 -0.1305342595  0.2632829532  0.1704097237  0.1108311586
#> disp    0.39497686 -0.0637735955  0.2802026449  0.1047937729 -0.0382094689
#> hp      0.74981247 -0.1623186148  0.4505292112  0.2380416739  0.0321176974
#> drat   -0.09078980  0.1013319872  0.1794235027 -0.1961112823 -0.0559245145
#> wt      0.42760594 -0.0021656960  0.1678632295 -0.1149211638  0.1781528260
#> qsec   -0.65624923  0.3004344085 -0.4117828964 -0.1700768575  0.0460070836
#> vs     -0.56960714  0.2073668119 -0.2598144531 -0.0444072266  0.0008689411
#> am      0.05753435 -0.0759175000  0.0224438077  0.0608257300 -0.1685114141
#> gear    0.27407284 -0.0546238371  0.1449842137  0.0891976446 -0.0215433379
#> carb    1.00000000 -0.0385632052  0.3643116462  0.0398716964  0.1297752630
#> mpg.1  -0.03856321  1.0000000000 -0.0181565927  0.1315267739 -0.1878640243
#> cyl.1   0.36431165 -0.0181565927  1.0000000000  0.2303996291 -0.1867180932
#> disp.1  0.03987170  0.1315267739  0.2303996291  1.0000000000 -0.1870989935
#> hp.1    0.12977526 -0.1878640243 -0.1867180932 -0.1870989935  1.0000000000
#> drat.1 -0.08320699 -0.0329374091  0.2076309587  0.1244186787 -0.0039036674
#> wt.1   -0.26483444  0.1913768850 -0.3204407374  0.0007629529  0.1662027046
#> qsec.1  0.19305372 -0.0588150995  0.1582791970 -0.2996732571  0.2221993287
#> vs.1    0.02185671 -0.2248744523  0.0031894010  0.0247539052  0.0950001430
#> am.1   -0.44493869  0.3051612377 -0.0836274591 -0.1505895944 -0.0114920332
#> gear.1 -0.11042054  0.0364276384 -0.0219065643 -0.3237911074 -0.0998363533
#> carb.1  0.67970512 -0.1041151498  0.0944287315  0.0943336475  0.1823550242
#> mpg.2   0.07069055 -0.5378982066 -0.1659724993 -0.1118770043  0.1320314093
#> cyl.2   0.08154454 -0.0257572883  0.1584478961 -0.1267186953 -0.0137989753
#> disp.2 -0.37275059 -0.0696135211  0.0005693528  0.0235724217 -0.2815871655
#> hp.2    0.12849710  0.1317310478 -0.2173568827  0.0828961722  0.1526796092
#> drat.2 -0.18823191 -0.1670844950 -0.0113024010  0.1331392081  0.0556747907
#> wt.2   -0.03117890  0.2294660120  0.1088273444  0.0110475929  0.0628829783
#> qsec.2  0.10627269 -0.0201419688  0.1584393120  0.2790163840 -0.0302384972
#> vs.2    0.13414220 -0.1932159210  0.0140077755 -0.0809758383  0.2221784546
#> am.2    0.09612263 -0.1046619508  0.0558666311 -0.0655770758 -0.0719197908
#> gear.2 -0.03027754  0.0146173803 -0.2148552999 -0.0094015361  0.0938301504
#> carb.2  0.22857976  0.0377250578  0.0778226540 -0.1411965527  0.2657184449
#> mpg.3   0.03319031  0.0974485179  0.0477674378 -0.1187069549  0.0741777405
#> cyl.3  -0.10068803  0.0485369555  0.0766365965 -0.1227539298 -0.0691171914
#> disp.3  0.30694871  0.4158333821 -0.0866032687  0.3255701460 -0.2830201566
#> hp.3    0.21280001 -0.0009862322  0.1662938653  0.3194305793  0.0911875295
#> drat.3  0.03171613 -0.2337664849 -0.0842242684 -0.0399906743  0.2375025561
#> wt.3   -0.17289623  0.1998522947 -0.2548736347 -0.0794079598  0.3060417857
#> qsec.3  0.09896028 -0.1598858564 -0.1992188722 -0.1433178132  0.1139614825
#> vs.3    0.15267499 -0.3328642487  0.0230448806  0.2265718207  0.1513285335
#> am.3   -0.07490691 -0.3322208652 -0.2266902383 -0.2352892742 -0.1839525344
#> gear.3 -0.21042149  0.1683685567 -0.0153801198 -0.1765483419  0.0657965399
#> carb.3 -0.46516666 -0.1162253900  0.0736509133 -0.0823067853 -0.2642896157
#>              drat.1          wt.1      qsec.1         vs.1        am.1
#> mpg     0.059302398  0.1564186578 -0.11787100 -0.003453393  0.36594352
#> cyl     0.002694454 -0.0971302378  0.11557302 -0.009644146 -0.31125841
#> disp   -0.043228266 -0.1312059432  0.08634945 -0.087275693 -0.21209451
#> hp     -0.013150627 -0.2761479879  0.05729937  0.063391591 -0.32487678
#> drat    0.327411826 -0.0477974310  0.06220324  0.139036048  0.20152985
#> wt     -0.156481817 -0.0337433744  0.17514875 -0.130125183 -0.14459445
#> qsec   -0.063171857  0.2795851863  0.01656137 -0.176183245  0.35781259
#> vs      0.059783202  0.2533199233 -0.10746720 -0.027542188  0.28483867
#> am      0.187310624  0.0104468874 -0.22584718  0.245041108  0.06713830
#> gear   -0.024782958 -0.0687117622 -0.07649561  0.023585472 -0.06556994
#> carb   -0.083206995 -0.2648344410  0.19305372  0.021856706 -0.44493869
#> mpg.1  -0.032937409  0.1913768850 -0.05881510 -0.224874452  0.30516124
#> cyl.1   0.207630959 -0.3204407374  0.15827920  0.003189401 -0.08362746
#> disp.1  0.124418679  0.0007629529 -0.29967326  0.024753905 -0.15058959
#> hp.1   -0.003903667  0.1662027046  0.22219933  0.095000143 -0.01149203
#> drat.1  1.000000000 -0.0798394014  0.19721288  0.156162470  0.07617479
#> wt.1   -0.079839401  1.0000000000 -0.13101823  0.149433577  0.29266368
#> qsec.1  0.197212882 -0.1310182304  1.00000000  0.062978533  0.21719542
#> vs.1    0.156162470  0.1494335773  0.06297853  1.000000000  0.11117255
#> am.1    0.076174785  0.2926636837  0.21719542  0.111172552  1.00000000
#> gear.1 -0.028794693  0.0113037234  0.23624287 -0.010188948  0.26380716
#> carb.1 -0.061446584 -0.1911212113  0.10146456  0.073355528 -0.30179838
#> mpg.2  -0.018822407 -0.0756701138 -0.14978880 -0.046383594 -0.33005272
#> cyl.2  -0.060928511  0.1090853660 -0.19223533 -0.019503343 -0.01024636
#> disp.2  0.236124246 -0.1158661422  0.01891434  0.056770440  0.01469084
#> hp.2   -0.109038282  0.0097985760 -0.13395045 -0.009595901 -0.04386873
#> drat.2  0.112340168 -0.0655681321 -0.08316997  0.205209094 -0.18652709
#> wt.2   -0.290946983  0.0258898272 -0.21316348 -0.141670691  0.12796193
#> qsec.2  0.219085743  0.2062570885 -0.02218588  0.139586348 -0.23363389
#> vs.2   -0.403298942 -0.2734240144  0.11904103 -0.210118865 -0.21502925
#> am.2    0.425173178 -0.1297550428 -0.19825125 -0.031660021 -0.61915497
#> gear.2  0.053323462 -0.0104985683 -0.14512838  0.262859718  0.17672153
#> carb.2  0.057724452  0.0142870615  0.04428041  0.010559958 -0.09965584
#> mpg.3  -0.188010111  0.0817925707 -0.11686275  0.218429016  0.08961228
#> cyl.3   0.033879451  0.0026778584  0.26166023 -0.041797359  0.16953855
#> disp.3 -0.338127748  0.0150622249 -0.30441907 -0.217065531 -0.06168746
#> hp.3    0.196067026 -0.1098486651  0.27106092  0.055790936 -0.02952701
#> drat.3 -0.225993157 -0.2092317060 -0.17593921  0.164111017  0.07424629
#> wt.3   -0.053287255  0.3920922269 -0.02472483 -0.158232244 -0.02269415
#> qsec.3 -0.119191394 -0.0325721365  0.09949070 -0.284982147 -0.11915044
#> vs.3    0.021312672  0.2214082655  0.03959238  0.603301544 -0.12280324
#> am.3   -0.297843484  0.1004689363 -0.10954494 -0.069467512 -0.01153433
#> gear.3  0.027518285  0.1282068999 -0.02458584 -0.107387414  0.07868504
#> carb.3  0.306395377  0.0469298608 -0.06131531 -0.061002536  0.29053530
#>             gear.1        carb.1        mpg.2        cyl.2        disp.2
#> mpg     0.15528215 -0.1389857208  0.022197762 -0.023461366  0.2927847657
#> cyl    -0.06981710  0.0721956908 -0.025877541  0.094191428 -0.2155899867
#> disp    0.01017577 -0.0483061176 -0.056509535  0.152053024 -0.2533698847
#> hp     -0.18073623  0.3313405219  0.006905703  0.129165087 -0.2809862369
#> drat    0.04216641  0.1389952829 -0.012777562  0.024494218  0.2299931957
#> wt      0.04135396 -0.0007907968 -0.114885684  0.150393512 -0.3938102156
#> qsec    0.11332661 -0.4293995265 -0.322224029 -0.036402808  0.1021189434
#> vs      0.02602880 -0.2814420885  0.030146376 -0.082666140  0.2485640588
#> am     -0.05831372  0.4332758859  0.262249650 -0.058855582  0.0483739497
#> gear   -0.10251884  0.5092029586  0.251159919 -0.013961835 -0.0121733878
#> carb   -0.11042054  0.6797051204  0.070690554  0.081544540 -0.3727505911
#> mpg.1   0.03642764 -0.1041151498 -0.537898207 -0.025757288 -0.0696135211
#> cyl.1  -0.02190656  0.0944287315 -0.165972499  0.158447896  0.0005693528
#> disp.1 -0.32379111  0.0943336475 -0.111877004 -0.126718695  0.0235724217
#> hp.1   -0.09983635  0.1823550242  0.132031409 -0.013798975 -0.2815871655
#> drat.1 -0.02879469 -0.0614465839 -0.018822407 -0.060928511  0.2361242458
#> wt.1    0.01130372 -0.1911212113 -0.075670114  0.109085366 -0.1158661422
#> qsec.1  0.23624287  0.1014645638 -0.149788803 -0.192235332  0.0189143398
#> vs.1   -0.01018895  0.0733555275 -0.046383594 -0.019503343  0.0567704403
#> am.1    0.26380716 -0.3017983795 -0.330052718 -0.010246360  0.0146908364
#> gear.1  1.00000000 -0.2522884917 -0.084626388  0.308612357 -0.0790436125
#> carb.1 -0.25228849  1.0000000000  0.170168966 -0.081721689 -0.1881055164
#> mpg.2  -0.08462639  0.1701689663  1.000000000 -0.091342814 -0.0218257970
#> cyl.2   0.30861236 -0.0817216889 -0.091342814  1.000000000  0.0017384146
#> disp.2 -0.07904361 -0.1881055164 -0.021825797  0.001738415  1.0000000000
#> hp.2   -0.44046073  0.2788302322  0.014254948 -0.140831955 -0.3802602347
#> drat.2 -0.04763418 -0.3814928955  0.181111828 -0.260645997 -0.0559860942
#> wt.2    0.01786021  0.0923434744  0.010545493  0.078952290 -0.1166570109
#> qsec.2 -0.19581225  0.1065745784  0.218205608  0.043982012  0.0412050259
#> vs.2    0.05069899  0.0262948858  0.026194546  0.108963079 -0.0526757177
#> am.2   -0.02141065  0.0808027891  0.161667180  0.107730538  0.0954719007
#> gear.2  0.32969161  0.0125621118  0.154511036 -0.063030805 -0.1736826023
#> carb.2 -0.05118846  0.3133800241 -0.026346703 -0.112306460 -0.2521613365
#> mpg.3   0.07826151  0.1102584847  0.032072889 -0.025483221 -0.0134107720
#> cyl.3   0.17076604 -0.1788587030 -0.044154248 -0.334434038  0.0903483436
#> disp.3  0.02698354  0.2917254723  0.110670270  0.028398831 -0.2534103754
#> hp.3   -0.19008390  0.3321947603 -0.107184249 -0.064579049  0.0460185887
#> drat.3  0.04157124  0.0304144807  0.181007517  0.172725388 -0.0708938702
#> wt.3    0.08783937 -0.0407558041  0.098722332  0.086131153 -0.0780668590
#> qsec.3  0.03795551  0.3729250119  0.380776971 -0.121917657 -0.1017128108
#> vs.3   -0.14031700  0.1183542795  0.023510400 -0.022914572 -0.0112502911
#> am.3   -0.09892875  0.0610284883  0.264437494 -0.158568719 -0.0469675497
#> gear.3 -0.39100607 -0.2338269762 -0.217601256  0.012147236  0.0955425636
#> carb.3  0.19323477 -0.4196248182 -0.169327628  0.231234891  0.5567521994
#>                hp.2      drat.2         wt.2       qsec.2        vs.2
#> mpg    -0.073162916 -0.23965363  0.006294095  0.068860762 -0.05587349
#> cyl     0.024241534  0.31054488 -0.023387189 -0.122046887  0.06133899
#> disp    0.063604874  0.36955596  0.119357016 -0.215286499  0.05164889
#> hp      0.138960634  0.10079314 -0.013136456 -0.065645357  0.05308633
#> drat   -0.081459057 -0.31150952 -0.138749568  0.199516912 -0.13338091
#> wt      0.112247783  0.26020322  0.186838642 -0.259751316  0.13389152
#> qsec    0.024801056 -0.06801638  0.055294078 -0.151852385  0.02292831
#> vs     -0.139691199 -0.10546315  0.008819231  0.219359959 -0.06713806
#> am      0.159577131 -0.27980149 -0.017976208  0.301949258 -0.16270441
#> gear    0.156639267 -0.41687113 -0.015013859  0.258008926  0.10812293
#> carb    0.128497098 -0.18823191 -0.031178902  0.106272690  0.13414220
#> mpg.1   0.131731048 -0.16708450  0.229466012 -0.020141969 -0.19321592
#> cyl.1  -0.217356883 -0.01130240  0.108827344  0.158439312  0.01400778
#> disp.1  0.082896172  0.13313921  0.011047593  0.279016384 -0.08097584
#> hp.1    0.152679609  0.05567479  0.062882978 -0.030238497  0.22217845
#> drat.1 -0.109038282  0.11234017 -0.290946983  0.219085743 -0.40329894
#> wt.1    0.009798576 -0.06556813  0.025889827  0.206257088 -0.27342401
#> qsec.1 -0.133950445 -0.08316997 -0.213163482 -0.022185881  0.11904103
#> vs.1   -0.009595901  0.20520909 -0.141670691  0.139586348 -0.21011887
#> am.1   -0.043868732 -0.18652709  0.127961925 -0.233633890 -0.21502925
#> gear.1 -0.440460730 -0.04763418  0.017860213 -0.195812255  0.05069899
#> carb.1  0.278830232 -0.38149290  0.092343474  0.106574578  0.02629489
#> mpg.2   0.014254948  0.18111183  0.010545493  0.218205608  0.02619455
#> cyl.2  -0.140831955 -0.26064600  0.078952290  0.043982012  0.10896308
#> disp.2 -0.380260235 -0.05598609 -0.116657011  0.041205026 -0.05267572
#> hp.2    1.000000000 -0.05772034 -0.011722800  0.053421055 -0.08593714
#> drat.2 -0.057720341  1.00000000 -0.119108255 -0.048400636  0.05244117
#> wt.2   -0.011722800 -0.11910826  1.000000000 -0.256487090  0.27242389
#> qsec.2  0.053421055 -0.04840064 -0.256487090  1.000000000 -0.03715022
#> vs.2   -0.085937140  0.05244117  0.272423890 -0.037150223  1.00000000
#> am.2   -0.131975096  0.25143905 -0.242043133  0.238054003 -0.09531510
#> gear.2 -0.142980652 -0.01588302 -0.159507835 -0.140933023 -0.11494728
#> carb.2  0.100324747 -0.23006811  0.217671842 -0.056293495 -0.09310962
#> mpg.3   0.086473488 -0.16065761  0.121418264 -0.118661962 -0.08418984
#> cyl.3  -0.171426946  0.43287620  0.226416328 -0.363211563  0.11365263
#> disp.3  0.307348527 -0.33925386  0.123246266  0.211881932 -0.07674681
#> hp.3    0.118908581  0.12620778 -0.484663020  0.186866384 -0.04964611
#> drat.3 -0.076531827  0.04829912  0.338580082 -0.040664331  0.23187268
#> wt.3    0.002164409 -0.08077492  0.469564076 -0.213035442 -0.06332248
#> qsec.3  0.237681673 -0.12584249  0.137965443 -0.204048164 -0.03596754
#> vs.3   -0.151336589  0.20555443 -0.094584645  0.138891380  0.04948313
#> am.3    0.129728964 -0.21041923  0.018758688 -0.115397321 -0.02797071
#> gear.3  0.152746387  0.20691672  0.058648042 -0.009593386 -0.02075190
#> carb.3 -0.398066303  0.06271908 -0.050420762 -0.100423196 -0.01242082
#>                am.2       gear.2       carb.2        mpg.3        cyl.3
#> mpg    -0.191007318  0.130273652 -0.029254144 -0.005732779  0.039441619
#> cyl     0.115107334 -0.105498547 -0.069794457 -0.080286176  0.011830073
#> disp    0.044488097 -0.181878963 -0.031144758 -0.197690684  0.209383281
#> hp     -0.037207751 -0.014710492  0.101138498 -0.084248592 -0.028222970
#> drat    0.036698810  0.016594147  0.059507942  0.110492728  0.045186040
#> wt      0.052138254 -0.197388631  0.059975131 -0.077713949  0.212308527
#> qsec   -0.062943460 -0.037313717 -0.059424518 -0.031813567  0.135088187
#> vs     -0.076762978  0.129387001 -0.063090019 -0.056424102  0.033607666
#> am      0.017222122  0.159276971  0.119186099  0.218248358 -0.100224470
#> gear   -0.146963048  0.125225913  0.178788770  0.154252784 -0.093491835
#> carb    0.096122633 -0.030277539  0.228579757  0.033190310 -0.100688032
#> mpg.1  -0.104661951  0.014617380  0.037725058  0.097448518  0.048536955
#> cyl.1   0.055866631 -0.214855300  0.077822654  0.047767438  0.076636596
#> disp.1 -0.065577076 -0.009401536 -0.141196553 -0.118706955 -0.122753930
#> hp.1   -0.071919791  0.093830150  0.265718445  0.074177741 -0.069117191
#> drat.1  0.425173178  0.053323462  0.057724452 -0.188010111  0.033879451
#> wt.1   -0.129755043 -0.010498568  0.014287061  0.081792571  0.002677858
#> qsec.1 -0.198251252 -0.145128381  0.044280408 -0.116862754  0.261660227
#> vs.1   -0.031660021  0.262859718  0.010559958  0.218429016 -0.041797359
#> am.1   -0.619154966  0.176721526 -0.099655842  0.089612283  0.169538549
#> gear.1 -0.021410646  0.329691614 -0.051188459  0.078261505  0.170766044
#> carb.1  0.080802789  0.012562112  0.313380024  0.110258485 -0.178858703
#> mpg.2   0.161667180  0.154511036 -0.026346703  0.032072889 -0.044154248
#> cyl.2   0.107730538 -0.063030805 -0.112306460 -0.025483221 -0.334434038
#> disp.2  0.095471901 -0.173682602 -0.252161336 -0.013410772  0.090348344
#> hp.2   -0.131975096 -0.142980652  0.100324747  0.086473488 -0.171426946
#> drat.2  0.251439051 -0.015883017 -0.230068112 -0.160657613  0.432876203
#> wt.2   -0.242043133 -0.159507835  0.217671842  0.121418264  0.226416328
#> qsec.2  0.238054003 -0.140933023 -0.056293495 -0.118661962 -0.363211563
#> vs.2   -0.095315100 -0.114947276 -0.093109623 -0.084189835  0.113652627
#> am.2    1.000000000 -0.041360370 -0.005307659  0.062366756 -0.132207699
#> gear.2 -0.041360370  1.000000000  0.275847942  0.242284781 -0.263769599
#> carb.2 -0.005307659  0.275847942  1.000000000 -0.044311568 -0.031928337
#> mpg.3   0.062366756  0.242284781 -0.044311568  1.000000000 -0.189785677
#> cyl.3  -0.132207699 -0.263769599 -0.031928337 -0.189785677  1.000000000
#> disp.3 -0.212837865  0.138879700 -0.170167601  0.149662194 -0.347487494
#> hp.3    0.062709415 -0.133298706 -0.357428895  0.044322664 -0.129025912
#> drat.3 -0.208779766 -0.045713009 -0.169840630  0.008571660 -0.089300427
#> wt.3   -0.008947495 -0.066616592  0.310197983  0.078297779  0.219404357
#> qsec.3 -0.018759647  0.069353762  0.109117240  0.054329352  0.042720044
#> vs.3    0.102049948  0.074661749 -0.062218878  0.112860311  0.006210147
#> am.3   -0.095233605 -0.047138364 -0.125253003  0.403304284 -0.117825133
#> gear.3  0.008978255 -0.304007008  0.085571196 -0.242957476 -0.039892884
#> carb.3  0.059094801 -0.261009202 -0.557249934 -0.090238775  0.239491579
#>              disp.3          hp.3       drat.3         wt.3       qsec.3
#> mpg     0.071689257  0.0008116447 -0.106974009  0.148708559 -0.051862255
#> cyl    -0.048073850  0.1092101178  0.109931709 -0.168379495  0.009597514
#> disp   -0.089779654 -0.0705180660  0.114438816 -0.049696288  0.060669350
#> hp      0.089926849  0.1230086092  0.141169316 -0.255130218  0.097826851
#> drat    0.018302832  0.2315273033 -0.146257286 -0.053508037 -0.218983124
#> wt     -0.139678985 -0.1803799604  0.103256659  0.096161279  0.122679710
#> qsec   -0.166030352 -0.2380390439 -0.236807166  0.374165499 -0.028696515
#> vs      0.013692798 -0.1631759455 -0.145535914  0.286103949  0.019731919
#> am      0.285588134  0.1835445085 -0.077732484 -0.108203688 -0.020201962
#> gear    0.380983873  0.2185465456 -0.147961911 -0.014865711  0.084597597
#> carb    0.306948710  0.2128000064  0.031716129 -0.172896231  0.098960281
#> mpg.1   0.415833382 -0.0009862322 -0.233766485  0.199852295 -0.159885856
#> cyl.1  -0.086603269  0.1662938653 -0.084224268 -0.254873635 -0.199218872
#> disp.1  0.325570146  0.3194305793 -0.039990674 -0.079407960 -0.143317813
#> hp.1   -0.283020157  0.0911875295  0.237502556  0.306041786  0.113961483
#> drat.1 -0.338127748  0.1960670259 -0.225993157 -0.053287255 -0.119191394
#> wt.1    0.015062225 -0.1098486651 -0.209231706  0.392092227 -0.032572137
#> qsec.1 -0.304419074  0.2710609157 -0.175939210 -0.024724827  0.099490701
#> vs.1   -0.217065531  0.0557909362  0.164111017 -0.158232244 -0.284982147
#> am.1   -0.061687464 -0.0295270092  0.074246292 -0.022694148 -0.119150441
#> gear.1  0.026983543 -0.1900838966  0.041571240  0.087839371  0.037955506
#> carb.1  0.291725472  0.3321947603  0.030414481 -0.040755804  0.372925012
#> mpg.2   0.110670270 -0.1071842493  0.181007517  0.098722332  0.380776971
#> cyl.2   0.028398831 -0.0645790491  0.172725388  0.086131153 -0.121917657
#> disp.2 -0.253410375  0.0460185887 -0.070893870 -0.078066859 -0.101712811
#> hp.2    0.307348527  0.1189085815 -0.076531827  0.002164409  0.237681673
#> drat.2 -0.339253861  0.1262077830  0.048299121 -0.080774923 -0.125842495
#> wt.2    0.123246266 -0.4846630200  0.338580082  0.469564076  0.137965443
#> qsec.2  0.211881932  0.1868663845 -0.040664331 -0.213035442 -0.204048164
#> vs.2   -0.076746809 -0.0496461060  0.231872679 -0.063322484 -0.035967542
#> am.2   -0.212837865  0.0627094146 -0.208779766 -0.008947495 -0.018759647
#> gear.2  0.138879700 -0.1332987056 -0.045713009 -0.066616592  0.069353762
#> carb.2 -0.170167601 -0.3574288952 -0.169840630  0.310197983  0.109117240
#> mpg.3   0.149662194  0.0443226641  0.008571660  0.078297779  0.054329352
#> cyl.3  -0.347487494 -0.1290259123 -0.089300427  0.219404357  0.042720044
#> disp.3  1.000000000  0.1176546412 -0.004937711 -0.023454965  0.070371032
#> hp.3    0.117654641  1.0000000000 -0.073077098 -0.391181727 -0.049615158
#> drat.3 -0.004937711 -0.0730770984  1.000000000 -0.236008615 -0.014976010
#> wt.3   -0.023454965 -0.3911817269 -0.236008615  1.000000000  0.269189523
#> qsec.3  0.070371032 -0.0496151585 -0.014976010  0.269189523  1.000000000
#> vs.3   -0.258534113  0.0048468102  0.123101124  0.004080869 -0.077905903
#> am.3    0.052141692 -0.1761461319  0.096178972  0.023533440  0.475949636
#> gear.3 -0.314780802  0.0297531760  0.084261894 -0.126207702 -0.177321163
#> carb.3 -0.309419838  0.0737365603  0.091187531 -0.189165426 -0.183518027
#>                vs.3          am.3       gear.3       carb.3
#> mpg    -0.428910477 -0.1136072418 -0.232778868  0.242833877
#> cyl     0.205518527 -0.0934093823  0.199263207 -0.163843910
#> disp    0.164577019 -0.0454835031  0.283097394 -0.102134001
#> hp      0.209958233 -0.0059392450  0.046949787 -0.266491685
#> drat   -0.255653953 -0.1510835270 -0.162914929  0.257226800
#> wt      0.251187335  0.0272340604  0.238987554 -0.197410916
#> qsec   -0.087533497  0.0138608316  0.032750109  0.213076711
#> vs     -0.073483033 -0.0006565719 -0.170268236  0.201380071
#> am     -0.181082628  0.0665270380 -0.355192418  0.009918475
#> gear   -0.129087845  0.0158184257 -0.432788500 -0.126281146
#> carb    0.152674989 -0.0749069123 -0.210421488 -0.465166655
#> mpg.1  -0.332864249 -0.3322208652  0.168368557 -0.116225390
#> cyl.1   0.023044881 -0.2266902383 -0.015380120  0.073650913
#> disp.1  0.226571821 -0.2352892742 -0.176548342 -0.082306785
#> hp.1    0.151328534 -0.1839525344  0.065796540 -0.264289616
#> drat.1  0.021312672 -0.2978434844  0.027518285  0.306395377
#> wt.1    0.221408265  0.1004689363  0.128206900  0.046929861
#> qsec.1  0.039592384 -0.1095449370 -0.024585844 -0.061315307
#> vs.1    0.603301544 -0.0694675119 -0.107387414 -0.061002536
#> am.1   -0.122803237 -0.0115343312  0.078685044  0.290535298
#> gear.1 -0.140316996 -0.0989287486 -0.391006074  0.193234774
#> carb.1  0.118354280  0.0610284883 -0.233826976 -0.419624818
#> mpg.2   0.023510400  0.2644374940 -0.217601256 -0.169327628
#> cyl.2  -0.022914572 -0.1585687188  0.012147236  0.231234891
#> disp.2 -0.011250291 -0.0469675497  0.095542564  0.556752199
#> hp.2   -0.151336589  0.1297289642  0.152746387 -0.398066303
#> drat.2  0.205554434 -0.2104192328  0.206916719  0.062719076
#> wt.2   -0.094584645  0.0187586882  0.058648042 -0.050420762
#> qsec.2  0.138891380 -0.1153973208 -0.009593386 -0.100423196
#> vs.2    0.049483131 -0.0279707100 -0.020751898 -0.012420823
#> am.2    0.102049948 -0.0952336048  0.008978255  0.059094801
#> gear.2  0.074661749 -0.0471383638 -0.304007008 -0.261009202
#> carb.2 -0.062218878 -0.1252530027  0.085571196 -0.557249934
#> mpg.3   0.112860311  0.4033042841 -0.242957476 -0.090238775
#> cyl.3   0.006210147 -0.1178251326 -0.039892884  0.239491579
#> disp.3 -0.258534113  0.0521416920 -0.314780802 -0.309419838
#> hp.3    0.004846810 -0.1761461319  0.029753176  0.073736560
#> drat.3  0.123101124  0.0961789722  0.084261894  0.091187531
#> wt.3    0.004080869  0.0235334403 -0.126207702 -0.189165426
#> qsec.3 -0.077905903  0.4759496362 -0.177321163 -0.183518027
#> vs.3    1.000000000  0.1478576852 -0.115211764 -0.089425945
#> am.3    0.147857685  1.0000000000 -0.063168024  0.034463262
#> gear.3 -0.115211764 -0.0631680242  1.000000000  0.141983574
#> carb.3 -0.089425945  0.0344632622  0.141983574  1.000000000