Smooth a non-positive definite correlation matrix to make it positive definite

Make correlations positive definite using psych::cor.smooth. Note that this modifies the correlation values, and does not recompute the remaining indices (p-values, confidence intervals, etc.). Thus, once you used cor_smooth(), any other indices becomes irrelevant.

cor_smooth(x, method = "psych", verbose = TRUE, ...)

is.positive_definite(x, tol = .Machine$double.eps, ...)

Arguments

x
A correlation matrix.

method
Smoothing method. Can be psych (will use psych::cor.smooth()), hj (Jorjani et al., 2003) or lrs (Schaeffer, 2014). For the two last, will use mbend::bend() (check its documentation for details).

verbose
Set to FALSE to silence the function.

...
Other arguments to be passed to or from other functions.

tol
The minimum eigen values to be considered as acceptable.

Examples

set.seed(1)
data <- cbind(mtcars,
              mtcars * matrix(rnorm(32 * 11, sd = 0.05), ncol=11),
              mtcars * matrix(rnorm(32 * 11, sd = 0.05), ncol=11),
              mtcars * -matrix(rnorm(32 * 11, sd = 0.01), ncol=11))
x <- correlation(data)
is.positive_definite(x)
#> [1] FALSE

smoothed <- cor_smooth(x)
#> mpg - cyl: -0.85 -> -0.85
#> mpg - disp: -0.85 -> -0.85
#> mpg - hp: -0.78 -> -0.78
#> mpg - drat: 0.68 -> 0.68
#> mpg - wt: -0.87 -> -0.87
#> mpg - qsec: 0.42 -> 0.42
#> mpg - vs: 0.66 -> 0.66
#> mpg - am: 0.60 -> 0.60
#> mpg - gear: 0.48 -> 0.48
#> mpg - carb: -0.55 -> -0.55
#> mpg - mpg.1: 0.16 -> 0.16
#> mpg - cyl.1: -0.30 -> -0.30
#> mpg - disp.1: -0.09 -> -0.09
#> mpg - hp.1: -0.08 -> -0.08
#> mpg - drat.1: 0.06 -> 0.06
#> mpg - wt.1: 0.16 -> 0.16
#> mpg - qsec.1: -0.12 -> -0.12
#> mpg - vs.1: -3.45e-03 -> -3.45e-03
#> mpg - am.1: 0.37 -> 0.37
#> mpg - gear.1: 0.16 -> 0.16
#> mpg - carb.1: -0.14 -> -0.14
#> mpg - mpg.2: 0.02 -> 0.02
#> mpg - cyl.2: -0.02 -> -0.02
#> mpg - disp.2: 0.29 -> 0.29
#> mpg - hp.2: -0.07 -> -0.07
#> mpg - drat.2: -0.24 -> -0.24
#> mpg - wt.2: 6.29e-03 -> 6.29e-03
#> mpg - qsec.2: 0.07 -> 0.07
#> mpg - vs.2: -0.06 -> -0.06
#> mpg - am.2: -0.19 -> -0.19
#> mpg - gear.2: 0.13 -> 0.13
#> mpg - carb.2: -0.03 -> -0.03
#> mpg - mpg.3: -5.73e-03 -> -5.73e-03
#> mpg - cyl.3: 0.04 -> 0.04
#> mpg - disp.3: 0.07 -> 0.07
#> mpg - hp.3: 8.12e-04 -> 8.12e-04
#> mpg - drat.3: -0.11 -> -0.11
#> mpg - wt.3: 0.15 -> 0.15
#> mpg - qsec.3: -0.05 -> -0.05
#> mpg - vs.3: -0.43 -> -0.43
#> mpg - am.3: -0.11 -> -0.11
#> mpg - gear.3: -0.23 -> -0.23
#> mpg - carb.3: 0.24 -> 0.24
#> cyl - disp: 0.90 -> 0.90
#> cyl - hp: 0.83 -> 0.83
#> cyl - drat: -0.70 -> -0.70
#> cyl - wt: 0.78 -> 0.78
#> cyl - qsec: -0.59 -> -0.59
#> cyl - vs: -0.81 -> -0.81
#> cyl - am: -0.52 -> -0.52
#> cyl - gear: -0.49 -> -0.49
#> cyl - carb: 0.53 -> 0.53
#> cyl - mpg.1: -0.13 -> -0.13
#> cyl - cyl.1: 0.26 -> 0.26
#> cyl - disp.1: 0.17 -> 0.17
#> cyl - hp.1: 0.11 -> 0.11
#> cyl - drat.1: 2.69e-03 -> 2.69e-03
#> cyl - wt.1: -0.10 -> -0.10
#> cyl - qsec.1: 0.12 -> 0.12
#> cyl - vs.1: -9.64e-03 -> -9.64e-03
#> cyl - am.1: -0.31 -> -0.31
#> cyl - gear.1: -0.07 -> -0.07
#> cyl - carb.1: 0.07 -> 0.07
#> cyl - mpg.2: -0.03 -> -0.03
#> cyl - cyl.2: 0.09 -> 0.09
#> cyl - disp.2: -0.22 -> -0.22
#> cyl - hp.2: 0.02 -> 0.02
#> cyl - drat.2: 0.31 -> 0.31
#> cyl - wt.2: -0.02 -> -0.02
#> cyl - qsec.2: -0.12 -> -0.12
#> cyl - vs.2: 0.06 -> 0.06
#> cyl - am.2: 0.12 -> 0.12
#> cyl - gear.2: -0.11 -> -0.11
#> cyl - carb.2: -0.07 -> -0.07
#> cyl - mpg.3: -0.08 -> -0.08
#> cyl - cyl.3: 0.01 -> 0.01
#> cyl - disp.3: -0.05 -> -0.05
#> cyl - hp.3: 0.11 -> 0.11
#> cyl - drat.3: 0.11 -> 0.11
#> cyl - wt.3: -0.17 -> -0.17
#> cyl - qsec.3: 9.60e-03 -> 9.60e-03
#> cyl - vs.3: 0.21 -> 0.21
#> cyl - am.3: -0.09 -> -0.09
#> cyl - gear.3: 0.20 -> 0.20
#> cyl - carb.3: -0.16 -> -0.16
#> disp - hp: 0.79 -> 0.79
#> disp - drat: -0.71 -> -0.71
#> disp - wt: 0.89 -> 0.89
#> disp - qsec: -0.43 -> -0.43
#> disp - vs: -0.71 -> -0.71
#> disp - am: -0.59 -> -0.59
#> disp - gear: -0.56 -> -0.56
#> disp - carb: 0.39 -> 0.39
#> disp - mpg.1: -0.06 -> -0.06
#> disp - cyl.1: 0.28 -> 0.28
#> disp - disp.1: 0.10 -> 0.10
#> disp - hp.1: -0.04 -> -0.04
#> disp - drat.1: -0.04 -> -0.04
#> disp - wt.1: -0.13 -> -0.13
#> disp - qsec.1: 0.09 -> 0.09
#> disp - vs.1: -0.09 -> -0.09
#> disp - am.1: -0.21 -> -0.21
#> disp - gear.1: 0.01 -> 0.01
#> disp - carb.1: -0.05 -> -0.05
#> disp - mpg.2: -0.06 -> -0.06
#> disp - cyl.2: 0.15 -> 0.15
#> disp - disp.2: -0.25 -> -0.25
#> disp - hp.2: 0.06 -> 0.06
#> disp - drat.2: 0.37 -> 0.37
#> disp - wt.2: 0.12 -> 0.12
#> disp - qsec.2: -0.22 -> -0.22
#> disp - vs.2: 0.05 -> 0.05
#> disp - am.2: 0.04 -> 0.04
#> disp - gear.2: -0.18 -> -0.18
#> disp - carb.2: -0.03 -> -0.03
#> disp - mpg.3: -0.20 -> -0.20
#> disp - cyl.3: 0.21 -> 0.21
#> disp - disp.3: -0.09 -> -0.09
#> disp - hp.3: -0.07 -> -0.07
#> disp - drat.3: 0.11 -> 0.11
#> disp - wt.3: -0.05 -> -0.05
#> disp - qsec.3: 0.06 -> 0.06
#> disp - vs.3: 0.16 -> 0.16
#> disp - am.3: -0.05 -> -0.05
#> disp - gear.3: 0.28 -> 0.28
#> disp - carb.3: -0.10 -> -0.10
#> hp - drat: -0.45 -> -0.45
#> hp - wt: 0.66 -> 0.66
#> hp - qsec: -0.71 -> -0.71
#> hp - vs: -0.72 -> -0.72
#> hp - am: -0.24 -> -0.24
#> hp - gear: -0.13 -> -0.13
#> hp - carb: 0.75 -> 0.75
#> hp - mpg.1: -0.16 -> -0.16
#> hp - cyl.1: 0.45 -> 0.45
#> hp - disp.1: 0.24 -> 0.24
#> hp - hp.1: 0.03 -> 0.03
#> hp - drat.1: -0.01 -> -0.01
#> hp - wt.1: -0.28 -> -0.28
#> hp - qsec.1: 0.06 -> 0.06
#> hp - vs.1: 0.06 -> 0.06
#> hp - am.1: -0.32 -> -0.32
#> hp - gear.1: -0.18 -> -0.18
#> hp - carb.1: 0.33 -> 0.33
#> hp - mpg.2: 6.91e-03 -> 6.91e-03
#> hp - cyl.2: 0.13 -> 0.13
#> hp - disp.2: -0.28 -> -0.28
#> hp - hp.2: 0.14 -> 0.14
#> hp - drat.2: 0.10 -> 0.10
#> hp - wt.2: -0.01 -> -0.01
#> hp - qsec.2: -0.07 -> -0.07
#> hp - vs.2: 0.05 -> 0.05
#> hp - am.2: -0.04 -> -0.04
#> hp - gear.2: -0.01 -> -0.01
#> hp - carb.2: 0.10 -> 0.10
#> hp - mpg.3: -0.08 -> -0.08
#> hp - cyl.3: -0.03 -> -0.03
#> hp - disp.3: 0.09 -> 0.09
#> hp - hp.3: 0.12 -> 0.12
#> hp - drat.3: 0.14 -> 0.14
#> hp - wt.3: -0.26 -> -0.26
#> hp - qsec.3: 0.10 -> 0.10
#> hp - vs.3: 0.21 -> 0.21
#> hp - am.3: -5.94e-03 -> -5.94e-03
#> hp - gear.3: 0.05 -> 0.05
#> hp - carb.3: -0.27 -> -0.27
#> drat - wt: -0.71 -> -0.71
#> drat - qsec: 0.09 -> 0.09
#> drat - vs: 0.44 -> 0.44
#> drat - am: 0.71 -> 0.71
#> drat - gear: 0.70 -> 0.70
#> drat - carb: -0.09 -> -0.09
#> drat - mpg.1: 0.10 -> 0.10
#> drat - cyl.1: 0.18 -> 0.18
#> drat - disp.1: -0.20 -> -0.20
#> drat - hp.1: -0.06 -> -0.06
#> drat - drat.1: 0.33 -> 0.33
#> drat - wt.1: -0.05 -> -0.05
#> drat - qsec.1: 0.06 -> 0.06
#> drat - vs.1: 0.14 -> 0.14
#> drat - am.1: 0.20 -> 0.20
#> drat - gear.1: 0.04 -> 0.04
#> drat - carb.1: 0.14 -> 0.14
#> drat - mpg.2: -0.01 -> -0.01
#> drat - cyl.2: 0.02 -> 0.02
#> drat - disp.2: 0.23 -> 0.23
#> drat - hp.2: -0.08 -> -0.08
#> drat - drat.2: -0.31 -> -0.31
#> drat - wt.2: -0.14 -> -0.14
#> drat - qsec.2: 0.20 -> 0.20
#> drat - vs.2: -0.13 -> -0.13
#> drat - am.2: 0.04 -> 0.04
#> drat - gear.2: 0.02 -> 0.02
#> drat - carb.2: 0.06 -> 0.06
#> drat - mpg.3: 0.11 -> 0.11
#> drat - cyl.3: 0.05 -> 0.05
#> drat - disp.3: 0.02 -> 0.02
#> drat - hp.3: 0.23 -> 0.23
#> drat - drat.3: -0.15 -> -0.15
#> drat - wt.3: -0.05 -> -0.05
#> drat - qsec.3: -0.22 -> -0.22
#> drat - vs.3: -0.26 -> -0.26
#> drat - am.3: -0.15 -> -0.15
#> drat - gear.3: -0.16 -> -0.16
#> drat - carb.3: 0.26 -> 0.26
#> wt - qsec: -0.17 -> -0.17
#> wt - vs: -0.55 -> -0.55
#> wt - am: -0.69 -> -0.69
#> wt - gear: -0.58 -> -0.58
#> wt - carb: 0.43 -> 0.43
#> wt - mpg.1: -2.17e-03 -> -2.17e-03
#> wt - cyl.1: 0.17 -> 0.17
#> wt - disp.1: -0.11 -> -0.11
#> wt - hp.1: 0.18 -> 0.18
#> wt - drat.1: -0.16 -> -0.16
#> wt - wt.1: -0.03 -> -0.03
#> wt - qsec.1: 0.18 -> 0.18
#> wt - vs.1: -0.13 -> -0.13
#> wt - am.1: -0.14 -> -0.14
#> wt - gear.1: 0.04 -> 0.04
#> wt - carb.1: -7.91e-04 -> -7.91e-04
#> wt - mpg.2: -0.11 -> -0.11
#> wt - cyl.2: 0.15 -> 0.15
#> wt - disp.2: -0.39 -> -0.39
#> wt - hp.2: 0.11 -> 0.11
#> wt - drat.2: 0.26 -> 0.26
#> wt - wt.2: 0.19 -> 0.19
#> wt - qsec.2: -0.26 -> -0.26
#> wt - vs.2: 0.13 -> 0.13
#> wt - am.2: 0.05 -> 0.05
#> wt - gear.2: -0.20 -> -0.20
#> wt - carb.2: 0.06 -> 0.06
#> wt - mpg.3: -0.08 -> -0.08
#> wt - cyl.3: 0.21 -> 0.21
#> wt - disp.3: -0.14 -> -0.14
#> wt - hp.3: -0.18 -> -0.18
#> wt - drat.3: 0.10 -> 0.10
#> wt - wt.3: 0.10 -> 0.10
#> wt - qsec.3: 0.12 -> 0.12
#> wt - vs.3: 0.25 -> 0.25
#> wt - am.3: 0.03 -> 0.03
#> wt - gear.3: 0.24 -> 0.24
#> wt - carb.3: -0.20 -> -0.20
#> qsec - vs: 0.74 -> 0.74
#> qsec - am: -0.23 -> -0.23
#> qsec - gear: -0.21 -> -0.21
#> qsec - carb: -0.66 -> -0.66
#> qsec - mpg.1: 0.30 -> 0.30
#> qsec - cyl.1: -0.41 -> -0.41
#> qsec - disp.1: -0.17 -> -0.17
#> qsec - hp.1: 0.05 -> 0.05
#> qsec - drat.1: -0.06 -> -0.06
#> qsec - wt.1: 0.28 -> 0.28
#> qsec - qsec.1: 0.02 -> 0.02
#> qsec - vs.1: -0.18 -> -0.18
#> qsec - am.1: 0.36 -> 0.36
#> qsec - gear.1: 0.11 -> 0.11
#> qsec - carb.1: -0.43 -> -0.43
#> qsec - mpg.2: -0.32 -> -0.32
#> qsec - cyl.2: -0.04 -> -0.04
#> qsec - disp.2: 0.10 -> 0.10
#> qsec - hp.2: 0.02 -> 0.02
#> qsec - drat.2: -0.07 -> -0.07
#> qsec - wt.2: 0.06 -> 0.06
#> qsec - qsec.2: -0.15 -> -0.15
#> qsec - vs.2: 0.02 -> 0.02
#> qsec - am.2: -0.06 -> -0.06
#> qsec - gear.2: -0.04 -> -0.04
#> qsec - carb.2: -0.06 -> -0.06
#> qsec - mpg.3: -0.03 -> -0.03
#> qsec - cyl.3: 0.14 -> 0.14
#> qsec - disp.3: -0.17 -> -0.17
#> qsec - hp.3: -0.24 -> -0.24
#> qsec - drat.3: -0.24 -> -0.24
#> qsec - wt.3: 0.37 -> 0.37
#> qsec - qsec.3: -0.03 -> -0.03
#> qsec - vs.3: -0.09 -> -0.09
#> qsec - am.3: 0.01 -> 0.01
#> qsec - gear.3: 0.03 -> 0.03
#> qsec - carb.3: 0.21 -> 0.21
#> vs - am: 0.17 -> 0.17
#> vs - gear: 0.21 -> 0.21
#> vs - carb: -0.57 -> -0.57
#> vs - mpg.1: 0.21 -> 0.21
#> vs - cyl.1: -0.26 -> -0.26
#> vs - disp.1: -0.04 -> -0.04
#> vs - hp.1: 8.69e-04 -> 8.69e-04
#> vs - drat.1: 0.06 -> 0.06
#> vs - wt.1: 0.25 -> 0.25
#> vs - qsec.1: -0.11 -> -0.11
#> vs - vs.1: -0.03 -> -0.03
#> vs - am.1: 0.28 -> 0.28
#> vs - gear.1: 0.03 -> 0.03
#> vs - carb.1: -0.28 -> -0.28
#> vs - mpg.2: 0.03 -> 0.03
#> vs - cyl.2: -0.08 -> -0.08
#> vs - disp.2: 0.25 -> 0.25
#> vs - hp.2: -0.14 -> -0.14
#> vs - drat.2: -0.11 -> -0.11
#> vs - wt.2: 8.82e-03 -> 8.82e-03
#> vs - qsec.2: 0.22 -> 0.22
#> vs - vs.2: -0.07 -> -0.07
#> vs - am.2: -0.08 -> -0.08
#> vs - gear.2: 0.13 -> 0.13
#> vs - carb.2: -0.06 -> -0.06
#> vs - mpg.3: -0.06 -> -0.06
#> vs - cyl.3: 0.03 -> 0.03
#> vs - disp.3: 0.01 -> 0.01
#> vs - hp.3: -0.16 -> -0.16
#> vs - drat.3: -0.15 -> -0.15
#> vs - wt.3: 0.29 -> 0.29
#> vs - qsec.3: 0.02 -> 0.02
#> vs - vs.3: -0.07 -> -0.07
#> vs - am.3: -6.57e-04 -> -6.57e-04
#> vs - gear.3: -0.17 -> -0.17
#> vs - carb.3: 0.20 -> 0.20
#> am - gear: 0.79 -> 0.79
#> am - carb: 0.06 -> 0.06
#> am - mpg.1: -0.08 -> -0.08
#> am - cyl.1: 0.02 -> 0.02
#> am - disp.1: 0.06 -> 0.06
#> am - hp.1: -0.17 -> -0.17
#> am - drat.1: 0.19 -> 0.19
#> am - wt.1: 0.01 -> 0.01
#> am - qsec.1: -0.23 -> -0.23
#> am - vs.1: 0.25 -> 0.25
#> am - am.1: 0.07 -> 0.07
#> am - gear.1: -0.06 -> -0.06
#> am - carb.1: 0.43 -> 0.43
#> am - mpg.2: 0.26 -> 0.26
#> am - cyl.2: -0.06 -> -0.06
#> am - disp.2: 0.05 -> 0.05
#> am - hp.2: 0.16 -> 0.16
#> am - drat.2: -0.28 -> -0.28
#> am - wt.2: -0.02 -> -0.02
#> am - qsec.2: 0.30 -> 0.30
#> am - vs.2: -0.16 -> -0.16
#> am - am.2: 0.02 -> 0.02
#> am - gear.2: 0.16 -> 0.16
#> am - carb.2: 0.12 -> 0.12
#> am - mpg.3: 0.22 -> 0.22
#> am - cyl.3: -0.10 -> -0.10
#> am - disp.3: 0.29 -> 0.29
#> am - hp.3: 0.18 -> 0.18
#> am - drat.3: -0.08 -> -0.08
#> am - wt.3: -0.11 -> -0.11
#> am - qsec.3: -0.02 -> -0.02
#> am - vs.3: -0.18 -> -0.18
#> am - am.3: 0.07 -> 0.07
#> am - gear.3: -0.36 -> -0.36
#> am - carb.3: 9.92e-03 -> 9.92e-03
#> gear - carb: 0.27 -> 0.27
#> gear - mpg.1: -0.05 -> -0.05
#> gear - cyl.1: 0.14 -> 0.14
#> gear - disp.1: 0.09 -> 0.09
#> gear - hp.1: -0.02 -> -0.02
#> gear - drat.1: -0.02 -> -0.02
#> gear - wt.1: -0.07 -> -0.07
#> gear - qsec.1: -0.08 -> -0.08
#> gear - vs.1: 0.02 -> 0.02
#> gear - am.1: -0.07 -> -0.07
#> gear - gear.1: -0.10 -> -0.10
#> gear - carb.1: 0.51 -> 0.51
#> gear - mpg.2: 0.25 -> 0.25
#> gear - cyl.2: -0.01 -> -0.01
#> gear - disp.2: -0.01 -> -0.01
#> gear - hp.2: 0.16 -> 0.16
#> gear - drat.2: -0.42 -> -0.42
#> gear - wt.2: -0.02 -> -0.02
#> gear - qsec.2: 0.26 -> 0.26
#> gear - vs.2: 0.11 -> 0.11
#> gear - am.2: -0.15 -> -0.15
#> gear - gear.2: 0.13 -> 0.13
#> gear - carb.2: 0.18 -> 0.18
#> gear - mpg.3: 0.15 -> 0.15
#> gear - cyl.3: -0.09 -> -0.09
#> gear - disp.3: no change (0.38)
#> gear - hp.3: 0.22 -> 0.22
#> gear - drat.3: -0.15 -> -0.15
#> gear - wt.3: -0.01 -> -0.01
#> gear - qsec.3: 0.08 -> 0.08
#> gear - vs.3: -0.13 -> -0.13
#> gear - am.3: 0.02 -> 0.02
#> gear - gear.3: -0.43 -> -0.43
#> gear - carb.3: -0.13 -> -0.13
#> carb - mpg.1: -0.04 -> -0.04
#> carb - cyl.1: 0.36 -> 0.36
#> carb - disp.1: 0.04 -> 0.04
#> carb - hp.1: 0.13 -> 0.13
#> carb - drat.1: -0.08 -> -0.08
#> carb - wt.1: -0.26 -> -0.26
#> carb - qsec.1: 0.19 -> 0.19
#> carb - vs.1: 0.02 -> 0.02
#> carb - am.1: -0.44 -> -0.44
#> carb - gear.1: -0.11 -> -0.11
#> carb - carb.1: 0.68 -> 0.68
#> carb - mpg.2: 0.07 -> 0.07
#> carb - cyl.2: 0.08 -> 0.08
#> carb - disp.2: -0.37 -> -0.37
#> carb - hp.2: 0.13 -> 0.13
#> carb - drat.2: -0.19 -> -0.19
#> carb - wt.2: -0.03 -> -0.03
#> carb - qsec.2: 0.11 -> 0.11
#> carb - vs.2: 0.13 -> 0.13
#> carb - am.2: 0.10 -> 0.10
#> carb - gear.2: -0.03 -> -0.03
#> carb - carb.2: 0.23 -> 0.23
#> carb - mpg.3: 0.03 -> 0.03
#> carb - cyl.3: -0.10 -> -0.10
#> carb - disp.3: 0.31 -> 0.31
#> carb - hp.3: 0.21 -> 0.21
#> carb - drat.3: 0.03 -> 0.03
#> carb - wt.3: -0.17 -> -0.17
#> carb - qsec.3: 0.10 -> 0.10
#> carb - vs.3: 0.15 -> 0.15
#> carb - am.3: -0.07 -> -0.07
#> carb - gear.3: -0.21 -> -0.21
#> carb - carb.3: -0.47 -> -0.47
#> mpg.1 - cyl.1: -0.02 -> -0.02
#> mpg.1 - disp.1: 0.13 -> 0.13
#> mpg.1 - hp.1: -0.19 -> -0.19
#> mpg.1 - drat.1: -0.03 -> -0.03
#> mpg.1 - wt.1: 0.19 -> 0.19
#> mpg.1 - qsec.1: -0.06 -> -0.06
#> mpg.1 - vs.1: -0.22 -> -0.22
#> mpg.1 - am.1: 0.31 -> 0.31
#> mpg.1 - gear.1: 0.04 -> 0.04
#> mpg.1 - carb.1: -0.10 -> -0.10
#> mpg.1 - mpg.2: -0.54 -> -0.54
#> mpg.1 - cyl.2: -0.03 -> -0.03
#> mpg.1 - disp.2: -0.07 -> -0.07
#> mpg.1 - hp.2: 0.13 -> 0.13
#> mpg.1 - drat.2: -0.17 -> -0.17
#> mpg.1 - wt.2: 0.23 -> 0.23
#> mpg.1 - qsec.2: -0.02 -> -0.02
#> mpg.1 - vs.2: -0.19 -> -0.19
#> mpg.1 - am.2: -0.10 -> -0.10
#> mpg.1 - gear.2: 0.01 -> 0.01
#> mpg.1 - carb.2: 0.04 -> 0.04
#> mpg.1 - mpg.3: 0.10 -> 0.10
#> mpg.1 - cyl.3: 0.05 -> 0.05
#> mpg.1 - disp.3: 0.42 -> 0.42
#> mpg.1 - hp.3: -9.86e-04 -> -9.86e-04
#> mpg.1 - drat.3: -0.23 -> -0.23
#> mpg.1 - wt.3: 0.20 -> 0.20
#> mpg.1 - qsec.3: -0.16 -> -0.16
#> mpg.1 - vs.3: -0.33 -> -0.33
#> mpg.1 - am.3: -0.33 -> -0.33
#> mpg.1 - gear.3: 0.17 -> 0.17
#> mpg.1 - carb.3: -0.12 -> -0.12
#> cyl.1 - disp.1: 0.23 -> 0.23
#> cyl.1 - hp.1: -0.19 -> -0.19
#> cyl.1 - drat.1: 0.21 -> 0.21
#> cyl.1 - wt.1: -0.32 -> -0.32
#> cyl.1 - qsec.1: no change (0.16)
#> cyl.1 - vs.1: 3.19e-03 -> 3.19e-03
#> cyl.1 - am.1: -0.08 -> -0.08
#> cyl.1 - gear.1: -0.02 -> -0.02
#> cyl.1 - carb.1: 0.09 -> 0.09
#> cyl.1 - mpg.2: -0.17 -> -0.17
#> cyl.1 - cyl.2: 0.16 -> 0.16
#> cyl.1 - disp.2: 5.69e-04 -> 5.69e-04
#> cyl.1 - hp.2: -0.22 -> -0.22
#> cyl.1 - drat.2: -0.01 -> -0.01
#> cyl.1 - wt.2: 0.11 -> 0.11
#> cyl.1 - qsec.2: 0.16 -> 0.16
#> cyl.1 - vs.2: 0.01 -> 0.01
#> cyl.1 - am.2: 0.06 -> 0.06
#> cyl.1 - gear.2: -0.21 -> -0.21
#> cyl.1 - carb.2: 0.08 -> 0.08
#> cyl.1 - mpg.3: 0.05 -> 0.05
#> cyl.1 - cyl.3: 0.08 -> 0.08
#> cyl.1 - disp.3: -0.09 -> -0.09
#> cyl.1 - hp.3: 0.17 -> 0.17
#> cyl.1 - drat.3: -0.08 -> -0.08
#> cyl.1 - wt.3: -0.25 -> -0.25
#> cyl.1 - qsec.3: -0.20 -> -0.20
#> cyl.1 - vs.3: 0.02 -> 0.02
#> cyl.1 - am.3: -0.23 -> -0.23
#> cyl.1 - gear.3: -0.02 -> -0.02
#> cyl.1 - carb.3: 0.07 -> 0.07
#> disp.1 - hp.1: -0.19 -> -0.19
#> disp.1 - drat.1: 0.12 -> 0.12
#> disp.1 - wt.1: 7.63e-04 -> 7.63e-04
#> disp.1 - qsec.1: -0.30 -> -0.30
#> disp.1 - vs.1: 0.02 -> 0.02
#> disp.1 - am.1: -0.15 -> -0.15
#> disp.1 - gear.1: -0.32 -> -0.32
#> disp.1 - carb.1: 0.09 -> 0.09
#> disp.1 - mpg.2: -0.11 -> -0.11
#> disp.1 - cyl.2: -0.13 -> -0.13
#> disp.1 - disp.2: 0.02 -> 0.02
#> disp.1 - hp.2: 0.08 -> 0.08
#> disp.1 - drat.2: 0.13 -> 0.13
#> disp.1 - wt.2: 0.01 -> 0.01
#> disp.1 - qsec.2: 0.28 -> 0.28
#> disp.1 - vs.2: -0.08 -> -0.08
#> disp.1 - am.2: -0.07 -> -0.07
#> disp.1 - gear.2: -9.40e-03 -> -9.40e-03
#> disp.1 - carb.2: -0.14 -> -0.14
#> disp.1 - mpg.3: -0.12 -> -0.12
#> disp.1 - cyl.3: -0.12 -> -0.12
#> disp.1 - disp.3: 0.33 -> 0.33
#> disp.1 - hp.3: 0.32 -> 0.32
#> disp.1 - drat.3: -0.04 -> -0.04
#> disp.1 - wt.3: -0.08 -> -0.08
#> disp.1 - qsec.3: -0.14 -> -0.14
#> disp.1 - vs.3: 0.23 -> 0.23
#> disp.1 - am.3: -0.24 -> -0.24
#> disp.1 - gear.3: -0.18 -> -0.18
#> disp.1 - carb.3: -0.08 -> -0.08
#> hp.1 - drat.1: -3.90e-03 -> -3.90e-03
#> hp.1 - wt.1: 0.17 -> 0.17
#> hp.1 - qsec.1: 0.22 -> 0.22
#> hp.1 - vs.1: 0.10 -> 0.10
#> hp.1 - am.1: -0.01 -> -0.01
#> hp.1 - gear.1: -0.10 -> -0.10
#> hp.1 - carb.1: 0.18 -> 0.18
#> hp.1 - mpg.2: 0.13 -> 0.13
#> hp.1 - cyl.2: -0.01 -> -0.01
#> hp.1 - disp.2: -0.28 -> -0.28
#> hp.1 - hp.2: 0.15 -> 0.15
#> hp.1 - drat.2: 0.06 -> 0.06
#> hp.1 - wt.2: 0.06 -> 0.06
#> hp.1 - qsec.2: -0.03 -> -0.03
#> hp.1 - vs.2: 0.22 -> 0.22
#> hp.1 - am.2: -0.07 -> -0.07
#> hp.1 - gear.2: 0.09 -> 0.09
#> hp.1 - carb.2: 0.27 -> 0.27
#> hp.1 - mpg.3: 0.07 -> 0.07
#> hp.1 - cyl.3: -0.07 -> -0.07
#> hp.1 - disp.3: -0.28 -> -0.28
#> hp.1 - hp.3: 0.09 -> 0.09
#> hp.1 - drat.3: 0.24 -> 0.24
#> hp.1 - wt.3: 0.31 -> 0.31
#> hp.1 - qsec.3: 0.11 -> 0.11
#> hp.1 - vs.3: 0.15 -> 0.15
#> hp.1 - am.3: -0.18 -> -0.18
#> hp.1 - gear.3: 0.07 -> 0.07
#> hp.1 - carb.3: -0.26 -> -0.26
#> drat.1 - wt.1: -0.08 -> -0.08
#> drat.1 - qsec.1: 0.20 -> 0.20
#> drat.1 - vs.1: 0.16 -> 0.16
#> drat.1 - am.1: 0.08 -> 0.08
#> drat.1 - gear.1: -0.03 -> -0.03
#> drat.1 - carb.1: -0.06 -> -0.06
#> drat.1 - mpg.2: -0.02 -> -0.02
#> drat.1 - cyl.2: -0.06 -> -0.06
#> drat.1 - disp.2: 0.24 -> 0.24
#> drat.1 - hp.2: -0.11 -> -0.11
#> drat.1 - drat.2: 0.11 -> 0.11
#> drat.1 - wt.2: -0.29 -> -0.29
#> drat.1 - qsec.2: 0.22 -> 0.22
#> drat.1 - vs.2: -0.40 -> -0.40
#> drat.1 - am.2: 0.43 -> 0.43
#> drat.1 - gear.2: 0.05 -> 0.05
#> drat.1 - carb.2: 0.06 -> 0.06
#> drat.1 - mpg.3: -0.19 -> -0.19
#> drat.1 - cyl.3: 0.03 -> 0.03
#> drat.1 - disp.3: -0.34 -> -0.34
#> drat.1 - hp.3: 0.20 -> 0.20
#> drat.1 - drat.3: -0.23 -> -0.23
#> drat.1 - wt.3: -0.05 -> -0.05
#> drat.1 - qsec.3: -0.12 -> -0.12
#> drat.1 - vs.3: 0.02 -> 0.02
#> drat.1 - am.3: -0.30 -> -0.30
#> drat.1 - gear.3: 0.03 -> 0.03
#> drat.1 - carb.3: 0.31 -> 0.31
#> wt.1 - qsec.1: -0.13 -> -0.13
#> wt.1 - vs.1: 0.15 -> 0.15
#> wt.1 - am.1: 0.29 -> 0.29
#> wt.1 - gear.1: 0.01 -> 0.01
#> wt.1 - carb.1: -0.19 -> -0.19
#> wt.1 - mpg.2: -0.08 -> -0.08
#> wt.1 - cyl.2: 0.11 -> 0.11
#> wt.1 - disp.2: -0.12 -> -0.12
#> wt.1 - hp.2: 9.80e-03 -> 9.80e-03
#> wt.1 - drat.2: -0.07 -> -0.07
#> wt.1 - wt.2: 0.03 -> 0.03
#> wt.1 - qsec.2: 0.21 -> 0.21
#> wt.1 - vs.2: -0.27 -> -0.27
#> wt.1 - am.2: -0.13 -> -0.13
#> wt.1 - gear.2: -0.01 -> -0.01
#> wt.1 - carb.2: 0.01 -> 0.01
#> wt.1 - mpg.3: 0.08 -> 0.08
#> wt.1 - cyl.3: 2.68e-03 -> 2.68e-03
#> wt.1 - disp.3: 0.02 -> 0.02
#> wt.1 - hp.3: -0.11 -> -0.11
#> wt.1 - drat.3: -0.21 -> -0.21
#> wt.1 - wt.3: 0.39 -> 0.39
#> wt.1 - qsec.3: -0.03 -> -0.03
#> wt.1 - vs.3: 0.22 -> 0.22
#> wt.1 - am.3: 0.10 -> 0.10
#> wt.1 - gear.3: 0.13 -> 0.13
#> wt.1 - carb.3: 0.05 -> 0.05
#> qsec.1 - vs.1: 0.06 -> 0.06
#> qsec.1 - am.1: 0.22 -> 0.22
#> qsec.1 - gear.1: 0.24 -> 0.24
#> qsec.1 - carb.1: 0.10 -> 0.10
#> qsec.1 - mpg.2: -0.15 -> -0.15
#> qsec.1 - cyl.2: -0.19 -> -0.19
#> qsec.1 - disp.2: 0.02 -> 0.02
#> qsec.1 - hp.2: -0.13 -> -0.13
#> qsec.1 - drat.2: -0.08 -> -0.08
#> qsec.1 - wt.2: -0.21 -> -0.21
#> qsec.1 - qsec.2: -0.02 -> -0.02
#> qsec.1 - vs.2: 0.12 -> 0.12
#> qsec.1 - am.2: -0.20 -> -0.20
#> qsec.1 - gear.2: -0.15 -> -0.15
#> qsec.1 - carb.2: 0.04 -> 0.04
#> qsec.1 - mpg.3: -0.12 -> -0.12
#> qsec.1 - cyl.3: 0.26 -> 0.26
#> qsec.1 - disp.3: -0.30 -> -0.30
#> qsec.1 - hp.3: 0.27 -> 0.27
#> qsec.1 - drat.3: -0.18 -> -0.18
#> qsec.1 - wt.3: -0.02 -> -0.02
#> qsec.1 - qsec.3: 0.10 -> 0.10
#> qsec.1 - vs.3: 0.04 -> 0.04
#> qsec.1 - am.3: -0.11 -> -0.11
#> qsec.1 - gear.3: -0.02 -> -0.02
#> qsec.1 - carb.3: -0.06 -> -0.06
#> vs.1 - am.1: 0.11 -> 0.11
#> vs.1 - gear.1: -0.01 -> -0.01
#> vs.1 - carb.1: 0.07 -> 0.07
#> vs.1 - mpg.2: -0.05 -> -0.05
#> vs.1 - cyl.2: -0.02 -> -0.02
#> vs.1 - disp.2: 0.06 -> 0.06
#> vs.1 - hp.2: -9.60e-03 -> -9.60e-03
#> vs.1 - drat.2: 0.21 -> 0.21
#> vs.1 - wt.2: -0.14 -> -0.14
#> vs.1 - qsec.2: 0.14 -> 0.14
#> vs.1 - vs.2: -0.21 -> -0.21
#> vs.1 - am.2: -0.03 -> -0.03
#> vs.1 - gear.2: 0.26 -> 0.26
#> vs.1 - carb.2: 0.01 -> 0.01
#> vs.1 - mpg.3: 0.22 -> 0.22
#> vs.1 - cyl.3: -0.04 -> -0.04
#> vs.1 - disp.3: -0.22 -> -0.22
#> vs.1 - hp.3: 0.06 -> 0.06
#> vs.1 - drat.3: 0.16 -> 0.16
#> vs.1 - wt.3: -0.16 -> -0.16
#> vs.1 - qsec.3: -0.28 -> -0.28
#> vs.1 - vs.3: 0.60 -> 0.60
#> vs.1 - am.3: -0.07 -> -0.07
#> vs.1 - gear.3: -0.11 -> -0.11
#> vs.1 - carb.3: -0.06 -> -0.06
#> am.1 - gear.1: 0.26 -> 0.26
#> am.1 - carb.1: -0.30 -> -0.30
#> am.1 - mpg.2: -0.33 -> -0.33
#> am.1 - cyl.2: -0.01 -> -0.01
#> am.1 - disp.2: 0.01 -> 0.01
#> am.1 - hp.2: -0.04 -> -0.04
#> am.1 - drat.2: -0.19 -> -0.19
#> am.1 - wt.2: 0.13 -> 0.13
#> am.1 - qsec.2: -0.23 -> -0.23
#> am.1 - vs.2: -0.22 -> -0.22
#> am.1 - am.2: -0.62 -> -0.62
#> am.1 - gear.2: 0.18 -> 0.18
#> am.1 - carb.2: -0.10 -> -0.10
#> am.1 - mpg.3: 0.09 -> 0.09
#> am.1 - cyl.3: 0.17 -> 0.17
#> am.1 - disp.3: -0.06 -> -0.06
#> am.1 - hp.3: -0.03 -> -0.03
#> am.1 - drat.3: 0.07 -> 0.07
#> am.1 - wt.3: -0.02 -> -0.02
#> am.1 - qsec.3: -0.12 -> -0.12
#> am.1 - vs.3: -0.12 -> -0.12
#> am.1 - am.3: -0.01 -> -0.01
#> am.1 - gear.3: 0.08 -> 0.08
#> am.1 - carb.3: 0.29 -> 0.29
#> gear.1 - carb.1: -0.25 -> -0.25
#> gear.1 - mpg.2: -0.08 -> -0.08
#> gear.1 - cyl.2: 0.31 -> 0.31
#> gear.1 - disp.2: -0.08 -> -0.08
#> gear.1 - hp.2: -0.44 -> -0.44
#> gear.1 - drat.2: -0.05 -> -0.05
#> gear.1 - wt.2: 0.02 -> 0.02
#> gear.1 - qsec.2: -0.20 -> -0.20
#> gear.1 - vs.2: 0.05 -> 0.05
#> gear.1 - am.2: -0.02 -> -0.02
#> gear.1 - gear.2: 0.33 -> 0.33
#> gear.1 - carb.2: -0.05 -> -0.05
#> gear.1 - mpg.3: 0.08 -> 0.08
#> gear.1 - cyl.3: 0.17 -> 0.17
#> gear.1 - disp.3: 0.03 -> 0.03
#> gear.1 - hp.3: -0.19 -> -0.19
#> gear.1 - drat.3: 0.04 -> 0.04
#> gear.1 - wt.3: 0.09 -> 0.09
#> gear.1 - qsec.3: 0.04 -> 0.04
#> gear.1 - vs.3: -0.14 -> -0.14
#> gear.1 - am.3: -0.10 -> -0.10
#> gear.1 - gear.3: -0.39 -> -0.39
#> gear.1 - carb.3: 0.19 -> 0.19
#> carb.1 - mpg.2: 0.17 -> 0.17
#> carb.1 - cyl.2: -0.08 -> -0.08
#> carb.1 - disp.2: -0.19 -> -0.19
#> carb.1 - hp.2: 0.28 -> 0.28
#> carb.1 - drat.2: -0.38 -> -0.38
#> carb.1 - wt.2: 0.09 -> 0.09
#> carb.1 - qsec.2: 0.11 -> 0.11
#> carb.1 - vs.2: 0.03 -> 0.03
#> carb.1 - am.2: 0.08 -> 0.08
#> carb.1 - gear.2: 0.01 -> 0.01
#> carb.1 - carb.2: 0.31 -> 0.31
#> carb.1 - mpg.3: 0.11 -> 0.11
#> carb.1 - cyl.3: -0.18 -> -0.18
#> carb.1 - disp.3: 0.29 -> 0.29
#> carb.1 - hp.3: 0.33 -> 0.33
#> carb.1 - drat.3: 0.03 -> 0.03
#> carb.1 - wt.3: -0.04 -> -0.04
#> carb.1 - qsec.3: 0.37 -> 0.37
#> carb.1 - vs.3: 0.12 -> 0.12
#> carb.1 - am.3: 0.06 -> 0.06
#> carb.1 - gear.3: -0.23 -> -0.23
#> carb.1 - carb.3: -0.42 -> -0.42
#> mpg.2 - cyl.2: -0.09 -> -0.09
#> mpg.2 - disp.2: -0.02 -> -0.02
#> mpg.2 - hp.2: 0.01 -> 0.01
#> mpg.2 - drat.2: 0.18 -> 0.18
#> mpg.2 - wt.2: 0.01 -> 0.01
#> mpg.2 - qsec.2: 0.22 -> 0.22
#> mpg.2 - vs.2: 0.03 -> 0.03
#> mpg.2 - am.2: 0.16 -> 0.16
#> mpg.2 - gear.2: 0.15 -> 0.15
#> mpg.2 - carb.2: -0.03 -> -0.03
#> mpg.2 - mpg.3: 0.03 -> 0.03
#> mpg.2 - cyl.3: -0.04 -> -0.04
#> mpg.2 - disp.3: 0.11 -> 0.11
#> mpg.2 - hp.3: -0.11 -> -0.11
#> mpg.2 - drat.3: 0.18 -> 0.18
#> mpg.2 - wt.3: 0.10 -> 0.10
#> mpg.2 - qsec.3: 0.38 -> 0.38
#> mpg.2 - vs.3: 0.02 -> 0.02
#> mpg.2 - am.3: 0.26 -> 0.26
#> mpg.2 - gear.3: -0.22 -> -0.22
#> mpg.2 - carb.3: -0.17 -> -0.17
#> cyl.2 - disp.2: 1.74e-03 -> 1.74e-03
#> cyl.2 - hp.2: -0.14 -> -0.14
#> cyl.2 - drat.2: -0.26 -> -0.26
#> cyl.2 - wt.2: 0.08 -> 0.08
#> cyl.2 - qsec.2: 0.04 -> 0.04
#> cyl.2 - vs.2: 0.11 -> 0.11
#> cyl.2 - am.2: 0.11 -> 0.11
#> cyl.2 - gear.2: -0.06 -> -0.06
#> cyl.2 - carb.2: -0.11 -> -0.11
#> cyl.2 - mpg.3: -0.03 -> -0.03
#> cyl.2 - cyl.3: -0.33 -> -0.33
#> cyl.2 - disp.3: 0.03 -> 0.03
#> cyl.2 - hp.3: -0.06 -> -0.06
#> cyl.2 - drat.3: 0.17 -> 0.17
#> cyl.2 - wt.3: 0.09 -> 0.09
#> cyl.2 - qsec.3: -0.12 -> -0.12
#> cyl.2 - vs.3: -0.02 -> -0.02
#> cyl.2 - am.3: -0.16 -> -0.16
#> cyl.2 - gear.3: 0.01 -> 0.01
#> cyl.2 - carb.3: 0.23 -> 0.23
#> disp.2 - hp.2: -0.38 -> -0.38
#> disp.2 - drat.2: -0.06 -> -0.06
#> disp.2 - wt.2: -0.12 -> -0.12
#> disp.2 - qsec.2: 0.04 -> 0.04
#> disp.2 - vs.2: -0.05 -> -0.05
#> disp.2 - am.2: 0.10 -> 0.10
#> disp.2 - gear.2: -0.17 -> -0.17
#> disp.2 - carb.2: -0.25 -> -0.25
#> disp.2 - mpg.3: -0.01 -> -0.01
#> disp.2 - cyl.3: 0.09 -> 0.09
#> disp.2 - disp.3: -0.25 -> -0.25
#> disp.2 - hp.3: 0.05 -> 0.05
#> disp.2 - drat.3: -0.07 -> -0.07
#> disp.2 - wt.3: -0.08 -> -0.08
#> disp.2 - qsec.3: -0.10 -> -0.10
#> disp.2 - vs.3: -0.01 -> -0.01
#> disp.2 - am.3: -0.05 -> -0.05
#> disp.2 - gear.3: 0.10 -> 0.10
#> disp.2 - carb.3: 0.56 -> 0.56
#> hp.2 - drat.2: -0.06 -> -0.06
#> hp.2 - wt.2: -0.01 -> -0.01
#> hp.2 - qsec.2: 0.05 -> 0.05
#> hp.2 - vs.2: -0.09 -> -0.09
#> hp.2 - am.2: -0.13 -> -0.13
#> hp.2 - gear.2: -0.14 -> -0.14
#> hp.2 - carb.2: 0.10 -> 0.10
#> hp.2 - mpg.3: 0.09 -> 0.09
#> hp.2 - cyl.3: -0.17 -> -0.17
#> hp.2 - disp.3: 0.31 -> 0.31
#> hp.2 - hp.3: 0.12 -> 0.12
#> hp.2 - drat.3: -0.08 -> -0.08
#> hp.2 - wt.3: 2.16e-03 -> 2.16e-03
#> hp.2 - qsec.3: 0.24 -> 0.24
#> hp.2 - vs.3: -0.15 -> -0.15
#> hp.2 - am.3: 0.13 -> 0.13
#> hp.2 - gear.3: 0.15 -> 0.15
#> hp.2 - carb.3: -0.40 -> -0.40
#> drat.2 - wt.2: -0.12 -> -0.12
#> drat.2 - qsec.2: -0.05 -> -0.05
#> drat.2 - vs.2: 0.05 -> 0.05
#> drat.2 - am.2: 0.25 -> 0.25
#> drat.2 - gear.2: -0.02 -> -0.02
#> drat.2 - carb.2: -0.23 -> -0.23
#> drat.2 - mpg.3: -0.16 -> -0.16
#> drat.2 - cyl.3: 0.43 -> 0.43
#> drat.2 - disp.3: -0.34 -> -0.34
#> drat.2 - hp.3: 0.13 -> 0.13
#> drat.2 - drat.3: 0.05 -> 0.05
#> drat.2 - wt.3: -0.08 -> -0.08
#> drat.2 - qsec.3: -0.13 -> -0.13
#> drat.2 - vs.3: 0.21 -> 0.21
#> drat.2 - am.3: -0.21 -> -0.21
#> drat.2 - gear.3: 0.21 -> 0.21
#> drat.2 - carb.3: 0.06 -> 0.06
#> wt.2 - qsec.2: -0.26 -> -0.26
#> wt.2 - vs.2: 0.27 -> 0.27
#> wt.2 - am.2: -0.24 -> -0.24
#> wt.2 - gear.2: -0.16 -> -0.16
#> wt.2 - carb.2: 0.22 -> 0.22
#> wt.2 - mpg.3: 0.12 -> 0.12
#> wt.2 - cyl.3: 0.23 -> 0.23
#> wt.2 - disp.3: 0.12 -> 0.12
#> wt.2 - hp.3: -0.48 -> -0.48
#> wt.2 - drat.3: 0.34 -> 0.34
#> wt.2 - wt.3: 0.47 -> 0.47
#> wt.2 - qsec.3: 0.14 -> 0.14
#> wt.2 - vs.3: -0.09 -> -0.09
#> wt.2 - am.3: 0.02 -> 0.02
#> wt.2 - gear.3: 0.06 -> 0.06
#> wt.2 - carb.3: -0.05 -> -0.05
#> qsec.2 - vs.2: -0.04 -> -0.04
#> qsec.2 - am.2: 0.24 -> 0.24
#> qsec.2 - gear.2: -0.14 -> -0.14
#> qsec.2 - carb.2: -0.06 -> -0.06
#> qsec.2 - mpg.3: -0.12 -> -0.12
#> qsec.2 - cyl.3: -0.36 -> -0.36
#> qsec.2 - disp.3: 0.21 -> 0.21
#> qsec.2 - hp.3: 0.19 -> 0.19
#> qsec.2 - drat.3: -0.04 -> -0.04
#> qsec.2 - wt.3: -0.21 -> -0.21
#> qsec.2 - qsec.3: -0.20 -> -0.20
#> qsec.2 - vs.3: 0.14 -> 0.14
#> qsec.2 - am.3: -0.12 -> -0.12
#> qsec.2 - gear.3: -9.59e-03 -> -9.59e-03
#> qsec.2 - carb.3: -0.10 -> -0.10
#> vs.2 - am.2: -0.10 -> -0.10
#> vs.2 - gear.2: -0.11 -> -0.11
#> vs.2 - carb.2: -0.09 -> -0.09
#> vs.2 - mpg.3: -0.08 -> -0.08
#> vs.2 - cyl.3: 0.11 -> 0.11
#> vs.2 - disp.3: -0.08 -> -0.08
#> vs.2 - hp.3: -0.05 -> -0.05
#> vs.2 - drat.3: 0.23 -> 0.23
#> vs.2 - wt.3: -0.06 -> -0.06
#> vs.2 - qsec.3: -0.04 -> -0.04
#> vs.2 - vs.3: 0.05 -> 0.05
#> vs.2 - am.3: -0.03 -> -0.03
#> vs.2 - gear.3: -0.02 -> -0.02
#> vs.2 - carb.3: -0.01 -> -0.01
#> am.2 - gear.2: -0.04 -> -0.04
#> am.2 - carb.2: -5.31e-03 -> -5.31e-03
#> am.2 - mpg.3: 0.06 -> 0.06
#> am.2 - cyl.3: -0.13 -> -0.13
#> am.2 - disp.3: -0.21 -> -0.21
#> am.2 - hp.3: 0.06 -> 0.06
#> am.2 - drat.3: -0.21 -> -0.21
#> am.2 - wt.3: -8.95e-03 -> -8.95e-03
#> am.2 - qsec.3: -0.02 -> -0.02
#> am.2 - vs.3: 0.10 -> 0.10
#> am.2 - am.3: -0.10 -> -0.10
#> am.2 - gear.3: 8.98e-03 -> 8.98e-03
#> am.2 - carb.3: 0.06 -> 0.06
#> gear.2 - carb.2: 0.28 -> 0.28
#> gear.2 - mpg.3: 0.24 -> 0.24
#> gear.2 - cyl.3: -0.26 -> -0.26
#> gear.2 - disp.3: 0.14 -> 0.14
#> gear.2 - hp.3: -0.13 -> -0.13
#> gear.2 - drat.3: -0.05 -> -0.05
#> gear.2 - wt.3: -0.07 -> -0.07
#> gear.2 - qsec.3: 0.07 -> 0.07
#> gear.2 - vs.3: 0.07 -> 0.07
#> gear.2 - am.3: -0.05 -> -0.05
#> gear.2 - gear.3: -0.30 -> -0.30
#> gear.2 - carb.3: -0.26 -> -0.26
#> carb.2 - mpg.3: -0.04 -> -0.04
#> carb.2 - cyl.3: -0.03 -> -0.03
#> carb.2 - disp.3: -0.17 -> -0.17
#> carb.2 - hp.3: -0.36 -> -0.36
#> carb.2 - drat.3: -0.17 -> -0.17
#> carb.2 - wt.3: 0.31 -> 0.31
#> carb.2 - qsec.3: 0.11 -> 0.11
#> carb.2 - vs.3: -0.06 -> -0.06
#> carb.2 - am.3: -0.13 -> -0.13
#> carb.2 - gear.3: 0.09 -> 0.09
#> carb.2 - carb.3: -0.56 -> -0.56
#> mpg.3 - cyl.3: -0.19 -> -0.19
#> mpg.3 - disp.3: 0.15 -> 0.15
#> mpg.3 - hp.3: 0.04 -> 0.04
#> mpg.3 - drat.3: 8.57e-03 -> 8.57e-03
#> mpg.3 - wt.3: 0.08 -> 0.08
#> mpg.3 - qsec.3: 0.05 -> 0.05
#> mpg.3 - vs.3: 0.11 -> 0.11
#> mpg.3 - am.3: 0.40 -> 0.40
#> mpg.3 - gear.3: -0.24 -> -0.24
#> mpg.3 - carb.3: -0.09 -> -0.09
#> cyl.3 - disp.3: -0.35 -> -0.35
#> cyl.3 - hp.3: -0.13 -> -0.13
#> cyl.3 - drat.3: -0.09 -> -0.09
#> cyl.3 - wt.3: 0.22 -> 0.22
#> cyl.3 - qsec.3: 0.04 -> 0.04
#> cyl.3 - vs.3: 6.21e-03 -> 6.21e-03
#> cyl.3 - am.3: -0.12 -> -0.12
#> cyl.3 - gear.3: -0.04 -> -0.04
#> cyl.3 - carb.3: 0.24 -> 0.24
#> disp.3 - hp.3: 0.12 -> 0.12
#> disp.3 - drat.3: -4.94e-03 -> -4.94e-03
#> disp.3 - wt.3: -0.02 -> -0.02
#> disp.3 - qsec.3: 0.07 -> 0.07
#> disp.3 - vs.3: -0.26 -> -0.26
#> disp.3 - am.3: 0.05 -> 0.05
#> disp.3 - gear.3: -0.31 -> -0.31
#> disp.3 - carb.3: -0.31 -> -0.31
#> hp.3 - drat.3: -0.07 -> -0.07
#> hp.3 - wt.3: -0.39 -> -0.39
#> hp.3 - qsec.3: -0.05 -> -0.05
#> hp.3 - vs.3: 4.85e-03 -> 4.85e-03
#> hp.3 - am.3: -0.18 -> -0.18
#> hp.3 - gear.3: 0.03 -> 0.03
#> hp.3 - carb.3: 0.07 -> 0.07
#> drat.3 - wt.3: -0.24 -> -0.24
#> drat.3 - qsec.3: -0.01 -> -0.01
#> drat.3 - vs.3: 0.12 -> 0.12
#> drat.3 - am.3: 0.10 -> 0.10
#> drat.3 - gear.3: 0.08 -> 0.08
#> drat.3 - carb.3: 0.09 -> 0.09
#> wt.3 - qsec.3: no change (0.27)
#> wt.3 - vs.3: 4.08e-03 -> 4.08e-03
#> wt.3 - am.3: 0.02 -> 0.02
#> wt.3 - gear.3: -0.13 -> -0.13
#> wt.3 - carb.3: -0.19 -> -0.19
#> qsec.3 - vs.3: -0.08 -> -0.08
#> qsec.3 - am.3: 0.48 -> 0.48
#> qsec.3 - gear.3: -0.18 -> -0.18
#> qsec.3 - carb.3: -0.18 -> -0.18
#> vs.3 - am.3: 0.15 -> 0.15
#> vs.3 - gear.3: -0.12 -> -0.12
#> vs.3 - carb.3: -0.09 -> -0.09
#> am.3 - gear.3: -0.06 -> -0.06
#> am.3 - carb.3: 0.03 -> 0.03
#> gear.3 - carb.3: 0.14 -> 0.14
#>