Correlation Analysis

Performs a correlation analysis.

correlation(data, data2 = NULL, method = "pearson",
  p_adjust = "holm", ci = "default", bayesian = FALSE,
  bayesian_prior = "medium", bayesian_ci_method = "hdi",
  bayesian_test = c("pd", "rope", "bf"), redundant = FALSE,
  include_factors = FALSE, partial = FALSE, partial_bayesian = FALSE,
  multilevel = FALSE, ...)

Arguments

data
A dataframe.

data2
An optional dataframe.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson" (default), "kendall", or "spearman", "biserial", "polychoric", "tetrachoric", or "biweight". Setting "auto" will attempt at selecting the most relevant method (polychoric when ordinal factors involved, tetrachoric when dichotomous factors involved, point-biserial if one dichotomous and one continuous and pearson otherwise).

p_adjust
Correction method for frequentist correlations. One of "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".

ci
Confidence/Credible Interval level. If "default", then 0.95 for Frequentist and 0.89 for Bayesian (see documentation in the bayestestR package).

bayesian
If TRUE, will run the correlations under a Bayesian framework Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

bayesian_prior
For the prior argument, several named values are recognized: "medium.narrow", "medium", "wide", and "ultrawide". These correspond to scale values of 1/sqrt(27), 1/3, 1/sqrt(3) and 1, respectively. See the BayesFactor::correlationBF function.

bayesian_ci_method
See arguments in model_parameters for BayesFactor tests.

bayesian_test
See arguments in model_parameters for BayesFactor tests.

redundant
Shoud the data include redundant rows (where each given correlation is repeated two times).

include_factors
If TRUE, the factors are kept and eventually converted to numeric or used as random effects (depending of multilevel). If FALSE, factors are removed upfront.

partial
Can be TRUE or "semi" for partial and semi-partial correlations, respectively. This only works for Frequentist correlations.

partial_bayesian
If TRUE, will run the correlations under a Bayesian framework Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

multilevel
If TRUE, the factors are included as random factors. If FALSE (default), factors are binarized (dummified) and partialized out in the same than the other numeric variables.

...
Arguments passed to or from other methods.

Examples

library(dplyr)

#> Warning: package 'dplyr' was built under R version 3.6.1

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

cor <- correlation(iris)

cor

#> Parameter1   |   Parameter2 |     r |     t |  df |      p |         95% CI |  Method
#> -------------------------------------------------------------------------------------
#> Sepal.Length |  Sepal.Width | -0.12 | -1.44 | 148 | 0.15   | [-0.27,  0.04] | Pearson
#> Sepal.Length | Petal.Length |  0.87 | 21.65 | 148 | < .001 | [ 0.83,  0.91] | Pearson
#> Sepal.Length |  Petal.Width |  0.82 | 17.30 | 148 | < .001 | [ 0.76,  0.86] | Pearson
#> Sepal.Width  | Petal.Length | -0.43 | -5.77 | 148 | < .001 | [-0.55, -0.29] | Pearson
#> Sepal.Width  |  Petal.Width | -0.37 | -4.79 | 148 | < .001 | [-0.50, -0.22] | Pearson
#> Petal.Length |  Petal.Width |  0.96 | 43.39 | 148 | < .001 | [ 0.95,  0.97] | Pearson

summary(cor)

#> Parameter    | Petal.Width | Petal.Length | Sepal.Width
#> -------------------------------------------------------
#> Sepal.Length |     0.82*** |      0.87*** |       -0.12
#> Sepal.Width  |    -0.37*** |     -0.43*** |            
#> Petal.Length |     0.96*** |              |            

as.table(cor)

#> Parameter    | Sepal.Length | Sepal.Width | Petal.Length | Petal.Width
#> ----------------------------------------------------------------------
#> Sepal.Length |      1.00*** |       -0.12 |      0.87*** |     0.82***
#> Sepal.Width  |        -0.12 |     1.00*** |     -0.43*** |    -0.37***
#> Petal.Length |      0.87*** |    -0.43*** |      1.00*** |     0.96***
#> Petal.Width  |      0.82*** |    -0.37*** |      0.96*** |     1.00***

iris %>%
  group_by(Species) %>%
  correlation()

#> Group      |   Parameter1 |   Parameter2 |    r |     t | df |      p |         95% CI |  Method
#> ------------------------------------------------------------------------------------------------
#> setosa     | Sepal.Length |  Sepal.Width | 0.74 |  7.68 | 48 | < .001 | [ 0.59,  0.85] | Pearson
#> setosa     | Sepal.Length | Petal.Length | 0.27 |  1.92 | 48 | 0.20   | [-0.01,  0.51] | Pearson
#> setosa     | Sepal.Length |  Petal.Width | 0.28 |  2.01 | 48 | 0.20   | [ 0.00,  0.52] | Pearson
#> setosa     |  Sepal.Width | Petal.Length | 0.18 |  1.25 | 48 | 0.22   | [-0.11,  0.43] | Pearson
#> setosa     |  Sepal.Width |  Petal.Width | 0.23 |  1.66 | 48 | 0.21   | [-0.05,  0.48] | Pearson
#> setosa     | Petal.Length |  Petal.Width | 0.33 |  2.44 | 48 | 0.09   | [ 0.06,  0.56] | Pearson
#> versicolor | Sepal.Length |  Sepal.Width | 0.53 |  4.28 | 48 | < .001 | [ 0.29,  0.70] | Pearson
#> versicolor | Sepal.Length | Petal.Length | 0.75 |  7.95 | 48 | < .001 | [ 0.60,  0.85] | Pearson
#> versicolor | Sepal.Length |  Petal.Width | 0.55 |  4.52 | 48 | < .001 | [ 0.32,  0.72] | Pearson
#> versicolor |  Sepal.Width | Petal.Length | 0.56 |  4.69 | 48 | < .001 | [ 0.33,  0.73] | Pearson
#> versicolor |  Sepal.Width |  Petal.Width | 0.66 |  6.15 | 48 | < .001 | [ 0.47,  0.80] | Pearson
#> versicolor | Petal.Length |  Petal.Width | 0.79 |  8.83 | 48 | < .001 | [ 0.65,  0.87] | Pearson
#> virginica  | Sepal.Length |  Sepal.Width | 0.46 |  3.56 | 48 | < .01  | [ 0.20,  0.65] | Pearson
#> virginica  | Sepal.Length | Petal.Length | 0.86 | 11.90 | 48 | < .001 | [ 0.77,  0.92] | Pearson
#> virginica  | Sepal.Length |  Petal.Width | 0.28 |  2.03 | 48 | < .05  | [ 0.00,  0.52] | Pearson
#> virginica  |  Sepal.Width | Petal.Length | 0.40 |  3.03 | 48 | < .05  | [ 0.14,  0.61] | Pearson
#> virginica  |  Sepal.Width |  Petal.Width | 0.54 |  4.42 | 48 | < .001 | [ 0.31,  0.71] | Pearson
#> virginica  | Petal.Length |  Petal.Width | 0.32 |  2.36 | 48 | < .05  | [ 0.05,  0.55] | Pearson

correlation(mtcars[-2], method = "auto")

#> Parameter1 | Parameter2 |     r |      t | df |      p |         95% CI |      Method
#> -------------------------------------------------------------------------------------
#> mpg        |       disp | -0.85 |  -8.75 | 30 | < .001 | [-0.92, -0.71] |     Pearson
#> mpg        |         hp | -0.78 |  -6.74 | 30 | < .001 | [-0.89, -0.59] |     Pearson
#> mpg        |       drat |  0.68 |   5.10 | 30 | < .001 | [ 0.44,  0.83] |     Pearson
#> mpg        |         wt | -0.87 |  -9.56 | 30 | < .001 | [-0.93, -0.74] |     Pearson
#> mpg        |       qsec |  0.42 |   2.53 | 30 | 0.22   | [ 0.08,  0.67] |     Pearson
#> mpg        |         vs |  0.82 |   7.80 | 30 | < .001 | [ 0.66,  0.91] |    Biserial
#> mpg        |         am |  0.75 |   6.06 | 30 | < .001 | [ 0.54,  0.87] |    Biserial
#> mpg        |       gear |  0.48 |   3.00 | 30 | 0.09   | [ 0.16,  0.71] |     Pearson
#> mpg        |       carb | -0.55 |  -3.62 | 30 | < .05  | [-0.75, -0.25] |     Pearson
#> disp       |         hp |  0.79 |   7.08 | 30 | < .001 | [ 0.61,  0.89] |     Pearson
#> disp       |       drat | -0.71 |  -5.53 | 30 | < .001 | [-0.85, -0.48] |     Pearson
#> disp       |         wt |  0.89 |  10.58 | 30 | < .001 | [ 0.78,  0.94] |     Pearson
#> disp       |       qsec | -0.43 |  -2.64 | 30 | 0.20   | [-0.68, -0.10] |     Pearson
#> disp       |         vs | -0.88 |  -9.99 | 30 | < .001 | [-0.94, -0.77] |    Biserial
#> disp       |         am | -0.74 |  -5.87 | 30 | < .001 | [-0.86, -0.52] |    Biserial
#> disp       |       gear | -0.56 |  -3.66 | 30 | < .05  | [-0.76, -0.26] |     Pearson
#> disp       |       carb |  0.39 |   2.35 | 30 | 0.30   | [ 0.05,  0.65] |     Pearson
#> hp         |       drat | -0.45 |  -2.75 | 30 | 0.16   | [-0.69, -0.12] |     Pearson
#> hp         |         wt |  0.66 |   4.80 | 30 | < .001 | [ 0.40,  0.82] |     Pearson
#> hp         |       qsec | -0.71 |  -5.49 | 30 | < .001 | [-0.85, -0.48] |     Pearson
#> hp         |         vs | -0.90 | -10.87 | 30 | < .001 | [-0.95, -0.80] |    Biserial
#> hp         |         am | -0.30 |  -1.71 | 30 | 1.00   | [-0.59,  0.05] |    Biserial
#> hp         |       gear | -0.13 |  -0.69 | 30 | 1.00   | [-0.45,  0.23] |     Pearson
#> hp         |       carb |  0.75 |   6.21 | 30 | < .001 | [ 0.54,  0.87] |     Pearson
#> drat       |         wt | -0.71 |  -5.56 | 30 | < .001 | [-0.85, -0.48] |     Pearson
#> drat       |       qsec |  0.09 |   0.50 | 30 | 1.00   | [-0.27,  0.43] |     Pearson
#> drat       |         vs |  0.55 |   3.51 | 30 | < .05  | [ 0.24,  0.75] |    Biserial
#> drat       |         am |  0.89 |  10.41 | 30 | < .001 | [ 0.78,  0.94] |    Biserial
#> drat       |       gear |  0.70 |   5.36 | 30 | < .001 | [ 0.46,  0.84] |     Pearson
#> drat       |       carb | -0.09 |  -0.50 | 30 | 1.00   | [-0.43,  0.27] |     Pearson
#> wt         |       qsec | -0.17 |  -0.97 | 30 | 1.00   | [-0.49,  0.19] |     Pearson
#> wt         |         vs | -0.69 |  -5.10 | 30 | < .001 | [-0.84, -0.45] |    Biserial
#> wt         |         am | -0.86 |  -9.20 | 30 | < .001 | [-0.93, -0.74] |    Biserial
#> wt         |       gear | -0.58 |  -3.93 | 30 | < .01  | [-0.77, -0.29] |     Pearson
#> wt         |       carb |  0.43 |   2.59 | 30 | 0.20   | [ 0.09,  0.68] |     Pearson
#> qsec       |         vs |  0.92 |  12.88 | 30 | < .001 | [ 0.85,  0.96] |    Biserial
#> qsec       |         am | -0.29 |  -1.61 | 30 | 1.00   | [-0.58,  0.07] |    Biserial
#> qsec       |       gear | -0.21 |  -1.19 | 30 | 1.00   | [-0.52,  0.15] |     Pearson
#> qsec       |       carb | -0.66 |  -4.76 | 30 | < .001 | [-0.82, -0.40] |     Pearson
#> vs         |         am |  0.26 |   1.47 | 30 | 1.00   | [-0.09,  0.56] | Tetrachoric
#> vs         |       gear |  0.26 |   1.42 | 30 | 1.00   | [-0.10,  0.55] |    Biserial
#> vs         |       carb | -0.71 |  -5.37 | 30 | < .001 | [-0.85, -0.47] |    Biserial
#> am         |       gear |  0.99 |  37.07 | 30 | < .001 | [ 0.98,  0.99] |    Biserial
#> am         |       carb |  0.07 |   0.39 | 30 | 1.00   | [-0.28,  0.41] |    Biserial
#> gear       |       carb |  0.27 |   1.56 | 30 | 1.00   | [-0.08,  0.57] |     Pearson