Correlation Analysis

Performs a correlation analysis.

correlation(
  data,
  data2 = NULL,
  method = "pearson",
  p_adjust = "holm",
  ci = 0.95,
  bayesian = FALSE,
  bayesian_prior = "medium",
  bayesian_ci_method = "hdi",
  bayesian_test = c("pd", "rope", "bf"),
  redundant = FALSE,
  include_factors = FALSE,
  partial = FALSE,
  partial_bayesian = FALSE,
  multilevel = FALSE,
  robust = FALSE,
  ...
)

Arguments

data
A dataframe.

data2
An optional dataframe.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson" (default), "kendall", or "spearman", "biserial", "polychoric", "tetrachoric", "biweight", "distance", "percentage" (for percentage bend correlation) or "shepherd" (for Shepherd's Pi correlation). Setting "auto" will attempt at selecting the most relevant method (polychoric when ordinal factors involved, tetrachoric when dichotomous factors involved, point-biserial if one dichotomous and one continuous and pearson otherwise).

p_adjust
Correction method for frequentist correlations. Can be one of "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

bayesian
If TRUE, will run the correlations under a Bayesian framework. Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

bayesian_prior
For the prior argument, several named values are recognized: "medium.narrow", "medium", "wide", and "ultrawide". These correspond to scale values of 1/sqrt(27), 1/3, 1/sqrt(3) and 1, respectively. See the BayesFactor::correlationBF function.

bayesian_ci_method
See arguments in model_parameters for BayesFactor tests.

bayesian_test
See arguments in model_parameters for BayesFactor tests.

redundant
Shoud the data include redundant rows (where each given correlation is repeated two times).

include_factors
If TRUE, the factors are kept and eventually converted to numeric or used as random effects (depending of multilevel). If FALSE, factors are removed upfront.

partial
Can be TRUE or "semi" for partial and semi-partial correlations, respectively. This only works for Frequentist correlations.

partial_bayesian
If TRUE, will run the correlations under a Bayesian framework. Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

multilevel
If TRUE, the factors are included as random factors. Else, if FALSE (default), they are included as fixed effects in the simple regression model.

robust
If TRUE, will rank-transform the variables prior to estimating the correlation. Note that, for instance, a Pearson's correlation on rank-transformed data is equivalent to a Spearman's rank correlation. Thus, using robust=TRUE and method="spearman" is redundant. Nonetheless, it is an easy way to increase the robustness of the correlation (as well as obtaining Bayesian Spearman rank Correlations).

...
Arguments passed to or from other methods.

Value

A correlation object that can be displayed using the print, summary or table methods.

Multiple tests correction

About multiple tests corrections.

Details

Correlation Types

Pearson's correlation: The covariance of the two variables divided by the product of their standard deviations.

Spearman's rank correlation: A nonparametric measure of rank correlation (statistical dependence between the rankings of two variables). The Spearman correlation between two variables is equal to the Pearson correlation between the rank values of those two variables; while Pearson's correlation assesses linear relationships, Spearman's correlation assesses monotonic relationships (whether linear or not).

Kendall's rank correlation: In the normal case, the Kendall correlation is preferred than the Spearman correlation because of a smaller gross error sensitivity (GES) and a smaller asymptotic variance (AV), making it more robust and more efficient. However, the interpretation of Kendall's tau is less direct than that of Spearman's rho, in the sense that it quantifies the difference between the % of concordant and discordant pairs among all possible pairwise events.

Biweight midcorrelation: A measure of similarity between samples that is median-based, rather than mean-based, thus is less sensitive to outliers, and can be a robust alternative to other similarity metrics, such as Pearson correlation.

Distance correlation: Distance correlation measures both linear and nonlinear association between two random variables or random vectors. This is in contrast to Pearson's correlation, which can only detect linear association between two random variables.

Percentage bend correlation: Introduced by Wilcox (1994), it is based on a down-weight of a specified percentage of marginal observations deviating from the median (by default, 20%).

Shepherd's Pi correlation: Equivalent to a Spearman's rank correlation after outliers removal (by means of bootstrapped mahalanobis distance).

Point-Biserial correlation: Correlation coefficient used when one variable is dichotomous.

Polychoric correlation: Correlation between two theorised normally distributed continuous latent variables, from two observed ordinal variables.

Tetrachoric correlation: Special case of the polychoric correlation applicable when both observed variables are dichotomous.

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

cor <- correlation(iris)

cor

#> Parameter1   |   Parameter2 |     r |         95% CI |     t |  df |      p |  Method | n_Obs
#> ---------------------------------------------------------------------------------------------
#> Sepal.Length |  Sepal.Width | -0.12 | [-0.27,  0.04] | -1.44 | 148 | 0.152  | Pearson |   150
#> Sepal.Length | Petal.Length |  0.87 | [ 0.83,  0.91] | 21.65 | 148 | < .001 | Pearson |   150
#> Sepal.Length |  Petal.Width |  0.82 | [ 0.76,  0.86] | 17.30 | 148 | < .001 | Pearson |   150
#> Sepal.Width  | Petal.Length | -0.43 | [-0.55, -0.29] | -5.77 | 148 | < .001 | Pearson |   150
#> Sepal.Width  |  Petal.Width | -0.37 | [-0.50, -0.22] | -4.79 | 148 | < .001 | Pearson |   150
#> Petal.Length |  Petal.Width |  0.96 | [ 0.95,  0.97] | 43.39 | 148 | < .001 | Pearson |   150

summary(cor)

#> Parameter    | Petal.Width | Petal.Length | Sepal.Width
#> -------------------------------------------------------
#> Sepal.Length |     0.82*** |      0.87*** |       -0.12
#> Sepal.Width  |    -0.37*** |     -0.43*** |            
#> Petal.Length |     0.96*** |              |            

as.table(cor)

#> Parameter    | Sepal.Length | Sepal.Width | Petal.Length | Petal.Width
#> ----------------------------------------------------------------------
#> Sepal.Length |      1.00*** |       -0.12 |      0.87*** |     0.82***
#> Sepal.Width  |        -0.12 |     1.00*** |     -0.43*** |    -0.37***
#> Petal.Length |      0.87*** |    -0.43*** |      1.00*** |     0.96***
#> Petal.Width  |      0.82*** |    -0.37*** |      0.96*** |     1.00***

iris %>%
  group_by(Species) %>%
  correlation()

#> Group      |   Parameter1 |   Parameter2 |    r |        95% CI |     t | df |      p |  Method | n_Obs
#> -------------------------------------------------------------------------------------------------------
#> setosa     | Sepal.Length |  Sepal.Width | 0.74 | [ 0.59, 0.85] |  7.68 | 48 | < .001 | Pearson |    50
#> setosa     | Sepal.Length | Petal.Length | 0.27 | [-0.01, 0.51] |  1.92 | 48 | 0.202  | Pearson |    50
#> setosa     | Sepal.Length |  Petal.Width | 0.28 | [ 0.00, 0.52] |  2.01 | 48 | 0.202  | Pearson |    50
#> setosa     |  Sepal.Width | Petal.Length | 0.18 | [-0.11, 0.43] |  1.25 | 48 | 0.217  | Pearson |    50
#> setosa     |  Sepal.Width |  Petal.Width | 0.23 | [-0.05, 0.48] |  1.66 | 48 | 0.208  | Pearson |    50
#> setosa     | Petal.Length |  Petal.Width | 0.33 | [ 0.06, 0.56] |  2.44 | 48 | 0.093  | Pearson |    50
#> versicolor | Sepal.Length |  Sepal.Width | 0.53 | [ 0.29, 0.70] |  4.28 | 48 | < .001 | Pearson |    50
#> versicolor | Sepal.Length | Petal.Length | 0.75 | [ 0.60, 0.85] |  7.95 | 48 | < .001 | Pearson |    50
#> versicolor | Sepal.Length |  Petal.Width | 0.55 | [ 0.32, 0.72] |  4.52 | 48 | < .001 | Pearson |    50
#> versicolor |  Sepal.Width | Petal.Length | 0.56 | [ 0.33, 0.73] |  4.69 | 48 | < .001 | Pearson |    50
#> versicolor |  Sepal.Width |  Petal.Width | 0.66 | [ 0.47, 0.80] |  6.15 | 48 | < .001 | Pearson |    50
#> versicolor | Petal.Length |  Petal.Width | 0.79 | [ 0.65, 0.87] |  8.83 | 48 | < .001 | Pearson |    50
#> virginica  | Sepal.Length |  Sepal.Width | 0.46 | [ 0.20, 0.65] |  3.56 | 48 | 0.003  | Pearson |    50
#> virginica  | Sepal.Length | Petal.Length | 0.86 | [ 0.77, 0.92] | 11.90 | 48 | < .001 | Pearson |    50
#> virginica  | Sepal.Length |  Petal.Width | 0.28 | [ 0.00, 0.52] |  2.03 | 48 | 0.048  | Pearson |    50
#> virginica  |  Sepal.Width | Petal.Length | 0.40 | [ 0.14, 0.61] |  3.03 | 48 | 0.012  | Pearson |    50
#> virginica  |  Sepal.Width |  Petal.Width | 0.54 | [ 0.31, 0.71] |  4.42 | 48 | < .001 | Pearson |    50
#> virginica  | Petal.Length |  Petal.Width | 0.32 | [ 0.05, 0.55] |  2.36 | 48 | 0.045  | Pearson |    50

correlation(mtcars[-2], method = "auto")

#> Parameter1 | Parameter2 |     r |         95% CI |      t | df |      p |      Method | n_Obs
#> ---------------------------------------------------------------------------------------------
#> mpg        |       disp | -0.85 | [-0.92, -0.71] |  -8.75 | 30 | < .001 |     Pearson |    32
#> mpg        |         hp | -0.78 | [-0.89, -0.59] |  -6.74 | 30 | < .001 |     Pearson |    32
#> mpg        |       drat |  0.68 | [ 0.44,  0.83] |   5.10 | 30 | < .001 |     Pearson |    32
#> mpg        |         wt | -0.87 | [-0.93, -0.74] |  -9.56 | 30 | < .001 |     Pearson |    32
#> mpg        |       qsec |  0.42 | [ 0.08,  0.67] |   2.53 | 30 | 0.222  |     Pearson |    32
#> mpg        |         vs |  0.82 | [ 0.66,  0.91] |   7.80 | 30 | < .001 |    Biserial |    32
#> mpg        |         am |  0.75 | [ 0.54,  0.87] |   6.06 | 30 | < .001 |    Biserial |    32
#> mpg        |       gear |  0.48 | [ 0.16,  0.71] |   3.00 | 30 | 0.092  |     Pearson |    32
#> mpg        |       carb | -0.55 | [-0.75, -0.25] |  -3.62 | 30 | 0.021  |     Pearson |    32
#> disp       |         hp |  0.79 | [ 0.61,  0.89] |   7.08 | 30 | < .001 |     Pearson |    32
#> disp       |       drat | -0.71 | [-0.85, -0.48] |  -5.53 | 30 | < .001 |     Pearson |    32
#> disp       |         wt |  0.89 | [ 0.78,  0.94] |  10.58 | 30 | < .001 |     Pearson |    32
#> disp       |       qsec | -0.43 | [-0.68, -0.10] |  -2.64 | 30 | 0.197  |     Pearson |    32
#> disp       |         vs | -0.88 | [-0.94, -0.77] |  -9.99 | 30 | < .001 |    Biserial |    32
#> disp       |         am | -0.74 | [-0.86, -0.52] |  -5.87 | 30 | < .001 |    Biserial |    32
#> disp       |       gear | -0.56 | [-0.76, -0.26] |  -3.66 | 30 | 0.019  |     Pearson |    32
#> disp       |       carb |  0.39 | [ 0.05,  0.65] |   2.35 | 30 | 0.303  |     Pearson |    32
#> hp         |       drat | -0.45 | [-0.69, -0.12] |  -2.75 | 30 | 0.160  |     Pearson |    32
#> hp         |         wt |  0.66 | [ 0.40,  0.82] |   4.80 | 30 | < .001 |     Pearson |    32
#> hp         |       qsec | -0.71 | [-0.85, -0.48] |  -5.49 | 30 | < .001 |     Pearson |    32
#> hp         |         vs | -0.90 | [-0.95, -0.80] | -10.87 | 30 | < .001 |    Biserial |    32
#> hp         |         am | -0.30 | [-0.59,  0.05] |  -1.71 | 30 | 1.000  |    Biserial |    32
#> hp         |       gear | -0.13 | [-0.45,  0.23] |  -0.69 | 30 | 1.000  |     Pearson |    32
#> hp         |       carb |  0.75 | [ 0.54,  0.87] |   6.21 | 30 | < .001 |     Pearson |    32
#> drat       |         wt | -0.71 | [-0.85, -0.48] |  -5.56 | 30 | < .001 |     Pearson |    32
#> drat       |       qsec |  0.09 | [-0.27,  0.43] |   0.50 | 30 | 1.000  |     Pearson |    32
#> drat       |         vs |  0.55 | [ 0.24,  0.75] |   3.51 | 30 | 0.026  |    Biserial |    32
#> drat       |         am |  0.89 | [ 0.78,  0.94] |  10.41 | 30 | < .001 |    Biserial |    32
#> drat       |       gear |  0.70 | [ 0.46,  0.84] |   5.36 | 30 | < .001 |     Pearson |    32
#> drat       |       carb | -0.09 | [-0.43,  0.27] |  -0.50 | 30 | 1.000  |     Pearson |    32
#> wt         |       qsec | -0.17 | [-0.49,  0.19] |  -0.97 | 30 | 1.000  |     Pearson |    32
#> wt         |         vs | -0.69 | [-0.84, -0.45] |  -5.10 | 30 | < .001 |    Biserial |    32
#> wt         |         am | -0.86 | [-0.93, -0.74] |  -9.20 | 30 | < .001 |    Biserial |    32
#> wt         |       gear | -0.58 | [-0.77, -0.29] |  -3.93 | 30 | 0.010  |     Pearson |    32
#> wt         |       carb |  0.43 | [ 0.09,  0.68] |   2.59 | 30 | 0.205  |     Pearson |    32
#> qsec       |         vs |  0.92 | [ 0.85,  0.96] |  12.88 | 30 | < .001 |    Biserial |    32
#> qsec       |         am | -0.29 | [-0.58,  0.07] |  -1.61 | 30 | 1.000  |    Biserial |    32
#> qsec       |       gear | -0.21 | [-0.52,  0.15] |  -1.19 | 30 | 1.000  |     Pearson |    32
#> qsec       |       carb | -0.66 | [-0.82, -0.40] |  -4.76 | 30 | < .001 |     Pearson |    32
#> vs         |         am |  0.26 | [-0.09,  0.56] |   1.47 | 30 | 1.000  | Tetrachoric |    32
#> vs         |       gear |  0.26 | [-0.10,  0.55] |   1.42 | 30 | 1.000  |    Biserial |    32
#> vs         |       carb | -0.71 | [-0.85, -0.47] |  -5.37 | 30 | < .001 |    Biserial |    32
#> am         |       gear |  0.99 | [ 0.98,  0.99] |  37.07 | 30 | < .001 |    Biserial |    32
#> am         |       carb |  0.07 | [-0.28,  0.41] |   0.39 | 30 | 1.000  |    Biserial |    32
#> gear       |       carb |  0.27 | [-0.08,  0.57] |   1.56 | 30 | 1.000  |     Pearson |    32