Correlation Analysis

Performs a correlation analysis.

correlation(
  data,
  data2 = NULL,
  method = "pearson",
  p_adjust = "holm",
  ci = 0.95,
  bayesian = FALSE,
  bayesian_prior = "medium",
  bayesian_ci_method = "hdi",
  bayesian_test = c("pd", "rope", "bf"),
  redundant = FALSE,
  include_factors = FALSE,
  partial = FALSE,
  partial_bayesian = FALSE,
  multilevel = FALSE,
  robust = FALSE,
  ...
)

Arguments

data
A dataframe.

data2
An optional dataframe.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson" (default), "kendall", or "spearman", "biserial", "polychoric", "tetrachoric", "biweight", "distance", "percentage" (for percentage bend correlation) or "shepherd" (for Shepherd's Pi correlation). Setting "auto" will attempt at selecting the most relevant method (polychoric when ordinal factors involved, tetrachoric when dichotomous factors involved, point-biserial if one dichotomous and one continuous and pearson otherwise).

p_adjust
Correction method for frequentist correlations. Can be one of "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

bayesian
If TRUE, will run the correlations under a Bayesian framework. Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

bayesian_prior
For the prior argument, several named values are recognized: "medium.narrow", "medium", "wide", and "ultrawide". These correspond to scale values of 1/sqrt(27), 1/3, 1/sqrt(3) and 1, respectively. See the BayesFactor::correlationBF function.

bayesian_ci_method
See arguments in model_parameters for BayesFactor tests.

bayesian_test
See arguments in model_parameters for BayesFactor tests.

redundant
Shoud the data include redundant rows (where each given correlation is repeated two times).

include_factors
If TRUE, the factors are kept and eventually converted to numeric or used as random effects (depending of multilevel). If FALSE, factors are removed upfront.

partial
Can be TRUE or "semi" for partial and semi-partial correlations, respectively.

partial_bayesian
If TRUE, will run the correlations under a Bayesian framework. Note that for partial correlations, you will also need to set partial_bayesian to TRUE to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).

multilevel
If TRUE, the factors are included as random factors. Else, if FALSE (default), they are included as fixed effects in the simple regression model.

robust
If TRUE, will rank-transform the variables prior to estimating the correlation. Note that, for instance, a Pearson's correlation on rank-transformed data is equivalent to a Spearman's rank correlation. Thus, using robust=TRUE and method="spearman" is redundant. Nonetheless, it is an easy way to increase the robustness of the correlation (as well as obtaining Bayesian Spearman rank Correlations).

...
Arguments passed to or from other methods.

Value

A correlation object that can be displayed using the print, summary or table methods.

Multiple tests correction

About multiple tests corrections.

Details

Correlation Types

Pearson's correlation: The covariance of the two variables divided by the product of their standard deviations.

Spearman's rank correlation: A nonparametric measure of rank correlation (statistical dependence between the rankings of two variables). The Spearman correlation between two variables is equal to the Pearson correlation between the rank values of those two variables; while Pearson's correlation assesses linear relationships, Spearman's correlation assesses monotonic relationships (whether linear or not).

Kendall's rank correlation: In the normal case, the Kendall correlation is preferred than the Spearman correlation because of a smaller gross error sensitivity (GES) and a smaller asymptotic variance (AV), making it more robust and more efficient. However, the interpretation of Kendall's tau is less direct than that of Spearman's rho, in the sense that it quantifies the difference between the % of concordant and discordant pairs among all possible pairwise events.

Biweight midcorrelation: A measure of similarity between samples that is median-based, rather than mean-based, thus is less sensitive to outliers, and can be a robust alternative to other similarity metrics, such as Pearson correlation.

Distance correlation: Distance correlation measures both linear and nonlinear association between two random variables or random vectors. This is in contrast to Pearson's correlation, which can only detect linear association between two random variables.

Percentage bend correlation: Introduced by Wilcox (1994), it is based on a down-weight of a specified percentage of marginal observations deviating from the median (by default, 20%).

Shepherd's Pi correlation: Equivalent to a Spearman's rank correlation after outliers removal (by means of bootstrapped mahalanobis distance).

Point-Biserial and biserial correlation: Correlation coefficient used when one variable is continuous and the other is dichotomous (binary). Point-serial is equivalent to a Pearson's correlation, while Biserial should be used when the binary variable is assumed to have an underlying continuity. For example, anxiety level can be measured on a continuous scale, but can be classified dichotomously as high/low.

Polychoric correlation: Correlation between two theorised normally distributed continuous latent variables, from two observed ordinal variables.

Tetrachoric correlation: Special case of the polychoric correlation applicable when both observed variables are dichotomous.

Partial correlation: Correlation between two variables after adjusting for the (linear) the effect of one or more variable. The correlation test is here run after having partialized the dataset, independently from it. In other words, it considers partialization as an independent step generating a different dataset, rather than belonging to the same model. This is why small discrepancies are to be expected for the t- and the p-values (but not the correlation coefficient) compared to other implementations such as ppcor.

Multilevel correlation: Multilevel correlations are a special case of partial correlations where the variable to be adjusted for is a factor and is included as a random effect in a mixed model.

Notes

Kendall and Spearman correlations when bayesian=TRUE: These are technically Pearson Bayesian correlations of rank transformed data, rather than pure Bayesian rank correlations (which have different priors).

Examples

library(correlation)

cor <- correlation(iris)

cor

#> Parameter1   |   Parameter2 |     r |         95% CI |     t |  df |      p |  Method | n_Obs
#> ---------------------------------------------------------------------------------------------
#> Sepal.Length |  Sepal.Width | -0.12 | [-0.27,  0.04] | -1.44 | 148 | 0.152  | Pearson |   150
#> Sepal.Length | Petal.Length |  0.87 | [ 0.83,  0.91] | 21.65 | 148 | < .001 | Pearson |   150
#> Sepal.Length |  Petal.Width |  0.82 | [ 0.76,  0.86] | 17.30 | 148 | < .001 | Pearson |   150
#> Sepal.Width  | Petal.Length | -0.43 | [-0.55, -0.29] | -5.77 | 148 | < .001 | Pearson |   150
#> Sepal.Width  |  Petal.Width | -0.37 | [-0.50, -0.22] | -4.79 | 148 | < .001 | Pearson |   150
#> Petal.Length |  Petal.Width |  0.96 | [ 0.95,  0.97] | 43.39 | 148 | < .001 | Pearson |   150

summary(cor)

#> Parameter    | Petal.Width | Petal.Length | Sepal.Width
#> -------------------------------------------------------
#> Sepal.Length |     0.82*** |      0.87*** |       -0.12
#> Sepal.Width  |    -0.37*** |     -0.43*** |            
#> Petal.Length |     0.96*** |              |            

as.table(cor)

#> Parameter    | Sepal.Length | Sepal.Width | Petal.Length | Petal.Width
#> ----------------------------------------------------------------------
#> Sepal.Length |      1.00*** |       -0.12 |      0.87*** |     0.82***
#> Sepal.Width  |        -0.12 |     1.00*** |     -0.43*** |    -0.37***
#> Petal.Length |      0.87*** |    -0.43*** |      1.00*** |     0.96***
#> Petal.Width  |      0.82*** |    -0.37*** |      0.96*** |     1.00***

# Grouped dataframe
if (require("dplyr")) {
  library(dplyr)

  iris %>%
    group_by(Species) %>%
    correlation()

  correlation(mtcars[-2], method = "auto")
}

#> Loading required package: dplyr

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#> Parameter1 | Parameter2 |     r |         95% CI |     t | df |      p |         Method | n_Obs
#> -----------------------------------------------------------------------------------------------
#> mpg        |       disp | -0.85 | [-0.92, -0.71] | -8.75 | 30 | < .001 |        Pearson |    32
#> mpg        |         hp | -0.78 | [-0.89, -0.59] | -6.74 | 30 | < .001 |        Pearson |    32
#> mpg        |       drat |  0.68 | [ 0.44,  0.83] |  5.10 | 30 | < .001 |        Pearson |    32
#> mpg        |         wt | -0.87 | [-0.93, -0.74] | -9.56 | 30 | < .001 |        Pearson |    32
#> mpg        |       qsec |  0.42 | [ 0.08,  0.67] |  2.53 | 30 | 0.222  |        Pearson |    32
#> mpg        |         vs |  0.66 | [ 0.41,  0.82] |  4.86 | 30 | < .001 | Point-biserial |    32
#> mpg        |         am |  0.60 | [ 0.32,  0.78] |  4.11 | 30 | 0.007  | Point-biserial |    32
#> mpg        |       gear |  0.48 | [ 0.16,  0.71] |  3.00 | 30 | 0.097  |        Pearson |    32
#> mpg        |       carb | -0.55 | [-0.75, -0.25] | -3.62 | 30 | 0.021  |        Pearson |    32
#> disp       |         hp |  0.79 | [ 0.61,  0.89] |  7.08 | 30 | < .001 |        Pearson |    32
#> disp       |       drat | -0.71 | [-0.85, -0.48] | -5.53 | 30 | < .001 |        Pearson |    32
#> disp       |         wt |  0.89 | [ 0.78,  0.94] | 10.58 | 30 | < .001 |        Pearson |    32
#> disp       |       qsec | -0.43 | [-0.68, -0.10] | -2.64 | 30 | 0.197  |        Pearson |    32
#> disp       |         vs | -0.71 | [-0.85, -0.48] | -5.53 | 30 | < .001 | Point-biserial |    32
#> disp       |         am | -0.59 | [-0.78, -0.31] | -4.02 | 30 | 0.009  | Point-biserial |    32
#> disp       |       gear | -0.56 | [-0.76, -0.26] | -3.66 | 30 | 0.020  |        Pearson |    32
#> disp       |       carb |  0.39 | [ 0.05,  0.65] |  2.35 | 30 | 0.303  |        Pearson |    32
#> hp         |       drat | -0.45 | [-0.69, -0.12] | -2.75 | 30 | 0.170  |        Pearson |    32
#> hp         |         wt |  0.66 | [ 0.40,  0.82] |  4.80 | 30 | 0.001  |        Pearson |    32
#> hp         |       qsec | -0.71 | [-0.85, -0.48] | -5.49 | 30 | < .001 |        Pearson |    32
#> hp         |         vs | -0.72 | [-0.86, -0.50] | -5.73 | 30 | < .001 | Point-biserial |    32
#> hp         |         am | -0.24 | [-0.55,  0.12] | -1.37 | 30 | 1.000  | Point-biserial |    32
#> hp         |       gear | -0.13 | [-0.45,  0.23] | -0.69 | 30 | 1.000  |        Pearson |    32
#> hp         |       carb |  0.75 | [ 0.54,  0.87] |  6.21 | 30 | < .001 |        Pearson |    32
#> drat       |         wt | -0.71 | [-0.85, -0.48] | -5.56 | 30 | < .001 |        Pearson |    32
#> drat       |       qsec |  0.09 | [-0.27,  0.43] |  0.50 | 30 | 1.000  |        Pearson |    32
#> drat       |         vs |  0.44 | [ 0.11,  0.68] |  2.69 | 30 | 0.187  | Point-biserial |    32
#> drat       |         am |  0.71 | [ 0.48,  0.85] |  5.57 | 30 | < .001 | Point-biserial |    32
#> drat       |       gear |  0.70 | [ 0.46,  0.84] |  5.36 | 30 | < .001 |        Pearson |    32
#> drat       |       carb | -0.09 | [-0.43,  0.27] | -0.50 | 30 | 1.000  |        Pearson |    32
#> wt         |       qsec | -0.17 | [-0.49,  0.19] | -0.97 | 30 | 1.000  |        Pearson |    32
#> wt         |         vs | -0.55 | [-0.76, -0.26] | -3.65 | 30 | 0.020  | Point-biserial |    32
#> wt         |         am | -0.69 | [-0.84, -0.45] | -5.26 | 30 | < .001 | Point-biserial |    32
#> wt         |       gear | -0.58 | [-0.77, -0.29] | -3.93 | 30 | 0.011  |        Pearson |    32
#> wt         |       carb |  0.43 | [ 0.09,  0.68] |  2.59 | 30 | 0.205  |        Pearson |    32
#> qsec       |         vs |  0.74 | [ 0.53,  0.87] |  6.11 | 30 | < .001 | Point-biserial |    32
#> qsec       |         am | -0.23 | [-0.54,  0.13] | -1.29 | 30 | 1.000  | Point-biserial |    32
#> qsec       |       gear | -0.21 | [-0.52,  0.15] | -1.19 | 30 | 1.000  |        Pearson |    32
#> qsec       |       carb | -0.66 | [-0.82, -0.40] | -4.76 | 30 | 0.001  |        Pearson |    32
#> vs         |         am |  0.26 | [-0.09,  0.56] |  1.47 | 30 | 1.000  |    Tetrachoric |    32
#> vs         |       gear |  0.21 | [-0.15,  0.52] |  1.15 | 30 | 1.000  | Point-biserial |    32
#> vs         |       carb | -0.57 | [-0.77, -0.28] | -3.80 | 30 | 0.015  | Point-biserial |    32
#> am         |       gear |  0.79 | [ 0.62,  0.89] |  7.16 | 30 | < .001 | Point-biserial |    32
#> am         |       carb |  0.06 | [-0.30,  0.40] |  0.32 | 30 | 1.000  | Point-biserial |    32
#> gear       |       carb |  0.27 | [-0.08,  0.57] |  1.56 | 30 | 1.000  |        Pearson |    32