Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(iris[, 1:4])

#> # Description of Posterior Distributions
#> 
#> Distance | CI_low | CI_high
#> ---------------------------
#>    2.191 |  1.594 |   3.017
#>    2.925 |  2.233 |   3.666
#>    2.150 |  1.620 |   2.758
#>    2.531 |  1.977 |   3.226
#>    2.525 |  1.789 |   3.402
#>    3.991 |  2.887 |   5.455
#>    2.952 |  2.322 |   3.735
#>    1.887 |  1.328 |   2.490
#>    3.494 |  2.686 |   4.335
#>    2.453 |  1.961 |   3.010
#>    3.374 |  2.432 |   4.527
#>    2.872 |  2.140 |   3.617
#>    2.694 |  2.133 |   3.369
#>    3.718 |  2.963 |   4.663
#>    8.979 |  6.914 |  11.683
#>    9.959 |  7.499 |  13.246
#>    5.896 |  4.434 |   7.827
#>    2.365 |  1.703 |   3.193
#>    4.615 |  3.433 |   6.163
#>    3.541 |  2.456 |   4.657
#>    2.727 |  2.105 |   3.476
#>    3.006 |  2.238 |   4.028
#>    3.731 |  2.940 |   4.692
#>    2.309 |  1.747 |   2.977
#>    5.471 |  4.219 |   6.927
#>    2.518 |  1.905 |   3.111
#>    1.815 |  1.317 |   2.442
#>    2.260 |  1.652 |   3.063
#>    2.568 |  1.868 |   3.319
#>    2.541 |  1.951 |   3.228
#>    2.043 |  1.535 |   2.548
#>    4.740 |  3.591 |   5.942
#>    8.585 |  6.624 |  11.145
#>    7.415 |  5.597 |  10.049
#>    2.047 |  1.544 |   2.592
#>    3.525 |  2.744 |   4.377
#>    5.500 |  4.235 |   6.929
#>    3.575 |  2.625 |   4.531
#>    3.257 |  2.558 |   4.091
#>    1.945 |  1.426 |   2.633
#>    2.606 |  1.927 |   3.463
#>   11.729 |  8.855 |  15.458
#>    3.427 |  2.749 |   4.298
#>    3.824 |  3.081 |   4.836
#>    4.619 |  3.539 |   6.070
#>    3.066 |  2.308 |   3.877
#>    4.550 |  3.358 |   5.884
#>    2.441 |  1.888 |   3.127
#>    3.093 |  2.211 |   4.150
#>    1.980 |  1.442 |   2.587
#>    4.664 |  3.337 |   6.277
#>    0.684 |  0.398 |   1.078
#>    3.163 |  2.113 |   4.402
#>    3.677 |  2.666 |   4.987
#>    2.126 |  1.513 |   2.936
#>    3.411 |  2.458 |   4.519
#>    1.337 |  0.857 |   1.961
#>    4.387 |  3.475 |   5.814
#>    2.841 |  1.850 |   3.970
#>    3.290 |  2.390 |   4.291
#>    7.810 |  5.950 |  10.313
#>    0.464 |  0.220 |   0.795
#>    7.777 |  5.781 |   9.853
#>    1.603 |  1.015 |   2.292
#>    1.091 |  0.702 |   1.515
#>    2.899 |  2.045 |   3.847
#>    3.601 |  2.538 |   4.703
#>    3.453 |  2.405 |   4.469
#>    7.581 |  5.677 |   9.648
#>    2.232 |  1.593 |   2.888
#>    3.368 |  2.276 |   4.588
#>    1.418 |  0.944 |   1.937
#>    2.454 |  1.703 |   3.245
#>    4.674 |  3.327 |   6.327
#>    1.679 |  1.060 |   2.363
#>    2.435 |  1.706 |   3.288
#>    4.333 |  3.042 |   5.859
#>    1.511 |  0.989 |   2.121
#>    0.340 |  0.193 |   0.562
#>    2.174 |  1.522 |   2.935
#>    2.848 |  2.055 |   3.811
#>    3.198 |  2.343 |   4.155
#>    0.888 |  0.558 |   1.281
#>    2.580 |  1.873 |   3.447
#>    6.421 |  4.818 |   8.236
#>    2.876 |  2.021 |   3.957
#>    1.901 |  1.224 |   2.717
#>    6.744 |  5.036 |   8.653
#>    1.727 |  1.127 |   2.352
#>    2.038 |  1.497 |   2.832
#>    5.075 |  3.704 |   6.570
#>    1.234 |  0.788 |   1.845
#>    1.409 |  0.943 |   1.920
#>    4.639 |  3.400 |   6.149
#>    1.532 |  1.083 |   2.130
#>    2.760 |  1.851 |   3.661
#>    1.226 |  0.787 |   1.734
#>    0.694 |  0.346 |   1.083
#>    4.953 |  3.737 |   6.660
#>    0.657 |  0.404 |   1.016
#>    9.222 |  6.809 |  12.466
#>    3.061 |  2.132 |   4.100
#>    2.556 |  1.818 |   3.406
#>    3.680 |  2.694 |   4.962
#>    2.436 |  1.647 |   3.357
#>    6.325 |  4.233 |   8.553
#>   10.476 |  7.924 |  13.142
#>    7.876 |  5.624 |  10.951
#>    3.448 |  2.568 |   4.625
#>    7.640 |  5.649 |  10.286
#>    2.364 |  1.571 |   3.570
#>    1.297 |  0.944 |   1.653
#>    2.538 |  1.821 |   3.660
#>    4.946 |  3.682 |   6.530
#>   11.876 |  8.617 |  15.604
#>    6.074 |  4.286 |   8.705
#>    2.063 |  1.360 |   2.816
#>   13.258 |  9.092 |  17.771
#>    7.441 |  5.597 |   9.954
#>    4.504 |  3.380 |   5.725
#>    4.186 |  2.990 |   6.047
#>    5.591 |  4.144 |   7.587
#>    9.160 |  6.360 |  12.335
#>    1.872 |  1.344 |   2.588
#>    2.982 |  2.190 |   4.237
#>    6.122 |  4.079 |   8.466
#>    1.507 |  1.038 |   2.169
#>    1.180 |  0.682 |   1.740
#>    1.733 |  1.245 |   2.367
#>    7.345 |  5.209 |  10.241
#>    5.916 |  4.294 |   8.265
#>   13.640 |  9.367 |  18.321
#>    2.639 |  1.816 |   3.646
#>    2.497 |  1.557 |   3.373
#>   13.272 | 10.005 |  17.727
#>   10.066 |  7.811 |  13.058
#>    8.451 |  5.908 |  11.397
#>    3.360 |  2.410 |   4.534
#>    1.533 |  0.936 |   2.258
#>    3.630 |  2.550 |   5.076
#>    6.178 |  4.438 |   8.876
#>   12.779 |  9.181 |  17.337
#>    3.061 |  2.132 |   4.100
#>    3.212 |  2.345 |   4.682
#>    7.693 |  5.420 |  10.865
#>    9.307 |  6.662 |  12.862
#>    4.169 |  3.175 |   5.530
#>    1.843 |  1.151 |   2.722
#>    7.910 |  5.728 |  10.843
#>    3.596 |  2.511 |   4.746
#> 

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688