Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(iris[, 1:4])

#> # Description of Posterior Distributions
#> 
#> Distance | CI_low | CI_high
#> ---------------------------
#>    2.194 |  1.531 |   2.975
#>    2.945 |  2.213 |   3.673
#>    2.167 |  1.632 |   2.764
#>    2.543 |  1.923 |   3.182
#>    2.529 |  1.766 |   3.422
#>    3.973 |  2.940 |   5.527
#>    2.993 |  2.297 |   3.689
#>    1.890 |  1.344 |   2.538
#>    3.493 |  2.751 |   4.391
#>    2.456 |  1.936 |   3.008
#>    3.366 |  2.490 |   4.655
#>    2.898 |  2.153 |   3.653
#>    2.692 |  2.111 |   3.332
#>    3.720 |  2.891 |   4.587
#>    8.971 |  6.596 |  11.603
#>    9.971 |  7.330 |  13.035
#>    5.934 |  4.386 |   7.826
#>    2.387 |  1.640 |   3.188
#>    4.636 |  3.258 |   6.061
#>    3.532 |  2.489 |   4.688
#>    2.722 |  2.167 |   3.505
#>    3.027 |  2.170 |   4.002
#>    3.745 |  2.838 |   4.708
#>    2.322 |  1.721 |   2.962
#>    5.471 |  4.284 |   6.949
#>    2.521 |  1.973 |   3.161
#>    1.830 |  1.288 |   2.409
#>    2.258 |  1.580 |   3.024
#>    2.575 |  1.975 |   3.472
#>    2.561 |  1.919 |   3.210
#>    2.052 |  1.590 |   2.614
#>    4.752 |  3.597 |   6.025
#>    8.652 |  6.543 |  11.074
#>    7.405 |  5.418 |   9.778
#>    2.041 |  1.553 |   2.585
#>    3.534 |  2.756 |   4.419
#>    5.497 |  4.307 |   7.131
#>    3.584 |  2.662 |   4.601
#>    3.259 |  2.458 |   4.022
#>    1.941 |  1.428 |   2.627
#>    2.631 |  1.876 |   3.455
#>   11.673 |  8.604 |  15.189
#>    3.450 |  2.743 |   4.297
#>    3.857 |  3.007 |   4.799
#>    4.639 |  3.443 |   5.900
#>    3.086 |  2.294 |   3.918
#>    4.574 |  3.397 |   5.993
#>    2.451 |  1.814 |   3.087
#>    3.070 |  2.230 |   4.257
#>    1.986 |  1.435 |   2.570
#>    4.660 |  3.257 |   6.185
#>    0.687 |  0.383 |   1.046
#>    3.157 |  2.166 |   4.434
#>    3.659 |  2.673 |   4.860
#>    2.120 |  1.403 |   2.855
#>    3.418 |  2.404 |   4.548
#>    1.343 |  0.867 |   1.884
#>    4.392 |  3.361 |   5.673
#>    2.857 |  1.881 |   3.941
#>    3.278 |  2.383 |   4.289
#>    7.793 |  6.032 |  10.246
#>    0.466 |  0.195 |   0.764
#>    7.693 |  5.901 |  10.083
#>    1.606 |  1.027 |   2.298
#>    1.077 |  0.715 |   1.525
#>    2.939 |  2.069 |   3.932
#>    3.565 |  2.664 |   4.691
#>    3.444 |  2.500 |   4.520
#>    7.597 |  5.809 |   9.816
#>    2.207 |  1.629 |   3.025
#>    3.375 |  2.392 |   4.550
#>    1.404 |  0.893 |   1.928
#>    2.448 |  1.764 |   3.337
#>    4.656 |  3.320 |   6.238
#>    1.678 |  1.087 |   2.367
#>    2.448 |  1.667 |   3.299
#>    4.382 |  2.948 |   5.753
#>    1.493 |  0.965 |   2.145
#>    0.345 |  0.176 |   0.553
#>    2.145 |  1.556 |   2.992
#>    2.844 |  2.095 |   3.835
#>    3.164 |  2.429 |   4.302
#>    0.875 |  0.572 |   1.313
#>    2.587 |  1.789 |   3.460
#>    6.387 |  4.953 |   8.172
#>    2.896 |  2.081 |   3.830
#>    1.899 |  1.202 |   2.679
#>    6.733 |  5.002 |   8.681
#>    1.722 |  1.231 |   2.440
#>    2.032 |  1.407 |   2.691
#>    5.110 |  3.776 |   6.719
#>    1.229 |  0.761 |   1.824
#>    1.394 |  0.941 |   1.973
#>    4.630 |  3.512 |   6.122
#>    1.529 |  1.067 |   2.097
#>    2.769 |  1.958 |   3.772
#>    1.221 |  0.809 |   1.774
#>    0.689 |  0.364 |   1.095
#>    4.936 |  3.666 |   6.441
#>    0.659 |  0.387 |   0.952
#>    9.211 |  6.563 |  12.083
#>    3.073 |  2.199 |   4.080
#>    2.552 |  1.799 |   3.438
#>    3.652 |  2.649 |   4.880
#>    2.435 |  1.665 |   3.319
#>    6.298 |  4.083 |   8.516
#>   10.449 |  7.996 |  13.167
#>    7.828 |  5.565 |  10.718
#>    3.417 |  2.553 |   4.615
#>    7.608 |  5.768 |  10.331
#>    2.361 |  1.518 |   3.345
#>    1.284 |  0.954 |   1.704
#>    2.521 |  1.765 |   3.569
#>    4.911 |  3.543 |   6.224
#>   11.846 |  8.651 |  15.512
#>    6.092 |  4.121 |   8.350
#>    2.061 |  1.329 |   2.799
#>   13.234 |  8.966 |  17.286
#>    7.415 |  5.494 |   9.821
#>    4.497 |  3.313 |   5.747
#>    4.208 |  2.823 |   5.665
#>    5.570 |  3.928 |   7.231
#>    9.113 |  6.522 |  12.488
#>    1.850 |  1.328 |   2.567
#>    3.004 |  2.083 |   4.004
#>    6.083 |  4.205 |   8.447
#>    1.503 |  0.965 |   2.105
#>    1.188 |  0.696 |   1.705
#>    1.728 |  1.212 |   2.302
#>    7.269 |  5.059 |   9.978
#>    5.911 |  4.088 |   8.081
#>   13.560 |  9.681 |  18.446
#>    2.632 |  1.789 |   3.593
#>    2.501 |  1.647 |   3.410
#>   13.200 |  9.724 |  17.163
#>    9.976 |  7.745 |  12.979
#>    8.455 |  6.102 |  11.336
#>    3.358 |  2.410 |   4.519
#>    1.538 |  0.892 |   2.175
#>    3.603 |  2.545 |   4.976
#>    6.183 |  4.160 |   8.408
#>   12.832 |  9.384 |  17.693
#>    3.073 |  2.199 |   4.080
#>    3.229 |  2.148 |   4.350
#>    7.724 |  5.280 |  10.460
#>    9.369 |  6.606 |  12.888
#>    4.151 |  3.129 |   5.373
#>    1.853 |  1.074 |   2.607
#>    7.927 |  5.735 |  10.662
#>    3.577 |  2.700 |   4.759
#> 

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688