Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A dataframe.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(data = iris[, 1:4])

#>       Distance     CI_low    CI_high
#> 1    2.1818990  1.4820848  2.9795163
#> 63   2.9291871  2.2529685  3.7803565
#> 74   2.1525199  1.5584474  2.7431538
#> 85   2.5323924  1.9519146  3.1703203
#> 96   2.5159740  1.7557286  3.4087846
#> 107  3.9971712  2.8058196  5.3803409
#> 118  2.9761756  2.3530638  3.7425456
#> 129  1.8841677  1.3316875  2.5579169
#> 140  3.4875385  2.7083250  4.3556599
#> 2    2.4423767  1.9730931  3.0177403
#> 13   3.3563884  2.4639707  4.6358445
#> 24   2.8876836  2.2220945  3.7293674
#> 35   2.6852262  2.0809564  3.3346428
#> 46   3.7181485  3.0081375  4.7138835
#> 57   9.0598275  6.9492167 11.8906740
#> 59   9.9907511  7.3643967 12.9919363
#> 60   5.9436400  4.3197247  7.7263019
#> 61   2.3891745  1.6811261  3.1984306
#> 62   4.6672172  3.4083446  6.1797628
#> 64   3.5431546  2.5315355  4.6956161
#> 65   2.7222501  2.1105582  3.5238479
#> 66   3.0147368  2.2546429  4.0721634
#> 67   3.7354126  2.9292754  4.7161059
#> 68   2.3158231  1.7058599  2.9997281
#> 69   5.4915003  4.2154384  6.9289075
#> 70   2.5112060  1.9088243  3.1779537
#> 71   1.8258497  1.3121902  2.4582785
#> 72   2.2455944  1.6199127  3.0682013
#> 73   2.5822548  1.8612365  3.3489861
#> 75   2.5507717  1.9770203  3.2515766
#> 76   2.0464103  1.5811744  2.6204475
#> 77   4.7619280  3.6496278  6.0813552
#> 78   8.6183879  6.4940419 10.9491952
#> 79   7.4452041  5.4353377  9.7887178
#> 80   2.0414453  1.5947997  2.6407362
#> 81   3.5240006  2.6525157  4.3941389
#> 82   5.4899401  4.3060430  6.9440725
#> 83   3.5660032  2.7372693  4.6649366
#> 84   3.2506732  2.5993754  4.1477331
#> 86   1.9455523  1.3609762  2.6237605
#> 87   2.6292897  1.9426993  3.4863694
#> 88  11.7326454  8.8812009 15.5656519
#> 89   3.4528482  2.7630982  4.3032858
#> 90   3.8489452  3.0657112  4.8351538
#> 91   4.6561007  3.5587077  5.9340872
#> 92   3.0609480  2.2487996  3.9192252
#> 93   4.5715689  3.3710826  5.8795756
#> 94   2.4442314  1.9137625  3.1586214
#> 95   3.0733892  2.1331009  4.1879555
#> 97   1.9859226  1.4383950  2.6043479
#> 98   4.6821750  3.4082478  6.3116384
#> 99   0.6918153  0.3562691  1.0404017
#> 100  3.1822347  2.2348977  4.4291955
#> 101  3.6545902  2.7344326  4.9867753
#> 102  2.1300780  1.5220715  3.0080698
#> 103  3.4144873  2.4219042  4.5098846
#> 104  1.3462568  0.8600181  1.9266314
#> 105  4.4106165  3.4039425  5.7744364
#> 106  2.8822926  2.0204034  3.9800324
#> 108  3.2937900  2.3406867  4.2341267
#> 109  7.7764395  5.9012576 10.1000618
#> 110  0.4683252  0.1880124  0.7490009
#> 111  7.7159041  6.1295069 10.2505267
#> 112  1.5965037  1.1026426  2.3656427
#> 113  1.0983088  0.7526785  1.5645550
#> 114  2.9404934  2.0529149  3.8529002
#> 115  3.6057927  2.6826589  4.7372089
#> 116  3.4389982  2.5948813  4.5749867
#> 117  7.5579466  5.8619379  9.9337573
#> 119  2.2308452  1.6858002  3.0646353
#> 120  3.3760555  2.3340478  4.5384280
#> 121  1.4060764  0.9279879  1.9585474
#> 122  2.4525101  1.7818358  3.3327735
#> 123  4.6565942  3.4794155  6.2962459
#> 124  1.6766037  1.1210487  2.3559552
#> 125  2.4518171  1.6613079  3.2658873
#> 126  4.3803567  3.1808237  5.8797273
#> 127  1.5080117  0.9329638  2.1252960
#> 128  0.3453285  0.1845008  0.5557530
#> 130  2.1591699  1.5452482  3.0079233
#> 131  2.8402976  2.1473716  3.8854236
#> 132  3.1894739  2.3425313  4.2040498
#> 133  0.8807166  0.5614325  1.3080655
#> 134  2.5919033  1.8886204  3.5187292
#> 135  6.4265124  4.8711422  8.1065058
#> 136  2.9033029  2.0361911  3.8507454
#> 137  1.9065766  1.2458914  2.6821675
#> 138  6.7046533  5.1738966  8.8001671
#> 139  1.7262785  1.2318628  2.4225686
#> 141  2.0381863  1.4544627  2.7577081
#> 142  5.0934637  3.7613372  6.6388364
#> 143  1.2239620  0.7399287  1.7843970
#> 144  1.3995689  0.9526702  1.9714643
#> 145  4.6593809  3.4742778  6.1207864
#> 146  1.5324227  0.9932127  2.0567955
#> 147  2.7742319  1.9417428  3.7151737
#> 148  1.2287001  0.7900342  1.7401758
#> 149  0.6977073  0.3523517  1.0530568
#> 150  4.9643029  3.6332928  6.5898800
#> 3    0.6601987  0.3923025  0.9837687
#> 4    9.2355374  6.5144478 12.0444888
#> 5    3.0834580  2.1276626  4.0382030
#> 6    2.5852279  1.8292602  3.4177389
#> 7    3.6595716  2.6375802  4.8057456
#> 8    2.4571924  1.6900286  3.3287739
#> 9    6.3076547  4.4721031  8.6928330
#> 10  10.5166826  8.0185695 13.3355391
#> 11   7.8707118  5.3904539 10.6095713
#> 12   3.4226585  2.4685760  4.4833457
#> 14   7.6357743  5.8866571 10.3066197
#> 15   2.3775183  1.5376339  3.4365528
#> 16   1.2894292  0.9594762  1.7108211
#> 17   2.5612023  1.7416559  3.6591287
#> 18   4.9482680  3.7267126  6.3353537
#> 19  11.8955140  8.5129731 15.4213373
#> 20   6.1365761  4.0112839  8.2858831
#> 21   2.0445989  1.4071098  2.8572626
#> 22  13.1873888  9.6376568 18.0950940
#> 23   7.4562617  5.4798396  9.7230884
#> 25   4.4814276  3.4135387  5.7644241
#> 26   4.2241839  2.6896275  5.6611875
#> 27   5.5776723  3.7933810  7.2759300
#> 28   9.1562175  6.6024686 12.2271650
#> 29   1.8711577  1.3295307  2.5850523
#> 30   2.9887187  2.0801650  4.0274177
#> 31   6.0675886  3.9711759  8.3525907
#> 32   1.5063915  1.0094150  2.1507189
#> 33   1.1907258  0.6963304  1.7455118
#> 34   1.7456458  1.2632318  2.3863014
#> 36   7.3318091  5.1916250 10.0791594
#> 37   5.9338365  4.2412360  7.8669905
#> 38  13.5478540  9.6285294 18.1786312
#> 39   2.6489136  1.8257508  3.6524362
#> 40   2.4812957  1.6414217  3.3934399
#> 41  13.2284530 10.0777404 17.2568849
#> 42  10.0708089  7.5689680 13.1310963
#> 43   8.4700710  6.1438290 11.4528911
#> 44   3.3604271  2.4509530  4.5458458
#> 45   1.5291780  0.9530703  2.2753130
#> 47   3.6359843  2.4302517  4.9983082
#> 48   6.2401432  4.0661262  8.3948176
#> 49  12.7726684  9.4263587 17.5836174
#> 50   3.0834580  2.1276626  4.0382030
#> 51   3.2312385  2.1353184  4.3649308
#> 52   7.7940334  5.2952981 10.4361409
#> 53   9.4106775  6.7315037 12.7797513
#> 54   4.1833915  3.1193523  5.3782148
#> 55   1.8685954  1.0767542  2.6322983
#> 56   7.9436292  5.6976951 10.7149226
#> 58   3.5972247  2.6489620  4.8127617