Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Additional arguments (e.g., alternative) to be passed to other methods. See stats::cor.test for further details.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

library(correlation)

distance_mahalanobis(iris[, 1:4])
#>       Distance    CI_low    CI_high
#> 1    2.1481091 1.4592783  2.9312932
#> 2    2.9258419 2.1566795  3.7350165
#> 3    2.1445046 1.5638903  2.7413676
#> 4    2.5291185 1.9267164  3.1410498
#> 5    2.4906721 1.7619346  3.4569255
#> 6    3.9270907 2.7042340  5.3351660
#> 7    2.9415369 2.2614501  3.6437293
#> 8    1.8564764 1.2879478  2.4963284
#> 9    3.5029670 2.7430351  4.3873137
#> 10   2.4543236 1.9336762  3.0440367
#> 11   3.3290538 2.2535472  4.3514957
#> 12   2.8328959 2.1054764  3.5849722
#> 13   2.6924870 2.0440525  3.3762916
#> 14   3.7259430 2.8856591  4.6061920
#> 15   8.9291809 6.8124450 11.2835099
#> 16   9.8425503 7.4552324 12.9561898
#> 17   5.9233171 4.4830845  7.7403911
#> 18   2.3543503 1.6871276  3.1810875
#> 19   4.5940878 3.2805281  5.8526661
#> 20   3.4813652 2.4944942  4.7022664
#> 21   2.6988186 2.0194076  3.3967534
#> 22   2.9865071 2.1942698  4.0031097
#> 23   3.7082265 2.8514052  4.5819827
#> 24   2.3134114 1.7359223  2.9885692
#> 25   5.4664208 4.2560534  6.8160385
#> 26   2.5173894 1.9126711  3.2120180
#> 27   1.8035089 1.2582786  2.3571036
#> 28   2.2123774 1.5913314  3.0154841
#> 29   2.5496873 1.8894037  3.3530343
#> 30   2.5237783 1.9495910  3.1931510
#> 31   2.0501046 1.5429390  2.5859123
#> 32   4.7325833 3.6698520  5.9261209
#> 33   8.4872853 6.5880571 11.0629871
#> 34   7.3096035 5.4818932  9.9175439
#> 35   2.0402848 1.5511140  2.6834310
#> 36   3.5147768 2.6635044  4.4145017
#> 37   5.4541219 4.3146536  6.8436975
#> 38   3.4970283 2.6470409  4.4967092
#> 39   3.2646843 2.4987653  4.0498372
#> 40   1.9177597 1.3453089  2.5805586
#> 41   2.5888237 1.8888984  3.3983479
#> 42  11.8435659 8.6842897 15.4087834
#> 43   3.4000750 2.7508745  4.2293195
#> 44   3.8465680 2.9330684  4.7026696
#> 45   4.5474040 3.5341203  5.9365153
#> 46   3.0695379 2.3070992  4.0047440
#> 47   4.4824825 3.4600779  5.8787101
#> 48   2.4263825 1.8200713  3.0220418
#> 49   3.0338063 2.2018509  4.2535641
#> 50   1.9618639 1.4085227  2.6146634
#> 51   4.5894103 3.3371688  6.0528606
#> 52   0.6617733 0.3614879  1.0261144
#> 53   3.1135630 2.2173975  4.3575078
#> 54   3.7473671 2.8322115  5.1426984
#> 55   2.1263566 1.4690476  2.8486844
#> 56   3.4168806 2.4782237  4.5945986
#> 57   1.3359132 0.8891004  1.9442137
#> 58   4.4527679 3.3087536  5.6034550
#> 59   2.8486067 1.9571350  3.9021983
#> 60   3.3103369 2.4181402  4.2921453
#> 61   7.9087973 6.1531423 10.4626934
#> 62   0.4737280 0.2316305  0.7855733
#> 63   7.8931030 5.8969785  9.8576132
#> 64   1.5984358 1.0614570  2.2843856
#> 65   1.1159154 0.7483722  1.5702565
#> 66   2.8796241 2.1669988  3.8440978
#> 67   3.5953086 2.6290262  4.7061036
#> 68   3.4661594 2.6008148  4.6124555
#> 69   7.6360366 5.7562466  9.7377296
#> 70   2.2689593 1.6522007  2.9833687
#> 71   3.4311244 2.3711571  4.6161267
#> 72   1.4350213 1.0095162  2.0295907
#> 73   2.4842872 1.7822484  3.2321076
#> 74   4.6700008 3.2901109  6.1958165
#> 75   1.6740725 1.1711535  2.3172775
#> 76   2.4167442 1.7596814  3.2305212
#> 77   4.3344480 3.1842397  5.7216958
#> 78   1.4768902 0.9572565  2.0537597
#> 79   0.3523745 0.1781728  0.5599385
#> 80   2.1995224 1.4810613  2.9493810
#> 81   2.9186112 2.1661846  3.9367048
#> 82   3.2433774 2.4827271  4.3152252
#> 83   0.9052270 0.5434003  1.3024506
#> 84   2.5926014 1.8356575  3.4479139
#> 85   6.4232909 4.9997368  8.2888260
#> 86   2.9060959 2.0150029  3.8566659
#> 87   1.8573813 1.2425886  2.5862226
#> 88   6.8133850 5.1200009  8.6265860
#> 89   1.7336875 1.2570995  2.3953883
#> 90   2.0903977 1.4740134  2.8012667
#> 91   5.1006497 3.7955493  6.7535535
#> 92   1.2245486 0.7984601  1.8359558
#> 93   1.4343138 0.9497211  1.9751088
#> 94   4.7177431 3.5134123  6.1784422
#> 95   1.5441241 1.0849686  2.1380119
#> 96   2.7913992 1.9808395  3.7222491
#> 97   1.2446036 0.8387111  1.7905892
#> 98   0.6970672 0.4132971  1.0901891
#> 99   5.0622568 3.5702110  6.5657851
#> 100  0.6739076 0.4184095  1.0143292
#> 101  9.3278887 6.8058001 12.4133841
#> 102  3.0956821 2.2795570  4.1961787
#> 103  2.5342140 1.8177185  3.3247742
#> 104  3.6505467 2.5764663  4.7857765
#> 105  2.4760518 1.7717634  3.4326512
#> 106  6.2010753 4.2800942  8.3774973
#> 107 10.5344766 8.0941710 13.3429018
#> 108  7.7964663 5.5536045 10.6031734
#> 109  3.4377456 2.6111054  4.5270467
#> 110  7.6435208 5.7054712 10.2004968
#> 111  2.3497445 1.4558162  3.3830714
#> 112  1.2969704 0.9427634  1.6908307
#> 113  2.5135555 1.6979616  3.5563799
#> 114  4.9882240 3.6724823  6.3394160
#> 115 11.8237535 8.9170672 16.1422685
#> 116  6.1314003 4.1291051  8.5298179
#> 117  2.0362603 1.2736485  2.7620975
#> 118 13.2073778 9.3198183 17.6641999
#> 119  7.3596262 5.6760327  9.6270652
#> 120  4.5226358 3.4347832  5.7383410
#> 121  4.1640578 2.8135072  5.7853752
#> 122  5.5966157 4.1040375  7.5727674
#> 123  9.0181589 6.3743149 12.0231447
#> 124  1.8888622 1.3785735  2.6586418
#> 125  3.0189031 2.0998702  4.0035478
#> 126  6.0298218 3.9254286  8.2634390
#> 127  1.5391033 0.9957246  2.1884376
#> 128  1.2023265 0.7285813  1.8211460
#> 129  1.7467528 1.2308383  2.3530633
#> 130  7.2505497 5.1338191 10.0087192
#> 131  5.8390486 4.0729568  7.8070614
#> 132 13.4396800 9.3967187 17.9398168
#> 133  2.6599010 1.8331143  3.6556951
#> 134  2.4874846 1.6795616  3.3813573
#> 135 13.3047839 9.8405583 17.4563715
#> 136 10.0259097 7.5397212 12.4823200
#> 137  8.5513755 5.9430381 11.3449770
#> 138  3.3754109 2.2878837  4.4267930
#> 139  1.5557857 0.9339485  2.2803569
#> 140  3.6146262 2.4021307  4.9428211
#> 141  6.1936804 4.0023395  8.4897025
#> 142 12.7967366 9.2263992 17.5838369
#> 143  3.0956821 2.2795570  4.1961787
#> 144  3.2318011 2.2886715  4.5010442
#> 145  7.7457755 5.3348688 10.7161391
#> 146  9.3705467 6.4301473 12.8068379
#> 147  4.1869817 3.0632001  5.5067953
#> 148  1.8546543 1.1221829  2.6941815
#> 149  8.0181345 5.5332503 10.6220901
#> 150  3.6128294 2.6615942  4.8046991
distance_mahalanobis(iris[, 1:4], robust = FALSE)
#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688