Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

library(correlation)

distance_mahalanobis(iris[, 1:4])

#>       Distance    CI_low    CI_high
#> 1    2.1939486 1.5308144  2.9753712
#> 2    2.9452515 2.2134276  3.6732618
#> 3    2.1668118 1.6319719  2.7644472
#> 4    2.5432801 1.9228201  3.1819630
#> 5    2.5293043 1.7656174  3.4220359
#> 6    3.9730984 2.9401411  5.5271414
#> 7    2.9933759 2.2971980  3.6889041
#> 8    1.8901873 1.3436557  2.5381783
#> 9    3.4933898 2.7511624  4.3909201
#> 10   2.4555605 1.9356989  3.0084704
#> 11   3.3662109 2.4898272  4.6554997
#> 12   2.8976543 2.1533469  3.6528873
#> 13   2.6921140 2.1110313  3.3319651
#> 14   3.7199463 2.8906877  4.5865999
#> 15   8.9713164 6.5957812 11.6025229
#> 16   9.9706247 7.3304510 13.0346220
#> 17   5.9336448 4.3856773  7.8260912
#> 18   2.3869550 1.6397193  3.1875473
#> 19   4.6364251 3.2582627  6.0609483
#> 20   3.5318680 2.4892226  4.6884311
#> 21   2.7224574 2.1672004  3.5054736
#> 22   3.0267070 2.1698619  4.0022861
#> 23   3.7450238 2.8378911  4.7084155
#> 24   2.3222132 1.7213319  2.9617809
#> 25   5.4714662 4.2839648  6.9492221
#> 26   2.5210586 1.9727354  3.1609579
#> 27   1.8303391 1.2877516  2.4086501
#> 28   2.2583711 1.5802980  3.0239152
#> 29   2.5748876 1.9751736  3.4723203
#> 30   2.5610855 1.9193322  3.2099096
#> 31   2.0516871 1.5902416  2.6142721
#> 32   4.7519701 3.5968808  6.0246417
#> 33   8.6518879 6.5430534 11.0744079
#> 34   7.4050105 5.4177914  9.7776921
#> 35   2.0409405 1.5527077  2.5845486
#> 36   3.5340359 2.7562705  4.4190058
#> 37   5.4974074 4.3069055  7.1309019
#> 38   3.5843390 2.6623997  4.6013000
#> 39   3.2594485 2.4578871  4.0224775
#> 40   1.9413190 1.4284872  2.6269944
#> 41   2.6306862 1.8763802  3.4549008
#> 42  11.6725268 8.6042395 15.1888644
#> 43   3.4497863 2.7433049  4.2970111
#> 44   3.8565852 3.0074768  4.7990322
#> 45   4.6390743 3.4434201  5.8997820
#> 46   3.0860226 2.2935933  3.9178434
#> 47   4.5742442 3.3973308  5.9932829
#> 48   2.4511521 1.8135431  3.0868539
#> 49   3.0695170 2.2299163  4.2571137
#> 50   1.9856238 1.4348849  2.5702315
#> 51   4.6604660 3.2566066  6.1852063
#> 52   0.6872800 0.3831024  1.0464472
#> 53   3.1574128 2.1657880  4.4340329
#> 54   3.6592819 2.6732812  4.8603654
#> 55   2.1203583 1.4034531  2.8551729
#> 56   3.4183887 2.4039631  4.5480538
#> 57   1.3425958 0.8667316  1.8835668
#> 58   4.3920195 3.3609039  5.6729148
#> 59   2.8568348 1.8809093  3.9406888
#> 60   3.2782919 2.3830908  4.2888026
#> 61   7.7925038 6.0322074 10.2460224
#> 62   0.4660386 0.1954309  0.7636576
#> 63   7.6927688 5.9014278 10.0827987
#> 64   1.6056025 1.0270438  2.2981090
#> 65   1.0770972 0.7152997  1.5248750
#> 66   2.9386292 2.0688775  3.9320360
#> 67   3.5653346 2.6643629  4.6914449
#> 68   3.4436142 2.4999829  4.5203431
#> 69   7.5969531 5.8092508  9.8156279
#> 70   2.2069421 1.6294029  3.0253462
#> 71   3.3746447 2.3916572  4.5499854
#> 72   1.4036323 0.8925736  1.9278677
#> 73   2.4481068 1.7644406  3.3374179
#> 74   4.6557102 3.3200461  6.2382704
#> 75   1.6778737 1.0873535  2.3673932
#> 76   2.4479116 1.6674787  3.2990814
#> 77   4.3822378 2.9483687  5.7534615
#> 78   1.4934706 0.9647435  2.1447363
#> 79   0.3446961 0.1755916  0.5530109
#> 80   2.1452892 1.5562475  2.9920725
#> 81   2.8436155 2.0952067  3.8346498
#> 82   3.1641901 2.4289974  4.3019030
#> 83   0.8748924 0.5721288  1.3128007
#> 84   2.5869229 1.7885177  3.4596067
#> 85   6.3866663 4.9531005  8.1721503
#> 86   2.8961020 2.0805078  3.8302547
#> 87   1.8988872 1.2021103  2.6788180
#> 88   6.7330845 5.0023002  8.6813404
#> 89   1.7224729 1.2314377  2.4402360
#> 90   2.0321967 1.4067406  2.6905294
#> 91   5.1095861 3.7763793  6.7187782
#> 92   1.2291413 0.7613621  1.8243814
#> 93   1.3935670 0.9412766  1.9734868
#> 94   4.6298967 3.5120965  6.1221439
#> 95   1.5289116 1.0666091  2.0974078
#> 96   2.7685252 1.9578467  3.7722401
#> 97   1.2206453 0.8091156  1.7744576
#> 98   0.6888563 0.3641800  1.0952799
#> 99   4.9358786 3.6657637  6.4414380
#> 100  0.6586209 0.3874673  0.9523366
#> 101  9.2112120 6.5627805 12.0834063
#> 102  3.0729348 2.1993967  4.0799621
#> 103  2.5517192 1.7991153  3.4376638
#> 104  3.6515231 2.6492010  4.8803950
#> 105  2.4348025 1.6652325  3.3188207
#> 106  6.2981014 4.0827058  8.5162745
#> 107 10.4491686 7.9961901 13.1668571
#> 108  7.8282403 5.5649740 10.7178347
#> 109  3.4167699 2.5526827  4.6153999
#> 110  7.6084792 5.7676300 10.3307070
#> 111  2.3606395 1.5177669  3.3454925
#> 112  1.2836246 0.9541174  1.7037263
#> 113  2.5209330 1.7647864  3.5691865
#> 114  4.9110357 3.5427504  6.2239380
#> 115 11.8463490 8.6507845 15.5115732
#> 116  6.0922405 4.1210404  8.3495988
#> 117  2.0613705 1.3291637  2.7992665
#> 118 13.2344493 8.9664602 17.2863665
#> 119  7.4150422 5.4937964  9.8211676
#> 120  4.4966255 3.3133594  5.7467718
#> 121  4.2082401 2.8230115  5.6649414
#> 122  5.5695393 3.9284360  7.2305382
#> 123  9.1130762 6.5217353 12.4881840
#> 124  1.8498844 1.3279196  2.5671476
#> 125  3.0040663 2.0829517  4.0041430
#> 126  6.0832511 4.2054341  8.4467746
#> 127  1.5028665 0.9651014  2.1050331
#> 128  1.1884397 0.6963786  1.7049663
#> 129  1.7278698 1.2120644  2.3017444
#> 130  7.2689778 5.0590056  9.9782062
#> 131  5.9108757 4.0881282  8.0807906
#> 132 13.5596047 9.6812592 18.4458833
#> 133  2.6320558 1.7888792  3.5925956
#> 134  2.5005485 1.6465766  3.4099626
#> 135 13.2000685 9.7242430 17.1630973
#> 136  9.9762039 7.7446747 12.9785612
#> 137  8.4547406 6.1016908 11.3357995
#> 138  3.3583289 2.4104485  4.5190025
#> 139  1.5377789 0.8923430  2.1747399
#> 140  3.6030714 2.5449082  4.9761278
#> 141  6.1830031 4.1596090  8.4076623
#> 142 12.8317732 9.3843961 17.6934095
#> 143  3.0729348 2.1993967  4.0799621
#> 144  3.2290246 2.1480285  4.3503904
#> 145  7.7241270 5.2801807 10.4604492
#> 146  9.3688087 6.6060650 12.8878478
#> 147  4.1506059 3.1292660  5.3730753
#> 148  1.8530080 1.0742977  2.6069025
#> 149  7.9274658 5.7350632 10.6622216
#> 150  3.5773009 2.7002708  4.7587281

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688