Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

Usage

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data

A data frame.

ci

Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations

The number of draws to simulate/bootstrap (when robust is TRUE).

robust

If TRUE, will run a bootstrapped version of the function with i iterations.

...

Additional arguments (e.g., alternative) to be passed to other methods. See stats::cor.test for further details.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

library(correlation)

distance_mahalanobis(iris[, 1:4])
#>       Distance    CI_low    CI_high
#> 1    2.1799257 1.5565268  3.0005158
#> 2    2.9457730 2.2071887  3.7279316
#> 3    2.1506407 1.5788732  2.7130278
#> 4    2.5325294 1.9760969  3.1575635
#> 5    2.5149781 1.7644678  3.3761117
#> 6    3.9657247 2.7191188  5.2368263
#> 7    2.9468683 2.3274479  3.7104656
#> 8    1.8639158 1.3330939  2.5289280
#> 9    3.4860917 2.7684179  4.4263499
#> 10   2.4537735 1.9305348  2.9788041
#> 11   3.3593890 2.4427008  4.5203573
#> 12   2.8425693 2.1678151  3.6395169
#> 13   2.6983058 2.1267922  3.3507177
#> 14   3.7166666 2.9578907  4.6326696
#> 15   8.9733668 6.5316451 11.3374712
#> 16   9.9467263 7.3458909 13.0912169
#> 17   5.9252088 4.3464840  7.6883999
#> 18   2.3806662 1.7007167  3.2082258
#> 19   4.6051144 3.4224489  6.0874924
#> 20   3.5271295 2.4990006  4.6739519
#> 21   2.7095779 2.1260588  3.4404804
#> 22   3.0207706 2.1800805  3.9953516
#> 23   3.7161789 2.9322653  4.7175945
#> 24   2.3238096 1.7590611  3.0023440
#> 25   5.4564414 4.1699234  6.7573547
#> 26   2.5281342 1.9616009  3.1847483
#> 27   1.8236534 1.2953759  2.4138267
#> 28   2.2489952 1.5889857  3.0205748
#> 29   2.5718080 1.8491238  3.3366623
#> 30   2.5366374 1.9909425  3.2221128
#> 31   2.0484693 1.5895522  2.5725768
#> 32   4.7269491 3.6303729  6.0076251
#> 33   8.5706975 6.4278360 11.1422379
#> 34   7.3877677 5.3003459  9.8083300
#> 35   2.0456671 1.5510468  2.5939517
#> 36   3.5345338 2.7137995  4.4442384
#> 37   5.4766686 4.2429858  6.9760850
#> 38   3.5109163 2.6657625  4.5961237
#> 39   3.2554156 2.5576174  4.1014267
#> 40   1.9387146 1.3960064  2.5913874
#> 41   2.6194170 1.8928812  3.4409972
#> 42  11.7863614 8.8501012 15.7068085
#> 43   3.4032997 2.6785643  4.2004219
#> 44   3.8613772 3.0792824  4.8779552
#> 45   4.5899426 3.4304733  5.9302715
#> 46   3.0839026 2.3353832  3.9464898
#> 47   4.5390887 3.4207923  5.9300312
#> 48   2.4261806 1.8536006  3.0345163
#> 49   3.0771802 2.1703114  4.1149199
#> 50   1.9844633 1.4555667  2.6434264
#> 51   4.6103747 3.2955369  6.2850574
#> 52   0.6795722 0.3400308  1.0296483
#> 53   3.1385767 2.2087193  4.4868648
#> 54   3.7089710 2.7119988  5.0786425
#> 55   2.1417018 1.4297935  2.8644936
#> 56   3.4191000 2.4920713  4.6997809
#> 57   1.3337254 0.8208172  1.9350031
#> 58   4.4519711 3.3272619  5.6772787
#> 59   2.8588760 1.9774487  3.8937580
#> 60   3.2781633 2.3705403  4.2002364
#> 61   7.8657602 5.9265701 10.4643920
#> 62   0.4634968 0.2236665  0.7981675
#> 63   7.7951035 5.8256717  9.8139529
#> 64   1.6243704 1.0305402  2.2956003
#> 65   1.0979633 0.7200742  1.5521055
#> 66   2.9081570 2.0329802  3.8626004
#> 67   3.5697116 2.6111768  4.6183484
#> 68   3.4848933 2.4551117  4.5102808
#> 69   7.6337844 5.7536222  9.8647715
#> 70   2.2568148 1.6396831  2.9831847
#> 71   3.3399420 2.4619254  4.5165117
#> 72   1.4377535 0.9213024  1.9379339
#> 73   2.4822089 1.7528175  3.1986569
#> 74   4.7215866 3.3047490  6.3021563
#> 75   1.6871957 1.1235225  2.3326461
#> 76   2.4333109 1.7105596  3.2905236
#> 77   4.3477249 3.1377206  5.7797524
#> 78   1.4977659 0.9992981  2.1842052
#> 79   0.3505193 0.1766845  0.5553932
#> 80   2.1828667 1.4958457  2.9583277
#> 81   2.8970522 2.0480978  3.8761762
#> 82   3.2252676 2.4177439  4.2948977
#> 83   0.8946361 0.5484268  1.3289850
#> 84   2.6070842 1.8937499  3.5595749
#> 85   6.3672633 4.8851233  8.1024579
#> 86   2.8859615 2.1585670  3.8785436
#> 87   1.8885087 1.2228334  2.7111885
#> 88   6.7902462 5.0812227  8.6984081
#> 89   1.7205175 1.2279298  2.3313129
#> 90   2.0705182 1.5119855  2.8286114
#> 91   5.1196048 3.6165439  6.6337846
#> 92   1.2482461 0.7199392  1.8006978
#> 93   1.4223322 0.9774686  2.0314986
#> 94   4.6913453 3.4701325  6.2559566
#> 95   1.5390349 1.0966953  2.1336801
#> 96   2.7850354 1.9412542  3.7650031
#> 97   1.2315192 0.8092259  1.7504006
#> 98   0.6995785 0.3752462  1.0680103
#> 99   5.0108885 3.6691528  6.7185950
#> 100  0.6664748 0.4276794  1.0086441
#> 101  9.2167262 6.8077604 12.3235808
#> 102  3.0545212 2.2230820  4.0753201
#> 103  2.5528745 1.8471697  3.5734980
#> 104  3.6633128 2.7186981  4.9365905
#> 105  2.4199144 1.7638294  3.3856355
#> 106  6.2758011 4.3188051  8.8457511
#> 107 10.4283920 8.2562609 13.5348806
#> 108  7.8433733 5.2026684 10.7935903
#> 109  3.4507908 2.5982188  4.5783842
#> 110  7.5958448 5.5034900  9.9386449
#> 111  2.3581056 1.4323647  3.4340731
#> 112  1.2937694 0.9392944  1.7051572
#> 113  2.5345987 1.6883455  3.6346206
#> 114  4.9581509 3.7332414  6.3920636
#> 115 11.8992951 8.8999065 15.6336546
#> 116  6.0823693 4.2557181  8.8405393
#> 117  2.0594187 1.3581319  2.8609737
#> 118 13.2696773 9.2236678 18.2199122
#> 119  7.4320882 5.4213278  9.7774102
#> 120  4.5449746 3.3835834  5.7589356
#> 121  4.2268802 2.7191179  5.8024147
#> 122  5.5093273 3.9378567  7.3366539
#> 123  9.0567039 6.3633989 12.4269079
#> 124  1.8808572 1.2449112  2.5734584
#> 125  2.9973252 2.1940395  4.0304378
#> 126  6.1232314 4.0372606  8.8210570
#> 127  1.5230020 1.0013380  2.2238927
#> 128  1.1781381 0.7496348  1.7975297
#> 129  1.7208109 1.2200212  2.3660658
#> 130  7.3306051 4.7571659 10.0359637
#> 131  5.9093631 4.0916594  7.9287270
#> 132 13.5974973 9.5881514 19.0567328
#> 133  2.6342064 1.7693688  3.6638668
#> 134  2.5188442 1.6212746  3.4015995
#> 135 13.4120791 9.8219535 17.5669325
#> 136  9.9872352 7.6709419 13.0041651
#> 137  8.4583659 6.2175371 11.6183776
#> 138  3.3586717 2.3748021  4.5193924
#> 139  1.5181257 0.9272936  2.2454588
#> 140  3.6351981 2.5239437  5.1647130
#> 141  6.2109366 4.2404709  8.8040648
#> 142 12.9498425 9.2639774 17.5221834
#> 143  3.0545212 2.2230820  4.0753201
#> 144  3.2139728 2.2090900  4.3981269
#> 145  7.7268304 4.9603326 10.4888015
#> 146  9.4445209 6.6447569 12.9734657
#> 147  4.2241636 2.9495056  5.4058515
#> 148  1.8589921 1.1457758  2.7609759
#> 149  7.9258215 5.5809661 10.6512128
#> 150  3.5667623 2.6968015  4.8034187
distance_mahalanobis(iris[, 1:4], robust = FALSE)
#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688