Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(iris[, 1:4])

#> # Description of Posterior Distributions
#> 
#> Distance | CI_low | CI_high
#> ---------------------------
#>    2.182 |  1.482 |   2.980
#>    2.929 |  2.253 |   3.780
#>    2.153 |  1.558 |   2.743
#>    2.532 |  1.952 |   3.170
#>    2.516 |  1.756 |   3.409
#>    3.997 |  2.806 |   5.380
#>    2.976 |  2.353 |   3.743
#>    1.884 |  1.332 |   2.558
#>    3.488 |  2.708 |   4.356
#>    2.442 |  1.973 |   3.018
#>    3.356 |  2.464 |   4.636
#>    2.888 |  2.222 |   3.729
#>    2.685 |  2.081 |   3.335
#>    3.718 |  3.008 |   4.714
#>    9.060 |  6.949 |  11.891
#>    9.991 |  7.364 |  12.992
#>    5.944 |  4.320 |   7.726
#>    2.389 |  1.681 |   3.198
#>    4.667 |  3.408 |   6.180
#>    3.543 |  2.532 |   4.696
#>    2.722 |  2.111 |   3.524
#>    3.015 |  2.255 |   4.072
#>    3.735 |  2.929 |   4.716
#>    2.316 |  1.706 |   3.000
#>    5.492 |  4.215 |   6.929
#>    2.511 |  1.909 |   3.178
#>    1.826 |  1.312 |   2.458
#>    2.246 |  1.620 |   3.068
#>    2.582 |  1.861 |   3.349
#>    2.551 |  1.977 |   3.252
#>    2.046 |  1.581 |   2.620
#>    4.762 |  3.650 |   6.081
#>    8.618 |  6.494 |  10.949
#>    7.445 |  5.435 |   9.789
#>    2.041 |  1.595 |   2.641
#>    3.524 |  2.653 |   4.394
#>    5.490 |  4.306 |   6.944
#>    3.566 |  2.737 |   4.665
#>    3.251 |  2.599 |   4.148
#>    1.946 |  1.361 |   2.624
#>    2.629 |  1.943 |   3.486
#>   11.733 |  8.881 |  15.566
#>    3.453 |  2.763 |   4.303
#>    3.849 |  3.066 |   4.835
#>    4.656 |  3.559 |   5.934
#>    3.061 |  2.249 |   3.919
#>    4.572 |  3.371 |   5.880
#>    2.444 |  1.914 |   3.159
#>    3.073 |  2.133 |   4.188
#>    1.986 |  1.438 |   2.604
#>    4.682 |  3.408 |   6.312
#>    0.692 |  0.356 |   1.040
#>    3.182 |  2.235 |   4.429
#>    3.655 |  2.734 |   4.987
#>    2.130 |  1.522 |   3.008
#>    3.414 |  2.422 |   4.510
#>    1.346 |  0.860 |   1.927
#>    4.411 |  3.404 |   5.774
#>    2.882 |  2.020 |   3.980
#>    3.294 |  2.341 |   4.234
#>    7.776 |  5.901 |  10.100
#>    0.468 |  0.188 |   0.749
#>    7.716 |  6.130 |  10.251
#>    1.597 |  1.103 |   2.366
#>    1.098 |  0.753 |   1.565
#>    2.940 |  2.053 |   3.853
#>    3.606 |  2.683 |   4.737
#>    3.439 |  2.595 |   4.575
#>    7.558 |  5.862 |   9.934
#>    2.231 |  1.686 |   3.065
#>    3.376 |  2.334 |   4.538
#>    1.406 |  0.928 |   1.959
#>    2.453 |  1.782 |   3.333
#>    4.657 |  3.479 |   6.296
#>    1.677 |  1.121 |   2.356
#>    2.452 |  1.661 |   3.266
#>    4.380 |  3.181 |   5.880
#>    1.508 |  0.933 |   2.125
#>    0.345 |  0.185 |   0.556
#>    2.159 |  1.545 |   3.008
#>    2.840 |  2.147 |   3.885
#>    3.189 |  2.343 |   4.204
#>    0.881 |  0.561 |   1.308
#>    2.592 |  1.889 |   3.519
#>    6.427 |  4.871 |   8.107
#>    2.903 |  2.036 |   3.851
#>    1.907 |  1.246 |   2.682
#>    6.705 |  5.174 |   8.800
#>    1.726 |  1.232 |   2.423
#>    2.038 |  1.454 |   2.758
#>    5.093 |  3.761 |   6.639
#>    1.224 |  0.740 |   1.784
#>    1.400 |  0.953 |   1.971
#>    4.659 |  3.474 |   6.121
#>    1.532 |  0.993 |   2.057
#>    2.774 |  1.942 |   3.715
#>    1.229 |  0.790 |   1.740
#>    0.698 |  0.352 |   1.053
#>    4.964 |  3.633 |   6.590
#>    0.660 |  0.392 |   0.984
#>    9.236 |  6.514 |  12.044
#>    3.083 |  2.128 |   4.038
#>    2.585 |  1.829 |   3.418
#>    3.660 |  2.638 |   4.806
#>    2.457 |  1.690 |   3.329
#>    6.308 |  4.472 |   8.693
#>   10.517 |  8.019 |  13.336
#>    7.871 |  5.390 |  10.610
#>    3.423 |  2.469 |   4.483
#>    7.636 |  5.887 |  10.307
#>    2.378 |  1.538 |   3.437
#>    1.289 |  0.959 |   1.711
#>    2.561 |  1.742 |   3.659
#>    4.948 |  3.727 |   6.335
#>   11.896 |  8.513 |  15.421
#>    6.137 |  4.011 |   8.286
#>    2.045 |  1.407 |   2.857
#>   13.187 |  9.638 |  18.095
#>    7.456 |  5.480 |   9.723
#>    4.481 |  3.414 |   5.764
#>    4.224 |  2.690 |   5.661
#>    5.578 |  3.793 |   7.276
#>    9.156 |  6.602 |  12.227
#>    1.871 |  1.330 |   2.585
#>    2.989 |  2.080 |   4.027
#>    6.068 |  3.971 |   8.353
#>    1.506 |  1.009 |   2.151
#>    1.191 |  0.696 |   1.746
#>    1.746 |  1.263 |   2.386
#>    7.332 |  5.192 |  10.079
#>    5.934 |  4.241 |   7.867
#>   13.548 |  9.629 |  18.179
#>    2.649 |  1.826 |   3.652
#>    2.481 |  1.641 |   3.393
#>   13.228 | 10.078 |  17.257
#>   10.071 |  7.569 |  13.131
#>    8.470 |  6.144 |  11.453
#>    3.360 |  2.451 |   4.546
#>    1.529 |  0.953 |   2.275
#>    3.636 |  2.430 |   4.998
#>    6.240 |  4.066 |   8.395
#>   12.773 |  9.426 |  17.584
#>    3.083 |  2.128 |   4.038
#>    3.231 |  2.135 |   4.365
#>    7.794 |  5.295 |  10.436
#>    9.411 |  6.732 |  12.780
#>    4.183 |  3.119 |   5.378
#>    1.869 |  1.077 |   2.632
#>    7.944 |  5.698 |  10.715
#>    3.597 |  2.649 |   4.813
#> 

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688