Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A dataframe.

ci
Confidence/Credible Interval level. If "default", then 0.95 for Frequentist and 0.89 for Bayesian (see documentation in the bayestestR package).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Arguments passed to or from other methods.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

distance_mahalanobis(data = iris[, 1:4])

#>       Distance    CI_low    CI_high
#> 1    2.2010854 1.5502899  3.0012189
#> 63   2.9309339 2.1972588  3.6387842
#> 74   2.1659767 1.6130408  2.7447841
#> 85   2.5333743 1.9509300  3.1921641
#> 96   2.5278270 1.7458864  3.4054224
#> 107  3.9763735 2.8797868  5.4535057
#> 118  2.9880369 2.3374973  3.7618009
#> 129  1.8917183 1.3651862  2.5531916
#> 140  3.4901608 2.7508627  4.3550244
#> 2    2.4535900 1.9395377  3.0295629
#> 13   3.3682500 2.4359751  4.5906395
#> 24   2.8707752 2.1811969  3.6864841
#> 35   2.6893020 2.0770753  3.2894250
#> 46   3.7204134 2.9161268  4.5828553
#> 57   8.9952224 6.6775079 11.6218842
#> 59  10.0149615 7.3019406 13.0212098
#> 60   5.9275049 4.4718700  7.8940240
#> 61   2.3980207 1.6621428  3.1872309
#> 62   4.6448924 3.4079678  6.2081648
#> 64   3.5346548 2.5109559  4.7210285
#> 65   2.7233286 2.1392236  3.5049026
#> 66   3.0282336 2.1900499  4.0209691
#> 67   3.7393264 2.7875191  4.6791413
#> 68   2.3210294 1.7178380  2.9510788
#> 69   5.4661523 4.2707589  6.9634525
#> 70   2.5199395 1.9604810  3.1309390
#> 71   1.8303651 1.3225343  2.4346858
#> 72   2.2676285 1.5841477  3.0291162
#> 73   2.5745795 1.8937681  3.3966875
#> 75   2.5499324 1.9680288  3.2405455
#> 76   2.0466169 1.5800880  2.5849601
#> 77   4.7721884 3.6592831  6.1276668
#> 78   8.6121830 6.5403671 11.0793736
#> 79   7.4158272 5.4881309  9.7721537
#> 80   2.0418640 1.5490772  2.5538338
#> 81   3.5377899 2.7593728  4.4410467
#> 82   5.4908620 4.1116082  6.9459296
#> 83   3.5678670 2.6568823  4.5668598
#> 84   3.2568195 2.5093823  4.0291242
#> 86   1.9500322 1.4239350  2.6159666
#> 87   2.6352564 1.8596232  3.4226701
#> 88  11.6804442 8.8049389 15.3170359
#> 89   3.4366809 2.6934381  4.2490632
#> 90   3.8638273 3.0169754  4.8180026
#> 91   4.6252597 3.4749254  5.9417661
#> 92   3.0837961 2.3169573  3.8876701
#> 93   4.5584138 3.4567656  6.0404486
#> 94   2.4473379 1.8409195  3.0711383
#> 95   3.0889397 2.1267101  4.1619911
#> 97   1.9914608 1.4445866  2.5823227
#> 98   4.6657579 3.2740212  6.1665236
#> 99   0.6872840 0.3834710  1.0378106
#> 100  3.1630814 2.1666162  4.4237387
#> 101  3.6604184 2.6583014  4.8431309
#> 102  2.1181215 1.4250889  2.8797041
#> 103  3.4084116 2.4050870  4.5854432
#> 104  1.3440867 0.8255234  1.8483316
#> 105  4.4033808 3.4350116  5.7602500
#> 106  2.8718536 1.8565077  3.9639579
#> 108  3.2729893 2.4135028  4.3136940
#> 109  7.8008408 5.9349573 10.1868361
#> 110  0.4641209 0.2120661  0.7692501
#> 111  7.7325431 5.8520698 10.0348654
#> 112  1.6082973 1.0263616  2.2896574
#> 113  1.0797658 0.7016855  1.5197237
#> 114  2.9271853 2.0706021  3.9324284
#> 115  3.5614485 2.6685244  4.6742467
#> 116  3.4454352 2.5391220  4.5595083
#> 117  7.6078400 5.6864996  9.7586282
#> 119  2.2170814 1.6192749  2.9939923
#> 120  3.3652818 2.3779875  4.4913972
#> 121  1.4071745 0.9175084  1.9711408
#> 122  2.4449716 1.7888519  3.4018877
#> 123  4.6636351 3.4065874  6.3598438
#> 124  1.6844424 1.0199447  2.3254009
#> 125  2.4488966 1.6810250  3.3052427
#> 126  4.3805519 2.9009321  5.7500079
#> 127  1.4947224 0.9335046  2.1194581
#> 128  0.3447497 0.1750044  0.5481137
#> 130  2.1574458 1.5399748  2.9637014
#> 131  2.8468328 2.1093966  3.8547357
#> 132  3.1777683 2.3767379  4.2466929
#> 133  0.8775595 0.5561597  1.3036608
#> 134  2.5791924 1.7776683  3.4350226
#> 135  6.3917234 5.0124890  8.2683683
#> 136  2.8816002 2.0908257  3.8371363
#> 137  1.8973453 1.2332323  2.7099874
#> 138  6.7201296 4.9548313  8.6111517
#> 139  1.7201630 1.1750388  2.3617932
#> 141  2.0328842 1.4486124  2.7402071
#> 142  5.1033110 3.8066057  6.7438263
#> 143  1.2265695 0.7499918  1.7976362
#> 144  1.4008596 0.9557928  1.9918399
#> 145  4.6378877 3.5345652  6.1594789
#> 146  1.5303551 1.0476521  2.0612816
#> 147  2.7690902 1.9401129  3.7640100
#> 148  1.2226922 0.8388941  1.7879256
#> 149  0.6919426 0.3503974  1.0888321
#> 150  4.9437404 3.7083109  6.4765247
#> 3    0.6597819 0.3918376  0.9453994
#> 4    9.1379907 6.5791192 12.0994636
#> 5    3.0690807 2.2068252  4.0977848
#> 6    2.5594602 1.7848302  3.4331666
#> 7    3.6645718 2.6703948  4.8950507
#> 8    2.4269194 1.6911931  3.3068993
#> 9    6.2995611 4.1876477  8.6355803
#> 10  10.4710186 8.0028206 13.2926593
#> 11   7.8372076 5.5498233 10.6808711
#> 12   3.4112493 2.4929948  4.5993960
#> 14   7.6231652 5.7668858 10.2892358
#> 15   2.3690716 1.5494671  3.3826832
#> 16   1.2862313 0.9742616  1.7140025
#> 17   2.5284045 1.7929802  3.5986695
#> 18   4.9032933 3.5660505  6.2709273
#> 19  11.8215505 8.6452771 15.5133072
#> 20   6.0960847 4.1395135  8.3610945
#> 21   2.0598829 1.3237918  2.7673003
#> 22  13.2669254 9.1327759 17.3806102
#> 23   7.4213230 5.4976858  9.9531359
#> 25   4.5080128 3.3397203  5.7756584
#> 26   4.2183582 2.8743677  5.7676046
#> 27   5.5455672 3.9401652  7.2257906
#> 28   9.1243629 6.2327924 12.2954557
#> 29   1.8542008 1.3195774  2.5430114
#> 30   3.0037897 2.1264420  4.0056432
#> 31   6.0817334 4.1359413  8.4197206
#> 32   1.5027709 1.0010311  2.1529447
#> 33   1.1854325 0.7220898  1.7190226
#> 34   1.7263422 1.2354108  2.3231367
#> 36   7.2764923 5.0689887  9.9754813
#> 37   5.9135528 4.1924640  8.2266507
#> 38  13.5698646 9.8581207 18.6494080
#> 39   2.6278541 1.8502665  3.6529483
#> 40   2.4922719 1.6673769  3.4240001
#> 41  13.2694074 9.5821001 17.2351869
#> 42   9.9780554 7.6669113 12.8833015
#> 43   8.4526171 6.0192563 11.2977973
#> 44   3.3616227 2.4633971  4.5097649
#> 45   1.5319756 0.9133586  2.1608353
#> 47   3.6303940 2.5678867  5.0411032
#> 48   6.1960170 4.3158989  8.5983292
#> 49  12.8730209 9.3490245 17.5849162
#> 50   3.0690807 2.2068252  4.0977848
#> 51   3.2276897 2.2819646  4.4335244
#> 52   7.7223451 5.1500406 10.3150532
#> 53   9.3742205 6.6211824 12.8820090
#> 54   4.1593983 3.1289857  5.3672625
#> 55   1.8538989 1.2121273  2.7382019
#> 56   7.9072108 5.5989899 10.5271948
#> 58   3.5795633 2.6308368  4.6805539