Mahalanobis distance and confidence interval (CI)

The Mahalanobis distance (in squared units) measures the distance in multivariate space taking into account the covariance structure of the data. Because a few extreme outliers can skew the covariance estimate, the bootstrapped version is considered as more robust.

distance_mahalanobis(data, ci = 0.95, iterations = 1000, robust = TRUE, ...)

Arguments

data
A data frame.

ci
Confidence/Credible Interval level. If "default", then it is set to 0.95 (95% CI).

iterations
The number of draws to simulate/bootstrap (when robust is TRUE).

robust
If TRUE, will run a bootstrapped version of the function with i iterations.

...
Additional arguments (e.g., alternative) to be passed to other methods. See stats::cor.test for further details.

Value

Description of the Mahalanobis distance.

References

Schwarzkopf, D. S., De Haas, B., & Rees, G. (2012). Better ways to improve standards in brain-behavior correlation analysis. Frontiers in human neuroscience, 6, 200.

Examples

library(correlation)

distance_mahalanobis(iris[, 1:4])

#>       Distance     CI_low    CI_high
#> 1    2.1871428  1.5925338  3.0046820
#> 2    2.9463673  2.2271896  3.7201721
#> 3    2.1618444  1.6113513  2.7236471
#> 4    2.5362152  2.0051103  3.2087760
#> 5    2.5090579  1.7870518  3.3473464
#> 6    3.9867769  2.8003603  5.4071910
#> 7    2.9663401  2.3428104  3.6983278
#> 8    1.8739561  1.3955143  2.5228266
#> 9    3.5088952  2.7407855  4.3004721
#> 10   2.4674122  1.9327819  2.9692516
#> 11   3.3594614  2.4038423  4.4996410
#> 12   2.8813823  2.2266934  3.6488770
#> 13   2.7012799  2.0766763  3.3097689
#> 14   3.7280352  2.9986120  4.6658210
#> 15   9.0233230  6.6803233 11.5984119
#> 16   9.9650316  7.0959457 13.0140225
#> 17   5.9232517  4.3576333  7.6846073
#> 18   2.3927873  1.6940228  3.2141952
#> 19   4.6314116  3.3500070  6.0675908
#> 20   3.5331709  2.4837596  4.6643893
#> 21   2.7135401  2.0855507  3.4162015
#> 22   3.0030917  2.2356906  4.0304277
#> 23   3.7263395  2.9616920  4.6572931
#> 24   2.3285833  1.7044439  2.9940852
#> 25   5.4840895  4.2301203  6.8757330
#> 26   2.5206355  1.9766531  3.2175793
#> 27   1.8270391  1.3205419  2.4212150
#> 28   2.2492201  1.6530854  3.0353063
#> 29   2.5793615  1.8131039  3.2916730
#> 30   2.5407964  1.9662377  3.1992043
#> 31   2.0535159  1.5890727  2.5639459
#> 32   4.7376906  3.5497351  6.0016535
#> 33   8.6475728  6.5758484 11.0147336
#> 34   7.4393790  5.2582631  9.7467264
#> 35   2.0542907  1.5715648  2.6236721
#> 36   3.5312819  2.7712737  4.4596672
#> 37   5.5104326  4.2394616  6.9681599
#> 38   3.5593850  2.7303028  4.5588378
#> 39   3.2639414  2.5486202  4.0657552
#> 40   1.9557957  1.4363931  2.6087525
#> 41   2.6311451  1.8684929  3.3570597
#> 42  11.7452915  8.3754748 15.4453379
#> 43   3.4289880  2.7097134  4.2116111
#> 44   3.8724280  3.0465696  4.8417288
#> 45   4.6356489  3.5059028  6.0156277
#> 46   3.0921367  2.2814762  3.9507299
#> 47   4.5812511  3.3633556  5.8441651
#> 48   2.4417248  1.8937038  3.0523814
#> 49   3.0717261  2.2320181  4.1656409
#> 50   1.9884367  1.4263979  2.5592584
#> 51   4.6834695  3.3823413  6.3468328
#> 52   0.6845473  0.3912871  1.0820501
#> 53   3.1785718  2.1441884  4.4048638
#> 54   3.6803494  2.5654397  4.8997210
#> 55   2.1321382  1.4393813  2.9118767
#> 56   3.4191504  2.5192526  4.6798898
#> 57   1.3321340  0.8263656  1.9383281
#> 58   4.4295405  3.2409206  5.6092639
#> 59   2.8532374  1.8404832  3.8885567
#> 60   3.2872203  2.3606054  4.2665546
#> 61   7.8444745  5.8917775 10.4355819
#> 62   0.4666511  0.2040211  0.7807567
#> 63   7.7690880  6.0767542 10.2141553
#> 64   1.6124497  0.9977590  2.2430513
#> 65   1.0975528  0.6962772  1.5289754
#> 66   2.9206401  2.0167875  3.8681065
#> 67   3.5953764  2.5733840  4.6104997
#> 68   3.4421891  2.6003640  4.5782336
#> 69   7.6032434  5.8389226  9.9400979
#> 70   2.2440710  1.6527453  2.9728876
#> 71   3.3764103  2.3231332  4.5085219
#> 72   1.4176364  0.9881208  2.0004862
#> 73   2.4747609  1.7440358  3.3079684
#> 74   4.6506052  3.4004695  6.2779930
#> 75   1.6841176  1.0849987  2.3939982
#> 76   2.4506619  1.6829987  3.3128148
#> 77   4.3445317  3.1066231  5.9256131
#> 78   1.5146709  0.9992826  2.1641180
#> 79   0.3449145  0.1950378  0.5511105
#> 80   2.1718254  1.4875180  2.9433208
#> 81   2.8766870  2.0841016  3.8244464
#> 82   3.2136364  2.3501719  4.1644718
#> 83   0.8957974  0.5335066  1.2718718
#> 84   2.5777436  1.8850298  3.4599782
#> 85   6.4221348  4.7475078  7.9784367
#> 86   2.8684226  2.1353135  4.0087275
#> 87   1.9088820  1.2341388  2.6831193
#> 88   6.7600009  5.1842981  8.8452455
#> 89   1.7280126  1.1740133  2.3401501
#> 90   2.0413496  1.4928072  2.8166067
#> 91   5.0902214  3.8768007  6.7551448
#> 92   1.2296489  0.7356051  1.7810505
#> 93   1.4174961  0.9412757  1.9442844
#> 94   4.6617067  3.5348435  6.3296152
#> 95   1.5307531  1.0598133  2.1052827
#> 96   2.7803091  1.9889884  3.7872646
#> 97   1.2281239  0.7799709  1.7018149
#> 98   0.6940166  0.3576701  1.0965702
#> 99   4.9656023  3.7475931  6.8121571
#> 100  0.6534918  0.3880999  0.9685119
#> 101  9.1885965  6.7038645 12.1429149
#> 102  3.0688682  2.2365942  4.1921542
#> 103  2.5751196  1.8340177  3.4810799
#> 104  3.6659813  2.6653267  4.8630664
#> 105  2.4455619  1.7173460  3.4008478
#> 106  6.3216023  4.4122666  8.9471669
#> 107 10.5076107  8.2797187 13.4043660
#> 108  7.9116933  5.1558389 10.5912673
#> 109  3.4516555  2.5153476  4.5638428
#> 110  7.6387796  5.6348741 10.1374073
#> 111  2.3413960  1.5243534  3.4339104
#> 112  1.2984646  0.9647175  1.6676985
#> 113  2.5635130  1.7113586  3.5391779
#> 114  4.9641434  3.6557201  6.3351071
#> 115 11.8990131  8.6551512 15.6164542
#> 116  6.0879354  4.0824552  8.4076973
#> 117  2.0489733  1.3483459  2.8307587
#> 118 13.2006063  9.3693853 18.3047030
#> 119  7.4356802  5.5397612  9.8842725
#> 120  4.5151094  3.5134446  5.8036641
#> 121  4.1727786  2.8954989  5.9143384
#> 122  5.5594478  3.8723371  7.2455728
#> 123  9.1450135  6.1735828 12.3123770
#> 124  1.8811715  1.3167720  2.5610346
#> 125  2.9688095  2.0958636  4.1410325
#> 126  6.0969972  4.1611860  8.7312139
#> 127  1.5215967  0.9805858  2.1087791
#> 128  1.1849048  0.7270091  1.8000451
#> 129  1.7336197  1.2657551  2.3591909
#> 130  7.3732167  4.8216827 10.0472803
#> 131  5.9143197  4.0500435  7.9920991
#> 132 13.6409126  9.6335745 18.5769058
#> 133  2.6405726  1.8842083  3.7062964
#> 134  2.4907887  1.7221013  3.5298335
#> 135 13.2485124 10.0007738 17.6824040
#> 136 10.0503750  7.6602239 12.9652722
#> 137  8.4591735  6.0001255 11.3849836
#> 138  3.3749129  2.3430054  4.4680402
#> 139  1.5281014  0.9573695  2.2869982
#> 140  3.6644760  2.5028620  5.0401749
#> 141  6.1709422  4.2381091  8.6294268
#> 142 12.9520987  9.0912806 17.4578225
#> 143  3.0688682  2.2365942  4.1921542
#> 144  3.2281298  2.2296673  4.5425793
#> 145  7.6931748  5.4127818 10.6600374
#> 146  9.3722525  6.6710753 12.9822502
#> 147  4.1842848  3.1596806  5.5105395
#> 148  1.8555361  1.1938477  2.7388540
#> 149  7.9424734  5.7393340 10.7507509
#> 150  3.5929082  2.5246015  4.5931013

distance_mahalanobis(iris[, 1:4], robust = FALSE)

#>       Distance
#> 1    2.1344679
#> 2    2.8491187
#> 3    2.0813387
#> 4    2.4523816
#> 5    2.4621545
#> 6    3.8834177
#> 7    2.8621081
#> 8    1.8333003
#> 9    3.3840731
#> 10   2.3752179
#> 11   3.2831069
#> 12   2.7747975
#> 13   2.6132975
#> 14   3.6034324
#> 15   8.7375184
#> 16   9.7127899
#> 17   5.7605877
#> 18   2.3213894
#> 19   4.4996899
#> 20   3.4388658
#> 21   2.6360071
#> 22   2.9292496
#> 23   3.6134114
#> 24   2.2371731
#> 25   5.3023607
#> 26   2.4453103
#> 27   1.7658286
#> 28   2.1971806
#> 29   2.5027712
#> 30   2.4643980
#> 31   1.9849638
#> 32   4.5911380
#> 33   8.3583413
#> 34   7.2213139
#> 35   1.9820679
#> 36   3.4173031
#> 37   5.3372175
#> 38   3.4513350
#> 39   3.1549793
#> 40   1.8926197
#> 41   2.5485013
#> 42  11.4240288
#> 43   3.3144697
#> 44   3.7085855
#> 45   4.4840560
#> 46   2.9786602
#> 47   4.4333077
#> 48   2.3594309
#> 49   3.0076970
#> 50   1.9285641
#> 51   4.4528311
#> 52   0.6273996
#> 53   3.0186280
#> 54   3.6125278
#> 55   2.0404590
#> 56   3.3195858
#> 57   1.2763441
#> 58   4.3093305
#> 59   2.7426578
#> 60   3.1665229
#> 61   7.6832930
#> 62   0.4323176
#> 63   7.5614023
#> 64   1.5482870
#> 65   1.0372548
#> 66   2.7865631
#> 67   3.4732716
#> 68   3.3289796
#> 69   7.3923779
#> 70   2.1837666
#> 71   3.2615033
#> 72   1.3601100
#> 73   2.3885294
#> 74   4.5073307
#> 75   1.6095193
#> 76   2.3362612
#> 77   4.1982905
#> 78   1.4219664
#> 79   0.3194730
#> 80   2.1112223
#> 81   2.8147732
#> 82   3.1338323
#> 83   0.8579871
#> 84   2.5178289
#> 85   6.2164814
#> 86   2.7910313
#> 87   1.7971852
#> 88   6.5544318
#> 89   1.6677135
#> 90   2.0056841
#> 91   4.9630348
#> 92   1.1821941
#> 93   1.3713881
#> 94   4.5698858
#> 95   1.4932609
#> 96   2.6869276
#> 97   1.1870542
#> 98   0.6575354
#> 99   4.8207427
#> 100  0.6341019
#> 101  8.9395988
#> 102  2.9682833
#> 103  2.4451456
#> 104  3.5551146
#> 105  2.3541837
#> 106  6.0617111
#> 107 10.1378044
#> 108  7.5880086
#> 109  3.3301655
#> 110  7.3453376
#> 111  2.2611037
#> 112  1.2342978
#> 113  2.4316524
#> 114  4.7588845
#> 115 11.4105735
#> 116  5.8815414
#> 117  1.9743869
#> 118 12.8130732
#> 119  7.1768159
#> 120  4.3974981
#> 121  4.0399248
#> 122  5.3595987
#> 123  8.7973122
#> 124  1.7967966
#> 125  2.8868280
#> 126  5.8835660
#> 127  1.4473698
#> 128  1.1348285
#> 129  1.6688603
#> 130  7.0710485
#> 131  5.6935238
#> 132 13.1010925
#> 133  2.5441171
#> 134  2.4087367
#> 135 12.8803310
#> 136  9.6569355
#> 137  8.2028004
#> 138  3.2575803
#> 139  1.4687445
#> 140  3.4924068
#> 141  5.9713581
#> 142 12.4413843
#> 143  2.9682833
#> 144  3.1069367
#> 145  7.4592052
#> 146  9.0639497
#> 147  4.0366487
#> 148  1.7670035
#> 149  7.6824724
#> 150  3.4787688